### 2024-02-06 #6328 @editorialbot
[PRE REVIEW]: extendr: Frictionless bindings for R and Rust
**Submitting author:** @cgmossa (<a href="http://orcid.org/0009-0007-9297-1523">Mossa Reimert</a>)
**Repository:** https://github.com/extendr/extendr
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.6.0
