### 2024-02-06 #6328 @danielskatz
@openjournals/dev - for some reason, the `check repository` command that was automatically run when this issue was opened doesn't seem to have given the expected word count for the paper, which seems like a bug. Can you check on this? (I don't see any issue with the paper length, this is really a request to make sure the command works properly in all cases)
