### 2024-02-06 #6328 @danielskatz
👋 @CGMossa - thanks for your submission. Can you explain how this is research software, as I'm not sure I see it that way from the paper. What kinds of research papers would you expect to cite this work?
