### 2024-02-06 #6328 @CGMossa
Thanks for your prompt handling of our submission!  This software has broad applicability (as you can see from the diverse affiliations of the authors), but to give you a specific example:  the focus of the team at the University of Copenhagen is on research into infectious disease modelling.  Specifically, we are using extendR to facilitate implementation of agent-based models of disease spread, where there is a big need for researchers to code up complex models in an efficient language, where most/all of them are based around workflows in R.

My own PhD project is on African Swine Fever in domestic pigs and wild boar (and we would cite this paper in future work on specific models), but there are a lot of other similar applications in veterinary epidemiology and human infectious disease modelling.  We cited a few of the researchers using Rust for models of ASF in wild boar in the introduction to our paper, but there are many others that are currently using a combination of e.g. C++ and R but would benefit from switching to Rust with an interface in R that we expect to cite extendR.  In fact, we see extendR as fulfilling the same remit for R/Rust as Rcpp does for R/C++ - and the relevant article for Rcpp (https://scholar.google.com/scholar?hl=en&as_sdt=0%2C5&q=Rcpp&btnG=) has been cited almost 2000 times!

We also mention changeforest https://github.com/mlondschien/changeforest/ which is published research software & paper. They would probably have cited an extendr-reference if it was present at that time.

Researchers that mainly write simulation/analysis/visualisation code in R, that occassionally need performance-focused code are one potential target. changeforest is an example of that.

In JOSS there is a paper on JuliaCall https://github.com/Non-Contradiction/JuliaCall, https://joss.theoj.org/papers/10.21105/joss.01284 which fulfills the exact same niche we do. However, we cited jlrs, because it brings Rust to Julia (which makes more sense for our paper).

I hope that helps to alleviate your concerns but please let me know if you would like further explanation.
