### 2024-02-06 #6328 @JosiahParry
To provide a bit more context extendr plays a core roll in [this network merge paper](https://github.com/nptscot/networkmerge/) that was recently submitted. A key part of the algorithm is written in rust and bindings are provided to R via extendr. Those rust algorithms/routines are then used as a part of a larger algorithm to merge two transit networks.
