### 2024-02-06 #6328 @danielskatz
Thanks @CGMossa and @JosiahParry - I'm going to ask the JOSS editorial board to consider this issue and your responses, and then we'll make a decision about this, probably in a week or two.
