+++ paper.md
@@ -6,7 +6,6 @@ tags:
   - FFI
   - bindings
 subtitle: "interfacing Rust code within R packages"
-# <!-- This is just a suggestion, but the previous subtitle didn't read very well to me -->
 authors:
   - name: "Mossa Merhi Reimert"
     orcid: 0009-0007-9297-1523
@@ -58,20 +57,14 @@ in both commercial and research applications where performance is important. Com
 include web development and game development, and in the research domain Rust is increasingly being
 used in a wide range of contexts including change point detection [@JMLR:v24:22-0512], high-performance
 GIF encoding [@gifski_cran], and agent-based models of disease spread [@eval_rust_for_custom_abm; @epirust_paper; @forth_african_2022-1].
-<!-- Other references/examples of Rust usage in research would be useful! -->
 
 However, typical workflows in research domains, such as disease modelling, often rely on higher-level programming languages due to lower entry barriers.
 This results in broader adoption within scientific communities, compared to the use of low-level languages like C++ and Rust.
 The statistical programming language[R]( https://www.r-project.org) is one of the most widely used
 high-level languages in research. R's official interpreter is written in C, and it provides a C API as well as
 tools for building dynamic libraries in Fortran/C/C++ natively.
-<!-- On Windows, the R-project provides
-a toolchain, Rtools, which bundles a developer environment for Fortran, C and C++. -->
-<!-- The R-project provides documentation for developing extensions in [R-internals](https://cran.r-project.org/doc/manuals/R-ints.html)
-and [R-extenstions](https://cran.r-project.org/doc/manuals/R-exts.html). -->
 The 'Extending R' book [@chambers2017extending] also describes interfacing with other languages
 such as Python and Julia.
-<!--Java?-->
 
 The strength of R is its ecosystem of packages, the vast majority of which are available from [CRAN](https://cran.r-project.org).
 They are primarily written by research scientists, specialists, and professionals.
@@ -79,7 +72,7 @@ Another important use case of R packages is being a front-end for other language
 Automated toolings that provide scaffolding and boilerplate code are widely used to simplify cross-language integration.
 For example, embedding C++ code is a good way to resolve performance bottlenecks within R packages, and it can be easily accomplished using cpp11 [@cpp11] or Rcpp [@rcpp_jss].
 Rust demonstrates similar performance to C++, but it also offers other beneficial features such as declarative memory management, which provides compile-time guarantees for memory safety in the absence of a garbage collector.
-<!-- Rust and R also share support for WebAssembly [webR](https://docs.r-wasm.org/webr/latest/). -->
+
 We note that other scientific computing communities have already introduced plug-ins for Rust, including Python via [PyO3](https://github.com/PyO3/pyo3),
 and Julia via [jlrs](https://github.com/Taaitaaiger/jlrs).
 
@@ -98,10 +91,10 @@ Rcpp [@rcpp_cran] and cpp11 [@cpp11] for C++, Java via rJava [@rJava], Python wi
 V8 R-package [@v8_cran] are among the most used.
 In contrast, bindings between Rust and R, such as [`gifski`](https://crates.io/crates/gifski) [@gifski_cran], are currently mostly written by hand.
 
-<!--This contradicts the statement above that we do not know other bindings provider-->
+
 We note that there exist other software packages providing bindings between R and Rust.
 The Rust crate / R-package `roxido` / [`cargo`](https://github.com/dbdahl/cargo-framework) [@cargo_cran] provides a mechanism for embedding Rust code within R packages.
-<!-- Looks like this should be cited:  David B. Dahl (2021), Writing R Extensions in Rust, arXiv:2108.07179 [cs.PL], URL https://arxiv.org/abs/2108.07179. -->
+
[…]
