### 2024-02-07 #6328 @mdenwood
@danielskatz thanks for looking at the paper.

Could I also add another point to the discussion with the JOSS editorial board:  our paper is very similar in scope and purpose to https://joss.theoj.org/papers/10.21105/joss.01284 - which is one of the reasons we sent the extendR paper to JOSS. According to google scholar, the JuliaCall paper has been cited 15 times in a variety of fields, but I actually see the use case for R/Rust integration to be substantially more compelling than that for R/Julia integration - and perhaps eventually (if Rust continues its current growth in popularity) even similar to the use case for R/C++ integration (where the Rcpp article in JSS that @CGMossa mentioned has been very frequently cited).

I would also like to note that my own JSS article on interfacing R to JAGS (another similar use case to our paper) is by far the most frequently cited of my >100 research publications (https://scholar.google.co.uk/citations?user=Zos2XTkAAAAJ&hl=en), so I am a big believer in the research impact of this kind of software.
