### 2024-02-09 #6328 @danielskatz
Note: if this passes the scope review, @jbytecode would like to edit it
