### 2024-02-13 #6328 @CGMossa
@danielskatz That's good - I'm very happy that somebody has already volunteered for editorial duties!

I was also contacted by @luciorq who expressed an interest in reviewing the paper, if it passes editorial review. They heard about the paper submission via our discord server but are independent of the extendR project and we have not previously been in contact, so there is no conflict of interest. I couldn't find any other mechanism to suggest reviewers so have done so here - hope that's OK.
