### 2024-02-13 #6328 @danielskatz
@CGMossa - please see the first comment in this issue, which includes

>@cgmossa if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

so basically, yes, we are happy with recommendations for reviewers, but in general don't want them tagged so they don't get notified before an editor chooses to ask them, to not send extra notifications to people...
