### 2024-02-13 94a20a2
paper: moved commands in the getting started sections
