### 2024-02-14 #6328 @jbytecode
@CGMossa  - Hi, thank you for submitting your paper on such a valuable software to JOSS. I am the handling editor. We will be handling whole the stuff during the review process until a final decision is made by our track editor. Upon initial review, I noticed a couple of things that need correction in your BibTeX file (as well as the manuscript), but these can also be addressed after the review as they require only minor changes.

We need to identify at least 2 suitable reviewers. It appears you have already suggested one reviewer, and I would be happy to invite them as they seem to have expertise in R.

You may also suggest additional reviewers from the potential reviewers' list provided at the top of this issue. I believe including at least one Rustacean who is familiar with the R environment would be beneficial. Please provide the list of potential reviewers without using the '@' character to avoid triggering notifications.

Thank you in advance!
