### 2024-02-14 #6328 @jbytecode
@CGMossa - you can also investigate the suggested papers (Five most similar historical JOSS papers) and we can give priority to their authors.
