### 2024-02-14 #6328 @CGMossa
Thank you very much the guidance on the next steps.

- I know a Rustacean with familiarity of the R ecosystem, that is not affiliated with extendr:
Dccsillag

 - I have looked at the list of authors on the similar papers list. It does not seem like I can find a Rust user amongst them, who also happens to be an R user. Atleast not according to the hosted repositories on their github pages.
Looking at https://reviewers.joss.theoj.org/, I believe you can also ask:
Esteinig (Eike Steinig)
Robmoss
lskatz
Dhvalden
Jeremylt (no R repositories on github, but seems like a good choice)
DanielRivasMD
MaaniBeigy

- One last thing, am I allowed to push commits to the paper now that I’ve submitted it? Currently, I have 2 un-pushed commits: (1) removes all commented text, as I think it might have messed with the word count algorithm, see https://github.com/openjournals/joss-reviews/issues/6328#issuecomment-1930002710 (2) moved a line of code in “Getting started” section from one code block, to another code-block above it.
