### 2024-02-14 #6328 @jbytecode
@CGMossa - please trigger a `@editorialbot generate pdf` task to update your manuscript when you've finished your changes.
