### 2024-02-16 c22a9ed
Update toml_edit requirement from 0.21.0 to 0.22.6
Updates the requirements on [toml_edit](https://github.com/toml-rs/toml) to permit the latest version.
- [Commits](https://github.com/toml-rs/toml/compare/v0.21.0...v0.22.6)

---
updated-dependencies:
- dependency-name: toml_edit
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
