+++ xtask/Cargo.toml
@@ -11,5 +11,5 @@ publish = false
 
 [dependencies]
 clap = { version = "4.4.6", features = ["derive"] }
-toml_edit = "0.21.0"
+toml_edit = "0.22.6"
 xshell = "0.2.5"
