### 2024-02-18 #6328 @jbytecode
👋👋👋 Dear @robmoss 👋👋👋

**extendr: Frictionless bindings for R and Rust**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6328).

Thank you in advance!
