+++ trait.impl/core/clone/trait.Clone.js
@@ -1,4 +1,4 @@
 (function() {var implementors = {
-"extendr_api":[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/complexes/struct.Complexes.html\" title=\"struct extendr_api::wrapper::complexes::Complexes\">Complexes</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/list/struct.ListIter.html\" title=\"struct extendr_api::wrapper::list::ListIter\">ListIter</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/symbol/struct.Symbol.html\" title=\"struct extendr_api::wrapper::symbol::Symbol\">Symbol</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/lang/struct.Language.html\" title=\"struct extendr_api::wrapper::lang::Language\">Language</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/expr/struct.Expressions.html\" title=\"struct extendr_api::wrapper::expr::Expressions\">Expressions</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/iter/struct.StrIter.html\" title=\"struct extendr_api::iter::StrIter\">StrIter</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/s4/struct.S4.html\" title=\"struct extendr_api::wrapper::s4::S4\">S4</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/scalar/rbool/struct.Rbool.html\" title=\"struct extendr_api::scalar::rbool::Rbool\">Rbool</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/strings/struct.Strings.html\" title=\"struct extendr_api::wrapper::strings::Strings\">Strings</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/clone/trait.Clone.html\" title=\"trait core::clone::Clone\">Clone</a> for <a class=\"struct\" href=\"extendr_api/wrapper/primitive/struct.Primitive.html\" tit
[…]

[…]
