### 2024-02-19 #6328 @robmoss
@jbytecode in principle I'd be interested in reviewing this paper, but my capacity is extremely limited at the moment — I have impending deadlines for several research contracts —  and I don't believe I could complete the review in a timely manner. I have another JOSS review underway, and once I finish it I might be able to review this paper, if you're still searching for reviewers at that point. Sorry!
