### 2024-02-19 #6328 @jbytecode
@robmoss - thank you for the response. Please keep in touch.

@CGMossa - You mentioned that `luciorq` has accepted to review if the submission passed the scope check, but it seems they haven't catch our notifications yet. Maybe you can reach them on discord if possible.
