### 2024-02-22 #6328 @luciorq
Dear @jbytecode,

Thank you for the invitation to review "extendr: Frictionless bindings for R and Rust" for JOSS. I am intrigued by the subject, given my background in R and bindings to other languages.

Looking forward to contributing to the JOSS community.
