### 2024-02-22 #6328 @lskatz
Hi, thank you for including me. I do not feel like I am enough of an expert in R and Rust to view language bindings between them, but I am comforted that you found enough reviewers to move forward.
