### 2024-02-29 0d82b42
Add user-documentation to S4 (#680)
* Explain `set_class` and add advice on R-package...

* Added an R example
