+++ trait.impl/serde/de/trait.MapAccess.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.MapAccess.html\" title=\"trait serde::de::MapAccess\">MapAccess</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.NamedListGetter.html\" title=\"struct extendr_api::deserializer::NamedListGetter\">NamedListGetter</a>&lt;'de&gt;"]]
+"extendr_api":[["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.197/serde/de/trait.MapAccess.html\" title=\"trait serde::de::MapAccess\">MapAccess</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.NamedListGetter.html\" title=\"struct extendr_api::deserializer::NamedListGetter\">NamedListGetter</a>&lt;'de&gt;"]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
