+++ trait.impl/serde/de/trait.Visitor.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.DoublesVisitor.html\" title=\"struct extendr_api::deserializer::DoublesVisitor\">DoublesVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.IntegersVisitor.html\" title=\"struct extendr_api::deserializer::IntegersVisitor\">IntegersVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.LogicalsVisitor.html\" title=\"struct extendr_api::deserializer::LogicalsVisitor\">LogicalsVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.RboolVisitor.html\" title=\"struct extendr_api::deserializer::RboolVisitor\">RboolVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.RfloatVisitor.html\" title=\"struct extendr_api::deserializer::RfloatVisitor\">RfloatVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.RintVisitor.html\" title=\"struct extendr_api::deserializer::RintVisitor\">RintVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.RobjVisitor.html\" title=\"struct extendr_api::deserializer::RobjVisitor\">RobjVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/trait.Visitor.html\" title=\"trait serde::de::Visitor\">Visitor</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.RstrVisitor.html\" title=\"struct extendr_api::deserializer::RstrVisitor\">RstrVisitor</a>"],["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.196/serde/de/t
[…]

[…]
