### 2024-03-03 7af4f53
Move faer to optional
