### 2024-03-03 7cfe74d
Move faer to optional
