### 2024-03-03 d972424
Implement traits for MatRef<f64>
