### 2024-03-03 c40a436
Add TryFrom
