### 2024-03-04 5d0627d
Add better impl traits
