### 2024-03-04 201e9da
Bump faer to v0.18.0
