+++ extendr-api/Cargo.toml
@@ -20,7 +20,7 @@ libc = { version = "0.2", optional = true }
 ndarray = { version = "0.15.3", optional = true }
 num-complex = { version = "0.4", optional = true }
 serde = { version = "1.0", features = ["derive"], optional = true }
-faer = { version = "0.17.1", optional = true }
+faer = { version = "0.18.0", optional = true }
 
 [dev-dependencies]
 extendr-engine = { path = "../extendr-engine" }
