+++ extendr-api/src/optional/mod.rs
@@ -3,7 +3,7 @@ A set of optional features and third-party crate integrations, usually hidden be
 */
 #[cfg(feature = "either")]
 pub mod either;
-#[cfg(feature = "ndarray")]
-pub mod ndarray;
 #[cfg(feature = "faer")]
 pub mod faer;
+#[cfg(feature = "ndarray")]
+pub mod ndarray;
