### 2024-03-08 298dd6e
Add R tests
