+++ tests/extendrtests/R/extendr-wrappers.R
@@ -86,6 +86,8 @@ hello_submodule <- function() .Call(wrap__hello_submodule)
 #' @export
 euclidean_dist <- function(a) .Call(wrap__euclidean_dist, a)
 
+faer_mat <- function(x) .Call(wrap__faer_mat, x)
+
 type_aware_sum <- function(input) .Call(wrap__type_aware_sum, input)
 
 #' Test raw identifiers (`r#`) in function arguments are parsed correctly.
