+++ xtask/src/main.rs
@@ -37,6 +37,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         cli::Commands::Doc => commands::generate_docs::run(&shell)?,
         cli::Commands::Msrv => commands::cargo_msrv::run(&shell)?,
         cli::Commands::DevtoolsTest => commands::devtools_test::run(&shell)?,
+        // TODO: Accept snapshots from command-line
+        // R -e "testthat::snapshot_accept(path = 'tests/extendrtests/tests/testthat', 'macro-snapshot')" 
     };
 
     Ok(())
