### 2024-03-16 6e947bf
xtask: use `DocumentMut` instead of `Document` from toml_edit
