+++ xtask/src/extendrtests/with_absolute_path.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 use std::path::Path;
 
-use toml_edit::{Document, InlineTable, Value};
+use toml_edit::{DocumentMut, InlineTable, Value};
 use xshell::Shell;
 
 use crate::extendrtests::path_helper::RCompatiblePath;
@@ -12,12 +12,12 @@ const RUST_FOLDER_PATH: &str = "tests/extendrtests/src/rust";
 const CARGO_TOML: &str = "Cargo.toml";
 
 #[derive(Debug)]
-pub(crate) struct DocumentHandle<'a> {
+pub(crate) struct DocumentMutHandle<'a> {
     document: Vec<u8>,
     shell: &'a Shell,
 }
 
-impl<'a> Drop for DocumentHandle<'a> {
+impl<'a> Drop for DocumentMutHandle<'a> {
     fn drop(&mut self) {
         let _rust_folder = self.shell.push_dir(RUST_FOLDER_PATH);
         self.shell
@@ -26,13 +26,14 @@ impl<'a> Drop for DocumentHandle<'a> {
     }
 }
 
-pub(crate) fn swap_extendr_api_path(shell: &Shell) -> Result<DocumentHandle, Box<dyn Error>> {
+pub(crate) fn swap_extendr_api_path(shell: &Shell) -> Result<DocumentMutHandle, Box<dyn Error>> {
     let current_path = shell.current_dir();
     let _rust_folder = shell.push_dir(RUST_FOLDER_PATH);
 
     let original_cargo_toml_bytes = read_file_with_line_ending(shell, CARGO_TOML)?;
 
-    let original_cargo_toml: Document = std::str::from_utf8(&original_cargo_toml_bytes)?.parse()?;
+    let original_cargo_toml: DocumentMut =
+        std::str::from_utf8(&original_cargo_toml_bytes)?.parse()?;
 
     let mut cargo_toml = original_cargo_toml.clone();
 
@@ -56,7 +57,7 @@ pub(crate) fn swap_extendr_api_path(shell: &Shell) -> Result<DocumentHandle, Box
     *libR_sys_entry = Value::InlineTable(replacement);
 
     shell.write_file(CARGO_TOML, cargo_toml.to_string())?;
-    Ok(DocumentHandle {
+    Ok(DocumentMutHandle {
         document: original_cargo_toml_bytes,
         shell,
     })
@@ -75,7 +76,7 @@ fn get_replacement_path_libR_sys(path: &Path) -> String {
     format!("{path}/libR-sys")
 }
 
-fn get_extendr_api_entry(document: &mut Document) -> Option<&mut Value> {
+fn get_extendr_api_entry(document: &mut DocumentMut) -> Option<&mut Value> {
     document
         .get_mut("patch")?
         .get_mut("crates-io")?
@@ -84,7 +85,7 @@ fn get_extendr_api_entry(document: &mut Document) -> Option<&mut Value> {
 }
 
 #[allow(non_snake_case)]
-fn get_libR_sys_entry(document: &mut Document) -> Option<&mut Value> {
+fn get_libR_sys_entry(document: &mut DocumentMut) -> Option<&mut Value> {
     document
         .get_mut("patch")?
         .get_mut("crates-io")?
