### 2024-03-16 7365b52
xtask: add `--snapshot-accept` / `-a` option to subcommand `cargo extendr devtools-test` that helps accept newly formed macro-snapshots.
