+++ xtask/src/main.rs
@@ -36,9 +36,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         )?,
         cli::Commands::Doc => commands::generate_docs::run(&shell)?,
         cli::Commands::Msrv => commands::cargo_msrv::run(&shell)?,
-        cli::Commands::DevtoolsTest => commands::devtools_test::run(&shell)?,
-        // TODO: Accept snapshots from command-line
-        // Rscript -e "testthat::snapshot_accept(path = 'tests/extendrtests/tests/testthat', 'macro-snapshot')"
+        cli::Commands::DevtoolsTest(args) => commands::devtools_test::run(&shell, args)?,
     };
 
     Ok(())
