+++ xtask/src/cli/devtools_test.rs
@@ -1,0 +1,12 @@
+use clap::Args;
+
+#[derive(Args, Debug)]
+pub(crate) struct DevtoolsTestArg {
+    #[arg(
+        long,
+        short = 'a',
+        default_value = "false",
+        help = "Accept newly generated macro-snapshots"
+    )]
+    pub(crate) snapshot_accept: bool,
+}
