### 2024-03-16 fce8dd6
xtask: added a `cargo extendr document` command that generates wrappers for `tests/extendrtests`, in case they are missing
