+++ xtask/src/commands/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod cargo_msrv;
 pub(crate) mod devtools_test;
 pub(crate) mod generate_docs;
 pub(crate) mod r_cmd_check;
+pub(crate) mod rextendr_document;
