### 2024-03-24 7dc11ab
fix ci: r-devel now uses rtools44 by default (#711)
