+++ trait.impl/syn/parse/trait.Parse.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_macros":[["impl <a class=\"trait\" href=\"https://docs.rs/syn/2.0.52/syn/parse/trait.Parse.html\" title=\"trait syn::parse::Parse\">Parse</a> for <a class=\"struct\" href=\"extendr_macros/call/struct.Call.html\" title=\"struct extendr_macros::call::Call\">Call</a>"],["impl <a class=\"trait\" href=\"https://docs.rs/syn/2.0.52/syn/parse/trait.Parse.html\" title=\"trait syn::parse::Parse\">Parse</a> for <a class=\"struct\" href=\"extendr_macros/extendr_module/struct.Module.html\" title=\"struct extendr_macros::extendr_module::Module\">Module</a>"],["impl <a class=\"trait\" href=\"https://docs.rs/syn/2.0.52/syn/parse/trait.Parse.html\" title=\"trait syn::parse::Parse\">Parse</a> for <a class=\"struct\" href=\"extendr_macros/pairs/struct.Pairs.html\" title=\"struct extendr_macros::pairs::Pairs\">Pairs</a>"]]
+"extendr_macros":[["impl <a class=\"trait\" href=\"https://docs.rs/syn/2.0.55/syn/parse/trait.Parse.html\" title=\"trait syn::parse::Parse\">Parse</a> for <a class=\"struct\" href=\"extendr_macros/call/struct.Call.html\" title=\"struct extendr_macros::call::Call\">Call</a>"],["impl <a class=\"trait\" href=\"https://docs.rs/syn/2.0.55/syn/parse/trait.Parse.html\" title=\"trait syn::parse::Parse\">Parse</a> for <a class=\"struct\" href=\"extendr_macros/extendr_module/struct.Module.html\" title=\"struct extendr_macros::extendr_module::Module\">Module</a>"],["impl <a class=\"trait\" href=\"https://docs.rs/syn/2.0.55/syn/parse/trait.Parse.html\" title=\"trait syn::parse::Parse\">Parse</a> for <a class=\"struct\" href=\"extendr_macros/pairs/struct.Pairs.html\" title=\"struct extendr_macros::pairs::Pairs\">Pairs</a>"]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
