### 2024-03-24 e6418ec
ci: upload r-cmd-check logs after failure
