### 2024-03-24 4645620
ci: added better name for the check-dir upload
