### 2024-03-24 13cc7f2
Add faer to full-functionality and test lists
