### 2024-03-24 beca459
Remove unnecessary matrix check
