### 2024-03-24 b4b9846
ci: Upload results of r-cmd-check (#712)
* ci: upload r-cmd-check logs after failure

* ci: added better name for the check-dir upload
