### 2024-03-24 e6bfffe
add assertions and reduce the size of the matrix that is printed
