### 2024-03-24 1e514b5
refactor: use lifetime formulation
