### 2024-03-24 b40d583
Refactor with is_matrix()
