### 2024-03-25 7de7c10
Initialize empty `RMatrix` (#654)
* this compiles

* Initialize empty matrices

* remove this

* `cargo fmt`

* added `Rbyte`..

* Added two variants, one with random entries, and one with NA.
the NA one mimics `matrix(nrow=10, ncol=10` in R.

* Noticed soundness issue

* forgot to check that in...

* `cargo fmt`?

* update doc strings

* remove 'empty' from doc

* add assertions and reduce the size of the matrix that is printed

* cargo fmt

* refactor: use lifetime formulation

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
