+++ extendr-api/tests/dataframe_tests.rs
@@ -29,3 +29,25 @@ fn test_derive_into_dataframe() {
         assert_eq!(list[1], r!(["0", "1"]));
     }
 }
+
+#[test]
+fn test_into_robj_dataframe() {
+    test! {
+        use extendr_api::prelude::*;
+
+        #[derive(Clone, Debug, IntoDataFrameRow)]
+        struct MyStruct {
+            x: Rint,
+            y: Rstr,
+        }
+
+        let v = vec![MyStruct { x: 0.into(), y: "abc".into() }, MyStruct { x: 1.into(), y: "xyz".into() }];
+        let df = v.into_dataframe()?;
+
+        assert_eq!(
+            df.clone().as_robj().clone(),
+            df.into_robj()
+        );
+
+    }
+}
