+++ extendr-api/src/wrapper/matrix.rs
@@ -50,10 +50,8 @@ where
         let sexptype = T::sexptype();
         let matrix = Robj::alloc_matrix(sexptype, nrow as _, ncol as _);
         // this is pretty much the same?
-        let mut robj = matrix;
-        let slice = robj.as_typed_slice_mut().unwrap();
-        let data = slice.as_mut_ptr();
-        RArray::from_parts(robj, data, [nrow, ncol])
+        let robj = matrix;
+        RArray::from_parts(robj, [nrow, ncol])
     }
 }
 
