+++ extendr-api/src/wrapper/matrix.rs
@@ -50,7 +50,9 @@ where
         let sexptype = T::sexptype();
         let matrix = Robj::alloc_matrix(sexptype, nrow as _, ncol as _);
         // this is pretty much the same?
-        let robj = matrix;
+        let mut robj = matrix;
+        let slice = robj.as_typed_slice_mut().unwrap();
+        let data = slice.as_mut_ptr();
         RArray::from_parts(robj, [nrow, ncol])
     }
 }
