### 2024-03-30 628a3e2
Remove `data`-field from `RArray` (#657)
* Removed the weird `data`-field pointer

* forgot to push this.. we
only need to do type-checking with the slice

* remove dead code

* fix: `new` now uses the new definition

* update changelog

* Update CHANGELOG.md

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
