+++ extendr-api/src/wrapper/matrix.rs
@@ -29,11 +29,9 @@ pub struct RArray<T, D> {
     /// Owning Robj (probably should be a Pin).
     robj: Robj,
 
-    /// Slice of the data references the Robj.
-    data: *mut T,
-
     /// Dimensions of the array.
     dim: D,
+    _data: std::marker::PhantomData<T>,
 }
 
 pub type RColumn<T> = RArray<T, [usize; 1]>;
@@ -55,7 +53,7 @@ where
         let mut robj = matrix;
         let slice = robj.as_typed_slice_mut().unwrap();
         let data = slice.as_mut_ptr();
-        RArray::from_parts(robj, data, [nrow, ncol])
+        RArray::from_parts(robj, [nrow, ncol])
     }
 }
 
@@ -131,14 +129,27 @@ impl<T> Offset<[usize; 3]> for RArray<T, [usize; 3]> {
     }
 }
 
-impl<T, D> RArray<T, D> {
-    pub fn from_parts(robj: Robj, data: *mut T, dim: D) -> Self {
-        Self { robj, data, dim }
+impl<'a, T, D> RArray<T, D>
+where
+    T: 'a,
+    Robj: AsTypedSlice<'a, T>,
+{
+    pub fn from_parts(robj: Robj, dim: D) -> Self {
+        Self {
+            robj,
+            dim,
+            _data: std::marker::PhantomData,
+        }
     }
 
-    /// Get the underlying data fro this array.
-    pub fn data(&self) -> &[T] {
-        unsafe { std::slice::from_raw_parts(self.data, self.robj.len()) }
+    /// Returns a flat representation of the array in col-major.
+    pub fn data(&self) -> &'a [T] {
+        self.as_typed_slice().unwrap()
+    }
+
+    /// Returns a flat mutable representation of the array in col-major.
+    pub fn data_mut(&mut self) -> &'a mut [T] {
+        self.as_typed_slice_mut().unwrap()
     }
 
     /// Get the dimensions for this array.
@@ -155,10 +166,8 @@ where
     pub fn new_column<F: FnMut(usize) -> T>(nrows: usize, f: F) -> Self {
         let mut robj = (0..nrows).map(f).collect_robj();
         let dim = [nrows];
-        let mut robj = robj.set_attrib(wrapper::symbol::dim_symbol(), dim).unwrap();
-        let slice = robj.as_typed_slice_mut().unwrap();
-        let data = slice.as_mut_ptr();
-        RArray::from_parts(robj, data, dim)
+        let robj = robj.set_attrib(wrapper::symbol::dim_symbol(), dim).unwrap();
+        RArray::from_parts(robj, dim)
     }
 
     /// Get the number of rows.
@@ -194,9 +203,8 @@ where
             })
             .collect_robj();
         let dim = [nrows, ncols];
-        let mut robj = robj.set_attrib(wrapper::symbol::dim_symbol(), dim).unwrap();
-        let data = robj.as_typed_slice_mut().unwrap().as_mut_ptr();
-        RArray::from_parts(robj, data, dim)
+        let robj = robj.set_attrib(wrapper::symbol::dim_symbol(), dim).unwrap();
+        RArray::from_parts(robj, dim)
     }
 
     /// Get the number of rows.
@@ -230,9 +238,8 @@ where
             })
             .collect_robj();
         let dim = [nrows, ncols, nmatrix];
-        let mut robj = robj.set_attrib(wrapper::symbol::dim_symbol(), dim).unwrap();
-        let data = robj.as_typed_slice_mut().unwrap().as_mut_ptr();
-        RArray::from_parts(robj, data, dim)
+        let robj = robj.set_attrib(wrapper::symbol::dim_symbol(), dim).unwrap();
+        RArray::from_parts(robj, dim)
     }
 
     /// Get the number of rows.
[…]
