+++ trait.impl/core/panic/unwind_safe/trait.UnwindSafe.js
@@ -1,5 +1,5 @@
 (function() {var implementors = {
-"extendr_api":[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/panic/unwind_safe/trait.UnwindSafe.html\" title=\"trait core::panic::unwind_safe::UnwindSafe\">UnwindSafe</a> for <a class=\"enum\" href=\"extendr_api/enum.Rtype.html\" title=\"enum extendr_api::Rtype\">Rtype</a>",1,["extendr_api::Rtype"]],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/panic/unwind_safe/trait.UnwindSafe.html\" title=\"trait core::panic::unwind_safe::UnwindSafe\">UnwindSafe</a> for <a class=\"enum\" href=\"extendr_api/error/enum.Error.html\" title=\"enum extendr_api::error::Error\">Error</a>",1,["extendr_api::error::Error"]],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/panic/unwind_safe/trait.UnwindSafe.html\" title=\"trait core::panic::unwind_safe::UnwindSafe\">UnwindSafe</a> for <a class=\"enum\" href=\"extendr_api/graphics/device_descriptor/enum.CanHAdjOption.html\" title=\"enum extendr_api::graphics::device_descriptor::CanHAdjOption\">CanHAdjOption</a>",1,["extendr_api::graphics::device_descriptor::CanHAdjOption"]],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/panic/unwind_safe/trait.UnwindSafe.html\" title=\"trait core::panic::unwind_safe::UnwindSafe\">UnwindSafe</a> for <a class=\"enum\" href=\"extendr_api/graphics/device_descriptor/enum.DevCapCapture.html\" title=\"enum extendr_api::graphics::device_descriptor::DevCapCapture\">DevCapCapture</a>",1,["extendr_api::graphics::device_descriptor::DevCapCapture"]],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/panic/unwind_safe/trait.UnwindSafe.html\" title=\"trait core::panic::unwind_safe::UnwindSafe\">UnwindSafe</a> for <a class=\"enum\" href=\"extendr_api/graphics/device_descriptor/enum.DevCapLocator.html\" title=\"enum extendr_api::graphics::device_descriptor::DevCapLocator\">DevCapLocator</a>",1,["extendr_api::graphics::device_descriptor::DevCapLocator"]],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/panic/unwind_safe/trait.UnwindSafe.html\" title=\"trait core::panic::unwind_safe::UnwindSafe\">UnwindSafe</a> for <a class=\"enum\" href=\"extendr_api/graphics/device_descriptor/enum.DevCapRaster.html\" title=\"enum extendr_api::graphics::device_descriptor::DevCapRaster\">DevCapRaster</a>",1,["extendr_api::graphics::device_descriptor::DevCapRaster"]],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/panic/unwind_safe/trait.UnwindSafe.html\" title=\"trait core::panic::unwind_safe::UnwindSafe\">UnwindSafe</a> for <a class=\"enum\" href=\"extendr_api/graphics/device_descriptor/enum.DevCapTransparency.html\" title=\"enum extendr_api::graphics::device_descriptor::DevCapTransparency\">DevCapTransparency</a>",1,["extendr_api::grap
[…]
