+++ type.impl/std/primitive.usize.js
@@ -1,3 +1,3 @@
 (function() {var type_impls = {
-"extendr_api":[["<details class=\"toggle implementors-toggle\" open><summary><section id=\"impl-IntoDimension-for-usize\" class=\"impl\"><a class=\"src rightside\" href=\"https://docs.rs/ndarray/0.15/src/ndarray/dimension/conversion.rs.html#47\">source</a><a href=\"#impl-IntoDimension-for-usize\" class=\"anchor\">§</a><h3 class=\"code-header\">impl <a class=\"trait\" href=\"extendr_api/prelude/trait.IntoDimension.html\" title=\"trait extendr_api::prelude::IntoDimension\">IntoDimension</a> for <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a></h3></section></summary><div class=\"impl-items\"><section id=\"associatedtype.Dim\" class=\"associatedtype trait-impl\"><a href=\"#associatedtype.Dim\" class=\"anchor\">§</a><h4 class=\"code-header\">type <a href=\"extendr_api/prelude/trait.IntoDimension.html#associatedtype.Dim\" class=\"associatedtype\">Dim</a> = <a class=\"struct\" href=\"extendr_api/prelude/struct.Dim.html\" title=\"struct extendr_api::prelude::Dim\">Dim</a>&lt;[<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a>; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.array.html\">1</a>]&gt;</h4></section><section id=\"method.into_dimension\" class=\"method trait-impl\"><a class=\"src rightside\" href=\"https://docs.rs/ndarray/0.15/src/ndarray/dimension/conversion.rs.html#50\">source</a><a href=\"#method.into_dimension\" class=\"anchor\">§</a><h4 class=\"code-header\">fn <a href=\"extendr_api/prelude/trait.IntoDimension.html#tymethod.into_dimension\" class=\"fn\">into_dimension</a>(self) -&gt; <a class=\"struct\" href=\"extendr_api/prelude/struct.Dim.html\" title=\"struct extendr_api::prelude::Dim\">Dim</a>&lt;[<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a>; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.array.html\">1</a>]&gt;</h4></section></div></details>","IntoDimension","extendr_api::prelude::Ix"],["<details class=\"toggle implementors-toggle\" open><summary><section id=\"impl-ToVectorValue-for-usize\" class=\"impl\"><a class=\"src rightside\" href=\"src/extendr_api/robj/into_robj.rs.html#254\">source</a><a href=\"#impl-ToVectorValue-for-usize\" class=\"anchor\">§</a><h3 class=\"code-header\">impl <a class=\"trait\" href=\"extendr_api/robj/into_robj/trait.ToVectorValue.html\" title=\"trait extendr_api::robj::into_robj::ToVectorValue\">ToVectorValue</a> for <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a></h3></section></summary><div class=\"impl-items\"><section id=\"method.sexptype\" class=\"method trait-impl\"><a class=\"src rightside\" href=\"src/extendr_api/robj/into_robj.rs.html#254\">source</a><a href=\"#method.sexptype\" class=\"anchor\">§</a><h4 class=\"code-header\">fn <a href=\"extendr_api/robj/into_robj/trait.ToVectorValue.
[…]
