+++ trait.impl/core/ops/index/trait.IndexMut.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl&lt;T&gt; <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/ops/index/trait.IndexMut.html\" title=\"trait core::ops::index::IndexMut\">IndexMut</a>&lt;[<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a>; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.array.html\">2</a>]&gt; for <a class=\"struct\" href=\"extendr_api/wrapper/matrix/struct.RArray.html\" title=\"struct extendr_api::wrapper::matrix::RArray\">RArray</a>&lt;T, [<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a>; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.array.html\">2</a>]&gt;"]]
+"extendr_api":[["impl&lt;'a, T&gt; <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/ops/index/trait.IndexMut.html\" title=\"trait core::ops::index::IndexMut\">IndexMut</a>&lt;[<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a>; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.array.html\">2</a>]&gt; for <a class=\"struct\" href=\"extendr_api/wrapper/matrix/struct.RArray.html\" title=\"struct extendr_api::wrapper::matrix::RArray\">RArray</a>&lt;T, [<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a>; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.array.html\">2</a>]&gt;<div class=\"where\">where\n    T: 'a,\n    <a class=\"struct\" href=\"extendr_api/robj/struct.Robj.html\" title=\"struct extendr_api::robj::Robj\">Robj</a>: <a class=\"trait\" href=\"extendr_api/robj/trait.AsTypedSlice.html\" title=\"trait extendr_api::robj::AsTypedSlice\">AsTypedSlice</a>&lt;'a, T&gt;,</div>"]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()

[…]
