+++ type.impl/extendr_api/wrapper/matrix/struct.RArray.js
@@ -1,3 +1,3 @@
 (function() {var type_impls = {
-"extendr_api":[["<details class=\"toggle implementors-toggle\" open><summary><section id=\"impl-Debug-for-RArray%3CT,+D%3E\" class=\"impl\"><a class=\"src rightside\" href=\"src/extendr_api/wrapper/matrix.rs.html#27\">source</a><a href=\"#impl-Debug-for-RArray%3CT,+D%3E\" class=\"anchor\">§</a><h3 class=\"code-header\">impl&lt;T: <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a>, D: <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a>&gt; <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"struct\" href=\"extendr_api/wrapper/matrix/struct.RArray.html\" title=\"struct extendr_api::wrapper::matrix::RArray\">RArray</a>&lt;T, D&gt;</h3></section></summary><div class=\"impl-items\"><details class=\"toggle method-toggle\" open><summary><section id=\"method.fmt\" class=\"method trait-impl\"><a class=\"src rightside\" href=\"src/extendr_api/wrapper/matrix.rs.html#27\">source</a><a href=\"#method.fmt\" class=\"anchor\">§</a><h4 class=\"code-header\">fn <a href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html#tymethod.fmt\" class=\"fn\">fmt</a>(&amp;self, f: &amp;mut <a class=\"struct\" href=\"https://doc.rust-lang.org/nightly/core/fmt/struct.Formatter.html\" title=\"struct core::fmt::Formatter\">Formatter</a>&lt;'_&gt;) -&gt; <a class=\"type\" href=\"https://doc.rust-lang.org/nightly/core/fmt/type.Result.html\" title=\"type core::fmt::Result\">Result</a></h4></section></summary><div class='docblock'>Formats the value using the given formatter. <a href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html#tymethod.fmt\">Read more</a></div></details></div></details>","Debug","extendr_api::wrapper::matrix::RColumn","extendr_api::wrapper::matrix::RMatrix","extendr_api::wrapper::matrix::RMatrix3D"],["<details class=\"toggle implementors-toggle\" open><summary><section id=\"impl-Deref-for-RArray%3CT,+D%3E\" class=\"impl\"><a class=\"src rightside\" href=\"src/extendr_api/wrapper/matrix.rs.html#421-427\">source</a><a href=\"#impl-Deref-for-RArray%3CT,+D%3E\" class=\"anchor\">§</a><h3 class=\"code-header\">impl&lt;T, D&gt; <a class=\"trait\" href=\"extendr_api/trait.Deref.html\" title=\"trait extendr_api::Deref\">Deref</a> for <a class=\"struct\" href=\"extendr_api/wrapper/matrix/struct.RArray.html\" title=\"struct extendr_api::wrapper::matrix::RArray\">RArray</a>&lt;T, D&gt;</h3></section></summary><div class=\"impl-items\"><details class=\"toggle\" open><summary><section id=\"associatedtype.Target\" class=\"associatedtype trait-impl\"><a href=\"#associatedtype.Target\" class=\"anchor\">§</a><h4 class=\"code-header\">type <a href=\"extendr_api/tr
[…]
