### 2024-03-30 1895bfc
fixex `throw_r_error` may crash R session #716 (#718)
