+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -41,7 +41,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "hello_submodule"),
+                              format_args!("user function panicked: {0}", "hello_submodule"),
                           );
                           res
                       };
@@ -158,7 +158,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "new"),
+                              format_args!("user function panicked: {0}", "new"),
                           );
                           res
                       };
@@ -237,7 +237,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "set_a"),
+                              format_args!("user function panicked: {0}", "set_a"),
                           );
                           res
                       };
@@ -328,7 +328,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "a"),
+                              format_args!("user function panicked: {0}", "a"),
                           );
                           res
                       };
@@ -414,7 +414,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "me"),
+                              format_args!("user function panicked: {0}", "me"),
                           );
                           res
                       };
@@ -745,7 +745,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "euclidean_dist"),
+                              format_args!("user function panicked: {0}", "euclidean_dist"),
                           );
                           res
                       };
@@ -940,7 +940,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "type_aware_sum"),
+                              format_args!("user function panicked: {0}", "type_aware_sum"),
                           );
                           res
                       };
@@ -1116,7 +1116,7 @@
                       let err_string = {
                           let res = ::alloc::fmt::format(
                               format_args!(
-                                  "user function panicked: {0}\0",
+                                  "user function panicked: {0}",
                                   "raw_identifier_in_fn_args",
                               ),
                           );
@@ -1200,7 +1200,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "r#true"),
+                              format_args!("user function panicked: {0}", "r#true"),
                           );
                           res
                       };
@@ -1275,7 +1275,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}\0", "r#false"),
+                              format_args!("user function panicked: {0}", "r#false"),
                           );
                           res
                       };
@@ -1449,7 +1449,7 @@
                       let err_string = {
                           let res = ::alloc::fmt::format(
                               format_args!(
-                                  "user function panicked: {0}\0",
+                                  "user function panicked: {0}",
                                   "leak_implicit_strings",
                               ),
                           );
@@ -1538,7 +1538,7 @@
                       let err_string = {
                           let res = ::alloc::fmt::format(
                               format_args!(
-                                  "user function panicked: {0}\0",
[…]
