### 2024-03-31 dd1b050
Use error instead of panic! in extendr-macros (#634)
* Use error instead of panic! in extendr-macros

* Derive macros

* Support impl
