+++ xtask/src/cli/mod.rs
@@ -27,7 +27,7 @@ pub(crate) enum Commands {
     Msrv,
     #[command(about = "Run devtools::test() on {extendrtests}")]
     DevtoolsTest(DevtoolsTestArg),
-    #[command(about = "Generate wrappers by `rextendr::document")]
+    #[command(about = "Generate wrappers by `rextendr::document`")]
     Document,
 }
 
