### 2024-03-31 9a4b7c2
xtask: update the README
