### 2024-03-31 910c58a
xtask: added support for vendored / submodule libR-sys / rextendr, or installed ones.
