### 2024-03-31 f809caf
add comments to ownership
