### 2024-04-02 d132eef
xtask: added comment on which `rextendr` is being used
