### 2024-04-02 6075f5f
xtask: `  -a, --accept-snapshot  Accept newly generated macro-snapshots`
