### 2024-04-02 0294a00
`xtask` add more features (#719)
* fix: extendrtests points to libR-sys submodule
updated `xtask` as well

* xtask: added `extendrtests` to the `check-fmt`.
Added a `cargo extendr fmt` that formats `extendr`-crates
and `extendrtests`.

* xtask: use `DocumentMut` instead of
`Document` from toml_edit

* xtask: add `--snapshot-accept` / `-a`
option to subcommand
`cargo extendr devtools-test` that helps
accept newly formed macro-snapshots.

* xtask: just pin `toml_edit` to a minor version,
instead of an arbitrary patch-version.

* xtask: added support for vendored / submodule libR-sys / rextendr, 
[…]
