+++ xtask/commands/sidebar-items.js
@@ -1,1 +1,1 @@
-window.SIDEBAR_ITEMS = {"mod":["cargo_fmt_check","cargo_msrv","devtools_test","generate_docs","r_cmd_check"]};
+window.SIDEBAR_ITEMS = {"mod":["cargo_fmt","cargo_fmt_check","cargo_msrv","devtools_test","generate_docs","r_cmd_check","rextendr_document"]};
