+++ trait.impl/core/convert/trait.Into.js
@@ -1,3 +1,0 @@
-(function() {var implementors = {
-"xtask":[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.Into.html\" title=\"trait core::convert::Into\">Into</a>&lt;<a class=\"enum\" href=\"xtask/commands/r_cmd_check/enum.RCmdCheckErrorOn.html\" title=\"enum xtask::commands::r_cmd_check::RCmdCheckErrorOn\">RCmdCheckErrorOn</a>&gt; for <a class=\"enum\" href=\"xtask/cli/r_cmd_check/enum.ErrorOn.html\" title=\"enum xtask::cli::r_cmd_check::ErrorOn\">ErrorOn</a>"]]
-};if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
