+++ xtask/commands/rextendr_document/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"fn":["run","run_rextendr_document"]};
