+++ xtask/cli/devtools_test/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"struct":["DevtoolsTestArg"]};
