+++ xtask/cli/sidebar-items.js
@@ -1,1 +1,1 @@
-window.SIDEBAR_ITEMS = {"enum":["Commands"],"fn":["parse"],"mod":["r_cmd_check"],"struct":["Cli"]};
+window.SIDEBAR_ITEMS = {"enum":["Commands"],"fn":["parse"],"mod":["devtools_test","r_cmd_check"],"struct":["Cli"]};

[…]
