### 2024-04-05 73ff014
first basic replace of 'as' with try_from or from
