+++ extendr-api/src/wrapper/complexes.rs
@@ -32,7 +32,7 @@ impl Complexes {
     pub fn get_region(&self, index: usize, dest: &mut [Rcplx]) -> usize {
         unsafe {
             let ptr: *mut Rcomplex = dest.as_mut_ptr() as *mut Rcomplex;
-            COMPLEX_GET_REGION(self.get(), index as R_xlen_t, dest.len() as R_xlen_t, ptr) as usize
+            COMPLEX_GET_REGION(self.get(), R_xlen_t::try_from(index).unwrap() , R_xlen_t::try_from(dest.len()).unwrap(), ptr).try_into().unwrap()
         }
     }
 }
