+++ extendr-api/src/wrapper/logicals.rs
@@ -36,7 +36,7 @@ impl Logicals {
     pub fn get_region(&self, index: usize, dest: &mut [Rbool]) -> usize {
         unsafe {
             let ptr: *mut i32 = dest.as_mut_ptr() as *mut i32;
-            LOGICAL_GET_REGION(self.get(), index as R_xlen_t, dest.len() as R_xlen_t, ptr) as usize
+            LOGICAL_GET_REGION(self.get(), R_xlen_t::try_from(index).unwrap(), R_xlen_t::try_from(dest.len()).unwrap(), ptr).try_into().unwrap()
         }
     }
 }
@@ -45,7 +45,7 @@ impl Logicals {
 impl Logicals {
     pub fn set_elt(&mut self, index: usize, val: Rbool) {
         single_threaded(|| unsafe {
-            SET_INTEGER_ELT(self.get_mut(), index as R_xlen_t, val.inner());
+            SET_INTEGER_ELT(self.get_mut(), R_xlen_t::try_from(index).unwrap(), val.inner());
         })
     }
 }

[…]
