### 2024-04-06 cc08196
extendr_impl: fixed Returning `self` from `extendr`-class creates a clone? #614
