### 2024-04-06 d6b1537
extendrtests: added an integration test
