+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -107,6 +107,16 @@
                   )
               }
           }
+          #[automatically_derived]
+          impl ::core::clone::Clone for MySubmoduleClass {
+              #[inline]
+              fn clone(&self) -> MySubmoduleClass {
+                  let _: ::core::clone::AssertParamIsClone<i32>;
+                  *self
+              }
+          }
+          #[automatically_derived]
+          impl ::core::marker::Copy for MySubmoduleClass {}
           /// Class for testing (exported)
           /// @examples
           /// x <- MySubmoduleClass$new()
@@ -128,8 +138,24 @@
               fn a(&self) -> i32 {
                   self.a
               }
-              /// Method for getting one's self.
-              fn me(&self) -> &Self {
+              /// Method for getting one's (by way of a copy) self.
+              fn me_owned(&self) -> Self {
+                  *self
+              }
+              /// Method for getting one's (ref) self.
+              fn me_ref(&self) -> &Self {
+                  self
+              }
+              /// Method for getting one's (ref mut) self.
+              fn me_mut(&mut self) -> &mut Self {
+                  self
+              }
+              /// Method for getting one's ref (explicit) self.
+              fn me_explicit_ref(&self) -> &MySubmoduleClass {
+                  self
+              }
+              /// Method for getting one's ref mut (explicit) self.
+              fn me_explicit_mut(&mut self) -> &mut MySubmoduleClass {
                   self
               }
           }
@@ -381,7 +407,7 @@
           }
           #[no_mangle]
           #[allow(non_snake_case, clippy::not_unsafe_ptr_arg_deref)]
-          pub extern "C" fn wrap__MySubmoduleClass__me(
+          pub extern "C" fn wrap__MySubmoduleClass__me_owned(
               _self: extendr_api::SEXP,
           ) -> extendr_api::SEXP {
               use extendr_api::robj::*;
@@ -396,7 +422,7 @@
                               extendr_api::unwrap_or_throw(
                                       <&MySubmoduleClass>::from_robj(&_self_robj),
                                   )
-                                  .me(),
+                                  .me_owned(),
                           ),
                       )
                   })
@@ -414,7 +440,7 @@
                       drop(unwind_err);
                       let err_string = {
                           let res = ::alloc::fmt::format(
-                              format_args!("user function panicked: {0}", "me"),
+                              format_args!("user function panicked: {0}", "me_owned"),
                           );
                           res
                       };
@@ -442,7 +468,9 @@
               }
           }
           #[allow(non_snake_case)]
-          fn meta__MySubmoduleClass__me(metadata: &mut Vec<extendr_api::metadata::Func>) {
+          fn meta__MySubmoduleClass__me_owned(
+              metadata: &mut Vec<extendr_api::metadata::Func>,
+          ) {
               let mut args = <[_]>::into_vec(
                   #[rustc_box]
                   ::alloc::boxed::Box::new([
@@ -455,13 +483,349 @@
               );
               metadata
                   .push(extendr_api::metadata::Func {
-                      doc: " Method for getting one's self.",
-                      rust_name: "me",
-                      r_name: "me",
-                      mod_name: "me",
+                      doc: " Method for getting one's (by way of a copy) self.",
+                      rust_name: "me_owned",
+                      r_name: "me_owned",
+                      mod_name: "me_owned",
                       args: args,
                       return_type: "Self",
-                      func_ptr: wrap__MySubmoduleClass__me as *const u8,
+                      func_ptr: wrap__MySubmoduleClass__me_owned as *const u8,
+                      hidden: false,
+                  })
+          }
+          #[no_mangle]
[…]
