### 2024-04-06 bf2c727
`From<Dataframe<T>> for Robj`
