### 2024-04-06 3aaabd4
test: fixes `Dataframe<T>` does not implement `ToVectorValue` #715
