### 2024-04-06 95b6fb0
Update extendr-macros/src/wrappers.rs
Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
