### 2024-04-06 13e7aad
Update extendr-macros/src/wrappers.rs
Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
