### 2024-04-06 e694101
fix typo &self -> &Self [skip ci]
