### 2024-04-06 c5894a0
ignore doc test since will not compile on windows
