### 2024-04-06 e61fc0a
extendr-impl: Support returning `&Self` and `&mut Self` (#722)
* extendr_impl: fixed Returning `self` from `extendr`-class creates a clone? #614

* tests

* extendrtests: added an integration test

* Update extendr-macros/src/wrappers.rs

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>

* Update extendr-macros/src/wrappers.rs

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>

* test: added `expect_identitcal` as well

* changelog

* explained things better

* snapshot tests

* update extendr-impl doc

* fix typo &self -> &Self [skip ci]

* ignore doc test since will not compile on w
[…]
