+++ tests/extendrtests/src/rust/src/submodule.rs
@@ -8,7 +8,7 @@ fn hello_submodule() -> &'static str {
 }
 
 // Class for testing
-#[derive(Default, Debug)]
+#[derive(Default, Debug, Clone, Copy)]
 struct MySubmoduleClass {
     a: i32,
 }
@@ -26,24 +26,50 @@ impl MySubmoduleClass {
     fn new() -> Self {
         Self { a: 0 }
     }
-    
+
     /// Method for setting stuff.
     /// @param x a number
-    fn set_a(& mut self, x: i32) {
+    fn set_a(&mut self, x: i32) {
         self.a = x;
     }
-    
+
     /// Method for getting stuff.
     fn a(&self) -> i32 {
         self.a
     }
-    
-    /// Method for getting one's self.
-    fn me(&self) -> &Self {
+
+    // NOTE: Cannot move ownership, as that concept is incompatible with bridging
+    // data from R to Rust
+    // fn myself(self) -> Self {
+    //     self
+    // }
+
+    /// Method for getting one's (by way of a copy) self.
+    fn me_owned(&self) -> Self {
+        // only possible due to `derive(Clone, Copy)`
+        *self
+    }
+
+    /// Method for getting one's (ref) self.
+    fn me_ref(&self) -> &Self {
+        self
+    }
+
+    /// Method for getting one's (ref mut) self.
+    fn me_mut(&mut self) -> &mut Self {
+        self
+    }
+
+    /// Method for getting one's ref (explicit) self.
+    fn me_explicit_ref(&self) -> &MySubmoduleClass {
+        self
+    }
+
+    /// Method for getting one's ref mut (explicit) self.
+    fn me_explicit_mut(&mut self) -> &mut MySubmoduleClass {
         self
     }
 }
-
 
 // Macro to generate exports
 extendr_module! {
