+++ extendr-api/src/ownership.rs
@@ -110,7 +110,6 @@ impl Ownership {
 
     // Garbage collect the tracking structures.
     unsafe fn garbage_collect(&mut self) {
-        // println!("garbage_collect {} {}", self.cur_index, self.max_index);
         let new_size = self.cur_index * 2 + EXTRA_PRESERVATION_SIZE;
         let new_sexp = Rf_allocVector(VECSXP, new_size as R_xlen_t);
         R_PreserveObject(new_sexp);
@@ -135,7 +134,7 @@ impl Ownership {
         }
 
         R_ReleaseObject(old_sexp);
-        self.preservation.set_inner(new_sexp);
+        self.preservation = (new_sexp).into();
         self.cur_index = j;
         self.max_index = new_size;
         self.objects = new_objects;
@@ -188,7 +187,7 @@ impl Ownership {
     }
 
     pub unsafe fn unprotect(&mut self, sexp: SEXP) {
-        let sexp_usize = sexp.into();
+        let send_sexp = sexp.into();
         let Ownership {
             preservation,
             cur_index: _,
@@ -196,7 +195,7 @@ impl Ownership {
             ref mut objects,
         } = self;
 
-        let mut entry = objects.entry(sexp_usize);
+        let mut entry = objects.entry(send_sexp);
         match entry {
             Entry::Occupied(ref mut occupied) => {
                 let object = occupied.get_mut();
@@ -228,8 +227,7 @@ impl Ownership {
             ref mut objects,
         } = *self;
 
-        let sexp_usize = sexp.into();
-        let mut entry = objects.entry(sexp_usize);
+        let mut entry = objects.entry(sexp.into());
         match entry {
             Entry::Occupied(ref mut occupied) => occupied.get().refcount,
             Entry::Vacant(_) => 0,
