### 2024-04-06 fe0e4c7
Prototype IntoIntegerish
