### 2024-04-06 1d47acf
add integration tests for returning Dataframe<T> from a function
