+++ tests/extendrtests/src/rust/src/rmatrix.rs
@@ -1,0 +1,1 @@
+use extendr_api::prelude::*;
