+++ extendr-api/src/lib.rs
@@ -369,6 +369,8 @@ pub mod na;
 
 pub mod optional;
 
+pub (crate) mod conversions;
+
 pub use std::convert::{TryFrom, TryInto};
 pub use std::ops::Deref;
 pub use std::ops::DerefMut;
