### 2024-04-06 5705775
add docstring to dataframe<T>
