+++ tests/extendrtests/src/rust/src/rmatrix.rs
@@ -1,1 +1,0 @@
-use extendr_api::prelude::*;
