### 2024-04-06 1b6a582
Fix `#[extendr]` conversions into `Dataframe<T>` (#723)
* added `FromRobj` for `Dataframe<T>`.

* `From<Dataframe<T>> for Robj`

* Minor

* test: fixes `Dataframe<T>` does not implement `ToVectorValue` #715

* following advice by @Ilia-Kosenkov

* add integration tests for returning Dataframe<T> from a function

* review

* update wrappers

* update snapshots

* add docstring to dataframe<T>

* Update extendr-api/src/wrapper/dataframe.rs [skip ci]

* remove empty matrix test file

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
