+++ tests/extendrtests/R/extendr-wrappers.R
@@ -187,36 +187,90 @@ MyClassUnexported$a <- function() .Call(wrap__MyClassUnexported__a, self)
 
 #' Class for testing (exported)
 #' @examples
-#' x <- MySubmoduleClass$new()
+#' x <- MyClass$new()
 #' x$a()
 #' x$set_a(10)
 #' x$a()
 #' @export
-MySubmoduleClass <- new.env(parent = emptyenv())
+MyClassTryFrom <- new.env(parent = emptyenv())
 
-MySubmoduleClass$new <- function() .Call(wrap__MySubmoduleClass__new)
+MyClassTryFrom$new <- function() .Call(wrap__MyClassTryFrom__new)
 
-MySubmoduleClass$set_a <- function(x) invisible(.Call(wrap__MySubmoduleClass__set_a, self, x))
+MyClassTryFrom$set_a <- function(x) invisible(.Call(wrap__MyClassTryFrom__set_a, self, x))
 
-MySubmoduleClass$a <- function() .Call(wrap__MySubmoduleClass__a, self)
+MyClassTryFrom$a <- function() .Call(wrap__MyClassTryFrom__a, self)
 
-MySubmoduleClass$me_owned <- function() .Call(wrap__MySubmoduleClass__me_owned, self)
+MyClassTryFrom$me <- function() .Call(wrap__MyClassTryFrom__me, self)
 
-MySubmoduleClass$me_ref <- function() .Call(wrap__MySubmoduleClass__me_ref, self)
+MyClassTryFrom$restore_from_robj <- function(robj) .Call(wrap__MyClassTryFrom__restore_from_robj, robj)
 
-MySubmoduleClass$me_mut <- function() .Call(wrap__MySubmoduleClass__me_mut, self)
+MyClassTryFrom$get_default_value <- function(x = 42) .Call(wrap__MyClassTryFrom__get_default_value, x)
 
-MySubmoduleClass$me_explicit_ref <- function() .Call(wrap__MySubmoduleClass__me_explicit_ref, self)
-
-MySubmoduleClass$me_explicit_mut <- function() .Call(wrap__MySubmoduleClass__me_explicit_mut, self)
-
-#' @rdname MySubmoduleClass
+#' @rdname MyClassTryFrom
 #' @usage NULL
 #' @export
-`$.MySubmoduleClass` <- function (self, name) { func <- MySubmoduleClass[[name]]; environment(func) <- environment(); func }
+`$.MyClassTryFrom` <- function (self, name) { func <- MyClassTryFrom[[name]]; environment(func) <- environment(); func }
 
 #' @export
-`[[.MySubmoduleClass` <- `$.MySubmoduleClass`
+`[[.MyClassTryFrom` <- `$.MyClassTryFrom`
+
+`__MyClassTryFrom` <- new.env(parent = emptyenv())
+
+`__MyClassTryFrom`$new <- function() .Call(wrap____MyClassTryFrom__new)
+
+`__MyClassTryFrom`$`__name_test` <- function() invisible(.Call(wrap____MyClassTryFrom____name_test, self))
+
+#' @export
+`$.__MyClassTryFrom` <- function (self, name) { func <- `__MyClassTryFrom`[[name]]; environment(func) <- environment(); func }
+
+#' @export
+`[[.__MyClassTryFrom` <- `$.__MyClassTryFrom`
+
+#' Class for testing (unexported)
+MyClassUnexportedTryFrom <- new.env(parent = emptyenv())
+
+MyClassUnexportedTryFrom$new <- function() .Call(wrap__MyClassUnexportedTryFrom__new)
+
[…]

[…]
