+++ tests/extendrtests/man/MySubmoduleClass.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\docType{data}
+\name{MySubmoduleClass}
+\alias{MySubmoduleClass}
+\alias{$.MySubmoduleClass}
+\title{Class for testing (exported)}
+\format{
+An object of class \code{environment} of length 8.
+}
+\usage{
+MySubmoduleClass
+}
+\description{
+Class for testing (exported)
+}
+\examples{
+x <- MySubmoduleClass$new()
+x$a()
+x$set_a(10)
+x$a()
+}
+\keyword{datasets}
