+++ tests/extendrtests/NAMESPACE
@@ -6,6 +6,7 @@ S3method("$",MyClass)
 S3method("$",MyClassTryFrom)
 S3method("$",MyClassUnexported)
 S3method("$",MyClassUnexportedTryFrom)
+S3method("$",MySubmoduleClass)
 S3method("$",MySubmoduleClassTryFrom)
 S3method("[[","__MyClass")
 S3method("[[","__MyClassTryFrom")
@@ -13,10 +14,12 @@ S3method("[[",MyClass)
 S3method("[[",MyClassTryFrom)
 S3method("[[",MyClassUnexported)
 S3method("[[",MyClassUnexportedTryFrom)
+S3method("[[",MySubmoduleClass)
 S3method("[[",MySubmoduleClassTryFrom)
 export("__00__special_function_name")
 export(MyClass)
 export(MyClassTryFrom)
+export(MySubmoduleClass)
 export(MySubmoduleClassTryFrom)
 export(euclidean_dist)
 export(false)
