+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -53,8 +53,8 @@
                   }
               }
           }
-          impl From<VecUsize> for Robj {
-              fn from(value: VecUsize) -> Self {
+          impl<'a> From<&'a VecUsize> for Robj {
+              fn from(value: &'a VecUsize) -> Self {
                   unsafe {
                       let ptr = Box::into_raw(Box::new(value));
                       let mut res = Robj::make_external_ptr(ptr, Robj::from(()));
@@ -64,8 +64,8 @@
                   }
               }
           }
-          impl<'a> From<&'a VecUsize> for Robj {
-              fn from(value: &'a VecUsize) -> Self {
+          impl From<VecUsize> for Robj {
+              fn from(value: VecUsize) -> Self {
                   unsafe {
                       let ptr = Box::into_raw(Box::new(value));
                       let mut res = Robj::make_external_ptr(ptr, Robj::from(()));
@@ -123,13 +123,18 @@
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
                   let _robj_robj = extendr_api::robj::Robj::from_sexp(robj);
-                  std::panic::catch_unwind(|| -> std::result::Result<Robj, extendr_api::Error> {
-                      Ok(
-                          extendr_api::Robj::from(
-                              new_usize(<Integers>::from_robj(&_robj_robj)?),
-                          ),
-                      )
-                  })
+                  std::panic::catch_unwind(
+                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                          Robj,
+                          extendr_api::Error,
+                      > {
+                          Ok(
+                              extendr_api::Robj::from(
+                                  new_usize(<Integers>::from_robj(&_robj_robj)?),
+                              ),
+                          )
+                      }),
+                  )
               };
               match wrap_result_state {
                   Ok(Ok(zz)) => {
@@ -246,9 +251,12 @@
                   std::result::Result<Robj, extendr_api::Error>,
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
-                  std::panic::catch_unwind(|| -> std::result::Result<Robj, extendr_api::Error> {
-                      Ok(extendr_api::Robj::from(tst_altstring()))
-                  })
+                  std::panic::catch_unwind(
+                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                          Robj,
+                          extendr_api::Error,
+                      > { Ok(extendr_api::Robj::from(tst_altstring())) }),
+                  )
               };
               match wrap_result_state {
                   Ok(Ok(zz)) => {
@@ -373,9 +381,12 @@
                   std::result::Result<Robj, extendr_api::Error>,
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
-                  std::panic::catch_unwind(|| -> std::result::Result<Robj, extendr_api::Error> {
-                      Ok(extendr_api::Robj::from(tst_altinteger()))
-                  })
+                  std::panic::catch_unwind(
+                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                          Robj,
+                          extendr_api::Error,
+                      > { Ok(extendr_api::Robj::from(tst_altinteger())) }),
+                  )
               };
               match wrap_result_state {
                   Ok(Ok(zz)) => {
@@ -605,9 +616,12 @@
                   std::result::Result<Robj, extendr_api::Error>,
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
-                  std::panic::catch_unwind(|| -> std::result::Result<Robj, extendr_api::Error> {
-                      Ok(extendr_api::Robj::from(test_derive_into_dataframe()))
-                  })
+                  std::panic::catch_unwind(
[…]
