+++ extendr-api/src/conversions/mod.rs
@@ -1,1 +1,1 @@
-pub (crate) mod from_integerish;
+pub(crate) mod try_into_int;
