### 2024-04-08 79544ca
Refactor from_robj conversions
