### 2024-04-08 a8e3bef
Update bool converter as well
