### 2024-04-09 1506480
Delegate implementation
