+++ extendr-api/src/scalar/rfloat.rs
@@ -183,13 +183,12 @@ impl TryFrom<&Robj> for Rfloat {
     type Error = Error;
 
     fn try_from(robj: &Robj) -> Result<Self> {
-        let f64_val : Result<f64> = robj.try_into();
-        match f64_val
-        {
+        let f64_val: Result<f64> = robj.try_into();
+        match f64_val {
             Ok(val) => Ok(Rfloat::from(val)),
             // TODO: Currently this results in an extra protection of robj
             Err(Error::MustNotBeNA(_)) => Ok(Rfloat::na()),
-            Err(e) => Err(e)
+            Err(e) => Err(e),
         }
     }
 }
