### 2024-04-09 fbb5647
Fix error display
