+++ extendr-api/Cargo.toml
@@ -23,7 +23,7 @@ serde = { version = "1.0", features = ["derive"], optional = true }
 
 [dev-dependencies]
 extendr-engine = { path = "../extendr-engine" }
-rstest = "0.18.1"
+rstest = "0.19.0"
 
 [features]
 
