+++ search.desc/extendr_engine/extendr_engine-desc-0-.js
@@ -1,0 +1,1 @@
+searchState.loadedDescShard("extendr_engine", 0, "A singleton instance of the R interpreter.\nClose down the R interpreter. Note you won’t be able to …\nEnsures that an embedded R instance is present when …")
