+++ trait.impl/core/convert/trait.From.js
@@ -1,4 +1,4 @@
 (function() {var implementors = {
-"extendr_api":[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.f64.html\">f64</a>&gt; for <a class=\"struct\" href=\"extendr_api/scalar/rfloat/struct.Rfloat.html\" title=\"struct extendr_api::scalar::rfloat::Rfloat\">Rfloat</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.i32.html\">i32</a>&gt; for <a class=\"struct\" href=\"extendr_api/scalar/rbool/struct.Rbool.html\" title=\"struct extendr_api::scalar::rbool::Rbool\">Rbool</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.i32.html\">i32</a>&gt; for <a class=\"struct\" href=\"extendr_api/scalar/rint/struct.Rint.html\" title=\"struct extendr_api::scalar::rint::Rint\">Rint</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.str.html\">str</a>&gt; for <a class=\"enum\" href=\"extendr_api/error/enum.Error.html\" title=\"enum extendr_api::error::Error\">Error</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.str.html\">str</a>&gt; for <a class=\"struct\" href=\"extendr_api/wrapper/rstr/struct.Rstr.html\" title=\"struct extendr_api::wrapper::rstr::Rstr\">Rstr</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;<a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.str.html\">str</a>&gt; for <a class=\"struct\" href=\"extendr_api/wrapper/symbol/struct.Symbol.html\" title=\"struct extendr_api::wrapper::symbol::Symbol\">Symbol</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;<a class=\"struct\" href=\"extendr_api/metadata/struct.Arg.html\" title=\"struct extendr_api::metadata::Arg\">Arg</a>&gt; for <a class=\"struct\" href=\"extendr_api/metadata/struct.RArg.html\" title=\"struct extendr_api::metadata::RArg\">RArg</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/t
[…]
