+++ search.desc/xtask/xtask-desc-0-.js
@@ -1,0 +1,1 @@
+searchState.loadedDescShard("xtask", 0, "Returns the argument unchanged.\nReturns the argument unchanged.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nReturns the argument unchanged.\nCalls <code>U::from(self)</code>.\nReturns the argument unchanged.\nReturns the argument unchanged.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nThis invokes <code>rextendr::document()</code> within <code>tests/extendrtests</code>…\nGenerates documentation like on the site extendr.github.io\nReturns the argument unchanged.\nCalls <code>U::from(self)</code>.\nReturns the argument unchanged.\nCalls <code>U::from(self)</code>.")

[…]
