### 2024-04-12 e679145
replace buggy number convers with tryfrom
