### 2024-04-20 185956c
remove set_inner() method from SendSEXP
