### 2024-04-20 ce3197f
Added unsafe `SendSEXP` to `ownership`-module (#666)
* Added a `SendSEXP` instead of conversion to
usize

* `cargo fmt`.

* remove unnecessary Deref/DerefMut impl

* add comments to ownership

* typos

* `cargo fmt`

* comments

* remove set_inner() method from SendSEXP

* rename sexp_usize to send_sexp to make the type:

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
