+++ extendr-api/src/ownership.rs
@@ -17,6 +17,38 @@ use libR_sys::{
     Rf_unprotect, LENGTH, SET_VECTOR_ELT, SEXP, VECSXP, VECTOR_ELT,
 };
 
+mod send_sexp {
+    //! Provide a wrapper around R's pointer type `SEXP` that is `Send`.
+    //!
+    //! This can lead to soundness issues, therefore accessing the `SEXP` has
+    //! to happen through the unsafe method [`SendSEXP::inner`].
+    //!
+    use libR_sys::SEXP;
+
+    /// A wrapper around R's pointer type `SEXP` that is `Send`.
+    #[derive(Debug, Clone, PartialEq, Eq, Hash)]
+    pub struct SendSEXP(SEXP);
+
+    impl From<SEXP> for SendSEXP {
+        fn from(value: SEXP) -> Self {
+            Self(value)
+        }
+    }
+
+    // Allows SendSEXP to be sent between threads even though unsafe
+    // Requires that the SEXP is not accessed concurrently.
+    unsafe impl Send for SendSEXP {}
+
+    impl SendSEXP {
+        /// Get the inner `SEXP`
+        pub unsafe fn inner(&self) -> SEXP {
+            self.0
+        }
+    }
+}
+
+use self::send_sexp::SendSEXP;
+
 static OWNERSHIP: Lazy<Mutex<Ownership>> = Lazy::new(|| Mutex::new(Ownership::new()));
 
 pub(crate) unsafe fn protect(sexp: SEXP) {
@@ -32,15 +64,20 @@ pub(crate) unsafe fn unprotect(sexp: SEXP) {
 pub const INITIAL_PRESERVATION_SIZE: usize = 100000;
 pub const EXTRA_PRESERVATION_SIZE: usize = 100000;
 
+// `Object` is a manual reference counting mechanism that is used for each SEXP.
+// `refcount` is the number of times the SEXP is accessed.
+// `index` is the index of the SEXP in the preservation vector.
+#[derive(Debug)]
 struct Object {
     refcount: usize,
     index: usize,
 }
 
 // A reference counted object with an index in the preservation vector.
+#[derive(Debug)]
 struct Ownership {
     // A growable vector containing all owned objects.
-    preservation: usize,
+    preservation: SendSEXP,
 
     // An incrementing count of objects through the vector.
     cur_index: usize,
@@ -49,7 +86,7 @@ struct Ownership {
     max_index: usize,
 
     // A hash map from SEXP address to object.
-    objects: HashMap<usize, Object>,
+    objects: HashMap<SendSEXP, Object>,
 }
 
 impl Ownership {
@@ -58,7 +95,7 @@ impl Ownership {
             let preservation = Rf_allocVector(VECSXP, INITIAL_PRESERVATION_SIZE as R_xlen_t);
             R_PreserveObject(preservation);
             Ownership {
-                preservation: preservation as usize,
+                preservation: preservation.into(),
                 cur_index: 0,
                 max_index: INITIAL_PRESERVATION_SIZE,
                 objects: HashMap::with_capacity(INITIAL_PRESERVATION_SIZE),
@@ -66,14 +103,54 @@ impl Ownership {
         }
     }
 
+    // Garbage collect the tracking structures.
+    unsafe fn garbage_collect(&mut self) {
+        let new_size = self.cur_index * 2 + EXTRA_PRESERVATION_SIZE;
+        let new_sexp = Rf_allocVector(VECSXP, new_size as R_xlen_t);
+        R_PreserveObject(new_sexp);
+        let old_sexp = self.preservation.inner();
+
+        let mut new_objects = HashMap::with_capacity(new_size);
+
+        // copy non-null elements to new vector and hashmap.
+        let mut j = 0;
+        for (addr, object) in self.objects.iter() {
[…]
