+++ trait.impl/core/fmt/trait.Debug.js
@@ -1,5 +1,5 @@
 (function() {var implementors = {
-"extendr_api":[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"enum\" href=\"extendr_api/enum.Rtype.html\" title=\"enum extendr_api::Rtype\">Rtype</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"enum\" href=\"extendr_api/error/enum.Error.html\" title=\"enum extendr_api::error::Error\">Error</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"enum\" href=\"extendr_api/graphics/enum.FontFace.html\" title=\"enum extendr_api::graphics::FontFace\">FontFace</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"enum\" href=\"extendr_api/graphics/enum.LineEnd.html\" title=\"enum extendr_api::graphics::LineEnd\">LineEnd</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"enum\" href=\"extendr_api/graphics/enum.LineJoin.html\" title=\"enum extendr_api::graphics::LineJoin\">LineJoin</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"enum\" href=\"extendr_api/graphics/enum.LineType.html\" title=\"enum extendr_api::graphics::LineType\">LineType</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"enum\" href=\"extendr_api/graphics/enum.Unit.html\" title=\"enum extendr_api::graphics::Unit\">Unit</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"struct\" href=\"extendr_api/graphics/struct.Device.html\" title=\"struct extendr_api::graphics::Device\">Device</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"struct\" href=\"extendr_api/graphics/struct.Pattern.html\" title=\"struct extendr_api::graphics::Pattern\">Pattern</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> for <a class=\"struct\" href=\"extendr_api/graphics/struct.TextMetric.html\" title=\"struct extendr_api::graphics::TextMetric\">TextMetric</a>"],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/n
[…]
