+++ trait.impl/serde/ser/trait.SerializeStructVariant.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl&lt;'a&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.197/serde/ser/trait.SerializeStructVariant.html\" title=\"trait serde::ser::SerializeStructVariant\">SerializeStructVariant</a> for <a class=\"struct\" href=\"extendr_api/serializer/struct.SerializeStructVariant.html\" title=\"struct extendr_api::serializer::SerializeStructVariant\">SerializeStructVariant</a>&lt;'a&gt;"]]
+"extendr_api":[["impl&lt;'a&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.198/serde/ser/trait.SerializeStructVariant.html\" title=\"trait serde::ser::SerializeStructVariant\">SerializeStructVariant</a> for <a class=\"struct\" href=\"extendr_api/serializer/struct.SerializeStructVariant.html\" title=\"struct extendr_api::serializer::SerializeStructVariant\">SerializeStructVariant</a>&lt;'a&gt;"]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
