+++ extendr_api/ownership/send_sexp/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"struct":["SendSEXP"]};

[…]
