### 2024-04-20 338f7e4
Add `TryFrom` for `ExternalPtr` (#730)
* this doesn't work, but it should work...

* make it work

* finally, shaping up

* macro-snapshot

* update changelog

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
