+++ tests/extendrtests/tests/testthat/test-test-submodules.R
@@ -21,3 +21,23 @@ test_that("Classes defined in submodules also works", {
     isTRUE(identical(x$me_owned(), x))
   )
 })
+
+test_that("Classes defined in submodules also works (use-try-from)", {
+  x <- MySubmoduleClassTryFrom$new()
+  x$set_a(10)
+  expect_equal(x$a(), 10)
+
+  expect_equal(x$me_ref(), x)
+  expect_equal(x$me_mut(), x)
+  expect_equal(x$me_explicit_ref(), x)
+  expect_equal(x$me_explicit_mut(), x)
+
+  expect_identical(x$me_ref(), x)
+  expect_identical(x$me_mut(), x)
+  expect_identical(x$me_explicit_ref(), x)
+  expect_identical(x$me_explicit_mut(), x)
+  # this "copies"
+  expect_false(
+    isTRUE(identical(x$me_owned(), x))
+  )
+})
