+++ tests/extendrtests/man/MySubmoduleClassTryFrom.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\docType{data}
+\name{MySubmoduleClassTryFrom}
+\alias{MySubmoduleClassTryFrom}
+\alias{$.MySubmoduleClassTryFrom}
+\title{Class for testing (exported)}
+\format{
+An object of class \code{environment} of length 8.
+}
+\usage{
+MySubmoduleClassTryFrom
+}
+\description{
+Class for testing (exported)
+}
+\examples{
+x <- MySubmoduleClassTryFrom$new()
+x$a()
+x$set_a(10)
+x$a()
+}
+\keyword{datasets}
