+++ tests/extendrtests/NAMESPACE
@@ -1,16 +1,26 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method("$","__MyClass")
+S3method("$","__MyClassTryFrom")
 S3method("$",MyClass)
+S3method("$",MyClassTryFrom)
 S3method("$",MyClassUnexported)
+S3method("$",MyClassUnexportedTryFrom)
 S3method("$",MySubmoduleClass)
+S3method("$",MySubmoduleClassTryFrom)
 S3method("[[","__MyClass")
+S3method("[[","__MyClassTryFrom")
 S3method("[[",MyClass)
+S3method("[[",MyClassTryFrom)
 S3method("[[",MyClassUnexported)
+S3method("[[",MyClassUnexportedTryFrom)
 S3method("[[",MySubmoduleClass)
+S3method("[[",MySubmoduleClassTryFrom)
 export("__00__special_function_name")
 export(MyClass)
+export(MyClassTryFrom)
 export(MySubmoduleClass)
+export(MySubmoduleClassTryFrom)
 export(euclidean_dist)
 export(false)
 export(hello_submodule)

[…]
