+++ extendr-api/src/optional/faer.rs
@@ -1,68 +1,66 @@
 use faer::{mat, Mat, MatRef};
 
+use crate::scalar::Rfloat;
 use crate::scalar::Scalar;
 use crate::*;
 
+/// Convert a `faer::Mat<f64>` into an `RMatrix<f64>` which is not NA aware.
+impl From<Mat<f64>> for RMatrix<f64> {
+    fn from(value: Mat<f64>) -> Self {
+        RMatrix::new_matrix(value.nrows(), value.ncols(), |i, j| value.read(i, j))
+    }
+}
+
 impl From<Mat<f64>> for Robj {
-    /// Convert a faer Mat<f64> into Robj.
-    fn from(mat: Mat<f64>) -> Self {
-        mat.col_chunks(1)
-            .flat_map(|c| c.row_chunks(1).map(|r| r.read(0, 0)))
-            .collect_robj()
+    fn from(value: Mat<f64>) -> Self {
+        RMatrix::<f64>::from(value).into_robj()
+    }
+}
+
+/// Convert a `faer::Mat<f64>` into an `RMatrix<f64>` which is not NA aware.
+impl From<MatRef<'_, f64>> for RMatrix<f64> {
+    /// Convert a faer MatRef<f64> into Robj.
+    fn from(value: MatRef<'_, f64>) -> Self {
+        RMatrix::new_matrix(value.nrows(), value.ncols(), |i, j| value.read(i, j))
     }
 }
 
 impl From<MatRef<'_, f64>> for Robj {
-    /// Convert a faer MatRef<f64> into Robj.
-    fn from(mat: MatRef<f64>) -> Self {
-        mat.col_chunks(1)
-            .flat_map(|c| c.row_chunks(1).map(|r| r.read(0, 0)))
-            .collect_robj()
+    fn from(value: MatRef<'_, f64>) -> Self {
+        RMatrix::<f64>::from(value).into_robj()
     }
 }
 
-impl<'a> FromRobj<'a> for Mat<f64> {
-    fn from_robj(robj: &'a Robj) -> std::result::Result<Self, &'static str> {
-        if !robj.is_matrix() {
-            return Err("R object is not a matrix");
-        }
-        let dim = robj.dim().expect("robj should be a matrix");
-        let nrows = (*dim).first().expect("dimension must exist").inner() as usize;
-        let ncols = (*dim).last().expect("dimension must exist").inner() as usize;
-
-        if let Some(slice) = robj.as_real_slice() {
-            let fmat = mat::from_column_major_slice::<f64>(&slice, nrows, ncols);
-            Ok(fmat.to_owned())
-        } else if let Some(slice) = robj.as_integer_slice() {
-            let fmat = Mat::<f64>::from_fn(nrows, ncols, |i, j| {
-                if slice[i + j * nrows].is_na() {
-                    f64::NAN
-                } else {
-                    slice[i + j * nrows] as f64
-                }
-            });
-            Ok(fmat)
-        } else {
-            Err("could not convert to matrix")
-        }
+impl From<Mat<f64>> for RMatrix<Rfloat> {
+    fn from(value: Mat<f64>) -> Self {
+        RMatrix::new_matrix(value.nrows(), value.ncols(), |i, j| value.read(i, j).into())
     }
 }
 
-impl<'a> FromRobj<'a> for MatRef<'a, f64> {
-    fn from_robj(robj: &'a Robj) -> std::result::Result<Self, &'static str> {
-        if !robj.is_matrix() {
-            return Err("R object is not a matrix");
-        }
-        let dim = robj.dim().expect("robj should be a matrix");
-        let nrows = (*dim).first().expect("dimension must exist").inner() as usize;
[…]
