### 2024-04-21 b3118a7
nit: an early-return branch should also UNPROTECT() (#741)
