### 2024-04-22 #6394 @CGMossa
Hello @jbytecode! I've attempted to ping @luciorq on our official ExtendR Discord Server, and I have not had a response yet. The message was sent on 04/16/2024 7:08 PM (CET), and the link to said message is [here](https://discord.com/channels/751419551132155925/751419551132155928/1229840762666090517).

I would suggest giving the reviewers some more time to respond. At the very least, we may wait until @dccsillag has finished their review. Reviewing is hard, and especially a software package of this magnitude, so I trust that, there are a lot of effort going into processing the paper, and extendr overall.
