### 2024-04-22 #6394 @jbytecode
@CGMossa  - Thank you for reaching out to our reviewer. I also want to thank you for the suggestion. We always allocate enough time to reviewers and provide additional time when needed. However, the review process is interactive in JOSS, so it's expected that reviewers and authors interact on specific subjects continuously. Since I haven't seen progress and interaction (and at least a life signal) for a long while, it seems that the elapsed time hasn't been used for the submission, which, as you mentioned before, is a comprehensive undertaking.

Because editorial work and reviewing are voluntary, I tend to be quite lenient, but not receiving a response makes things difficult.

I'll wait to receive a sign of progress for a while.
