### 2024-04-26 #6394 @dccsillag
Sorry for the excessive delay. I still haven't finished completely, but I feel like the remaining things will be relatively minor.

Extendr solves the relevant and important problem of making R bindings of Rust code (and vice versa), akin to what can be done for other languages already (e.g., for Python with PyO3). R is very prominent language that would benefit greatly from being able to easily write natively-compiled packages, so there is a great fit and I find ExtendR to be an excellent contribution to the community.

Moreover, Extendr seems fairly well-written and be the result of much care. I am confident that it is very well-suited for JOSS.

That said, I think there are two key factors to be improved prior to acceptance: documentation and it's behaviour on errors.

#### Documentation

The documentation is in dire need of improvement. In particular, there seem to be bits of the documentation in various semi-scattered places, and with no cohesive link between them. For example, when I was getting started, I started at the GitHub README. Of course, what I really wanted from there were the docs and/or a quickstart guide. Looking for this, I fell into the docs for the *development* version of Extendr, following the link at the bottom of the README (which was the only reference to documentation that I could find). I do not know if there is a better version of these docs (the ones generated by `rustdoc`), but its current presentation causes excessive confusion.

Perhaps the most prominent case of this is the sentence at the very top of the first page, stating "See `Robj` for much of the content of this crate. `Robj` provides a safe wrapper for the R object type." The docs should not point to `Robj` at the first paragraph. After all, the basic usage of the crate is all about the `#[extendr]` and `extendr_module!` macros, and those (which are presented right in the following section) should come first. From the user's perspective, the `Robj` API is not the most important part of the crate (even if from a developer perspective it is, because it's what implements the conversions). To a new user, it is a lot of effort to have to understand the `Robj` API before diving into anything else, especially when for their first usage of Extendr `Robj` will most likely be automatically used behind-the-scenes only.

That aside, I found it rather hard to understand how I should structure a Rust crate that provides R bindings. I ended up falling on the helloextendr repo, which I found rather clunky, to the degree that I was just wishing I could generaete them automatically. Fortunately, from there I ended up falling into the vignette at https://extendr.github.io/rextendr/articles/package.html, which explains how to do precisely that. My concern is (i) how _hidden_ that vignette felt, and (ii) how _scattered_ the docs are. So far, I've located docs in the form of GitHub READMEs, rustdoc-generated docs (for the development version), R docs, R vignettes (which also include things you would want to read even if you are just working on the Rust side), and probably a couple more. While I think it's understandable to have more than one location with documentation (especially considering Extendr bridges between languages), I think it's very important to try to (i) minimize their surface area, (ii) improve the presentation of each individually so that they are each reasonably self-sufficient, and (iii) better link between them all.

#### Error management

Dealing with computation that may fail is absolutely essential, and I understand this is something that is being worked on. However, I believe the current solutions must be improved. That said, it is possible that I've misunderstood something, so please correct me if I'm wrong.

Firstly, it seems that the easily accessible `throw_r_error` function leaks memory due to not calling destructors. I agree that this function should be part of the API (seeing as it is a key part of the R API), but it absolutely should not be easy to use, as it essentially violates Rust's "guarantees" (I know, I know, leaking is technically allowed in safe Rust. But that does not mean we should not avoid it, if possible). Rust code should be "safe-by-default", ideally. Therefore, I strongly think that this function should be marked as `unsafe` -- if the function is easy to use, then users *will* use it (especially as it has the appeal of being a mechanism to escape having to pipe `Result<T,E>`s through!).
If there is a strong desire to have some functionality like it but without the `unsafe`, then maybe an alternative version could be written that first panics, unwinding, and catching the panic and then calling the current `throw_r_error`.

As for the returning `Result<T,E>` option, as I understand, it makes the Rust side panic? I did not understand what truly happens in this case (and unfortunately have not been able to play around with it yet). But it does not seem sufficient, as the error does not get displayed to the R user?

Finally, there are the two experimental options. The list option sounds quite fine, though a bit more Rust-like then R-like (and thus maybe unfamiliar to R users). The error condition option also sounds quite nice, and is a bit more R-like, to the point that I wonder if it should not be the default. Could the authors clarify on why these are "experimental"?

That said, I wonder why there is not an option to return the `T` straight-up if it's an `Ok`, and *throw* an R error (stringify and then `stop`?) with `E` it's an `Err`. After all, as far as I'm aware, that is what most R packages do, and so what an R user of a ExtendR-backed Rust package would expect.

#### Additional minor concerns

Along the way, some relatively minor concerns:

- Should `test!` really be shown in the examples in the docs? While I get why it is (because then you can easily run the examples as tests), it is not standard Rust and so leads to some confusion when being used in reference pages.
- What is the necessity of the `r!` macros for creating ExtendR objects? It feels to me like just `.into()` (along with the associated `Robj::from`) would have been sufficient and minimize the API surface, but I may have missed something (or perhaps underestimate the convenience of having a macro for this).
- From the docs: "The minimal overhead of calling an extendr function is in the ballpark of 2-4us. Returning a condition or list increases the overhead to 4-8us. Checking & handling the result on R side will likely increase overall overhead to 8-16us, depending on how efficiently the result is handled." But the numbers here certainly depends on the specific hardware, right? Maybe it would be better to clarify the kind of hardware being considered here, or to speak in terms of relations (e.g., 40% slower) and not absolute numbers. (Yes, in the case of relations it's still a bit flawed, but much better than stating raw numbers.)

Overall, I'd like to congratulate the authors for their great work so far on Extendr, and once again apologise for my tardiness.
