### 2024-04-27 d841904
replaces `SEXPTYPE` and `Rboolean` to the new enum-versions from latest `libR-sys`.
