+++ extendr-api/src/wrapper/altrep.rs
@@ -155,22 +155,22 @@ fn manifest(x: SEXP) -> SEXP {
     single_threaded(|| unsafe {
         Rf_protect(x);
         let len = XLENGTH_EX(x);
-        let data2 = Rf_allocVector(TYPEOF(x) as u32, len as R_xlen_t);
+        let data2 = Rf_allocVector(TYPEOF(x), len as R_xlen_t);
         Rf_protect(data2);
-        match TYPEOF(x) as u32 {
-            INTSXP => {
+        match TYPEOF(x) {
+            SEXPTYPE::INTSXP => {
                 INTEGER_GET_REGION(x, 0, len as R_xlen_t, INTEGER(data2));
             }
-            LGLSXP => {
+            SEXPTYPE::LGLSXP => {
                 LOGICAL_GET_REGION(x, 0, len as R_xlen_t, LOGICAL(data2));
             }
-            REALSXP => {
+            SEXPTYPE::REALSXP => {
                 REAL_GET_REGION(x, 0, len as R_xlen_t, REAL(data2));
             }
-            RAWSXP => {
+            SEXPTYPE::RAWSXP => {
                 RAW_GET_REGION(x, 0, len as R_xlen_t, RAW(data2));
             }
-            CPLXSXP => {
+            SEXPTYPE::CPLXSXP => {
                 COMPLEX_GET_REGION(x, 0, len as R_xlen_t, COMPLEX(data2));
             }
             _ => {
@@ -515,7 +515,7 @@ impl Altrep {
 
             // Use R_RegisterCFinalizerEx() and set onexit to 1 (TRUE) to invoke
             // the finalizer on a shutdown of the R session as well.
-            R_RegisterCFinalizerEx(state, Some(finalizer::<StateType>), 1);
+            R_RegisterCFinalizerEx(state, Some(finalizer::<StateType>), Rboolean::TRUE);
 
             let class_ptr = R_altrep_class_t { ptr: class.get() };
             let sexp = R_new_altrep(class_ptr, state, R_NilValue);
@@ -589,7 +589,7 @@ impl Altrep {
 
         unsafe extern "C" fn altrep_Coerce<StateType: AltrepImpl + 'static>(
             x: SEXP,
-            ty: c_int,
+            ty: SEXPTYPE,
         ) -> SEXP {
             <StateType>::coerce(x, sxp_to_rtype(ty)).get()
         }
@@ -598,14 +598,14 @@ impl Altrep {
             x: SEXP,
             deep: Rboolean,
         ) -> SEXP {
-            <StateType>::duplicate(x, deep == 1).get()
+            <StateType>::duplicate(x, deep == Rboolean::TRUE).get()
         }
 
         unsafe extern "C" fn altrep_DuplicateEX<StateType: AltrepImpl + 'static>(
             x: SEXP,
             deep: Rboolean,
         ) -> SEXP {
-            <StateType>::duplicate_ex(x, deep == 1).get()
+            <StateType>::duplicate_ex(x, deep == Rboolean::TRUE).get()
         }
 
         unsafe extern "C" fn altrep_Inspect<StateType: AltrepImpl + 'static>(
@@ -628,7 +628,7 @@ impl Altrep {
             x: SEXP,
             writeable: Rboolean,
         ) -> *mut c_void {
-            <StateType>::dataptr(x, writeable != 0) as *mut c_void
+            <StateType>::dataptr(x, writeable != Rboolean::FALSE) as *mut c_void
         }
 
         unsafe extern "C" fn altvec_Dataptr_or_null<StateType: AltrepImpl + 'static>(
@@ -748,21 +748,27 @@ impl Altrep {
                 x: SEXP,
                 narm: Rboolean,
             ) -> SEXP {
[…]
