+++ extendr-api/src/robj/mod.rs
@@ -236,8 +236,8 @@ impl Robj {
 pub trait Types: GetSexp {
     #[doc(hidden)]
     /// Get the XXXSXP type of the object.
-    fn sexptype(&self) -> u32 {
-        unsafe { TYPEOF(self.get()) as u32 }
+    fn sexptype(&self) -> SEXPTYPE {
+        unsafe { TYPEOF(self.get()) }
     }
 
     /// Get the type of an R object.
@@ -263,6 +263,7 @@ pub trait Types: GetSexp {
     /// }
     /// ```
     fn rtype(&self) -> Rtype {
+        use SEXPTYPE::*;
         match self.sexptype() {
             NILSXP => Rtype::Null,
             SYMSXP => Rtype::Symbol,
@@ -293,6 +294,7 @@ pub trait Types: GetSexp {
     }
 
     fn as_any(&self) -> Rany {
+        use SEXPTYPE::*;
         unsafe {
             match self.sexptype() {
                 NILSXP => Rany::Null(std::mem::transmute(self.as_robj())),
@@ -346,6 +348,7 @@ impl Robj {
         } else {
             unsafe {
                 let sexp = self.get();
+                use SEXPTYPE::*;
                 match self.sexptype() {
                     STRSXP => STRING_ELT(sexp, 0) == libR_sys::R_NaString,
                     INTSXP => *(INTEGER(sexp)) == libR_sys::R_NaInt,
@@ -787,6 +790,7 @@ macro_rules! make_typed_slice {
     }
 }
 
+use SEXPTYPE::*;
 make_typed_slice!(Rbool, INTEGER, LGLSXP);
 make_typed_slice!(i32, INTEGER, INTSXP);
 make_typed_slice!(u32, INTEGER, INTSXP);
@@ -818,7 +822,7 @@ pub trait Attributes: Types + Length {
         Robj: From<N> + 'a,
     {
         let name = Robj::from(name);
-        if self.sexptype() == CHARSXP {
+        if self.sexptype() == SEXPTYPE::CHARSXP {
             None
         } else {
             // FIXME: this attribute does not need protection
@@ -838,7 +842,7 @@ pub trait Attributes: Types + Length {
         Robj: From<N> + 'a,
     {
         let name = Robj::from(name);
-        if self.sexptype() == CHARSXP {
+        if self.sexptype() == SEXPTYPE::CHARSXP {
             false
         } else {
             unsafe { Rf_getAttrib(self.get(), name.get()) != R_NilValue }
@@ -1068,7 +1072,7 @@ impl PartialEq<Robj> for Robj {
             }
 
             // see https://github.com/hadley/r-internals/blob/master/misc.md
-            R_compute_identical(self.get(), rhs.get(), 16) != 0
+            R_compute_identical(self.get(), rhs.get(), 16) != Rboolean::FALSE
         }
     }
 }
