+++ extendr-api/src/functions.rs
@@ -209,12 +209,12 @@ pub fn blank_scalar_string() -> Robj {
 pub fn parse(code: &str) -> Result<Expressions> {
     single_threaded(|| unsafe {
         use libR_sys::*;
-        let mut status = 0_u32;
-        let status_ptr = &mut status as _;
+        let mut status = ParseStatus::PARSE_NULL;
+        let status_ptr = &mut status as *mut _;
         let codeobj: Robj = code.into();
         let parsed = Robj::from_sexp(R_ParseVector(codeobj.get(), -1, status_ptr, R_NilValue));
         match status {
-            1 => parsed.try_into(),
+            ParseStatus::PARSE_OK => parsed.try_into(),
             _ => Err(Error::ParseError(code.into())),
         }
     })

[…]
