### 2024-04-27 88632f3
build.rs: Added a feature for using OBJSXP vs S4SXP
