### 2024-04-28 ec02b36
added dllinfo to the altrep class registration
