### 2024-04-29 3d46447
extendr-api: remove unnecessary constraints on `ExternalPtr<T>`
