### 2024-04-29 0fb1d70
Remove unnecessary constraints on `ExternalPtr<T>` on `ExternalPtr<T>`
