### 2024-04-29 3fabaae
extendr-api: remove unnecessary constraints on `ExternalPtr<T>`
