### 2024-04-30 7d42999
test.yml: slide rtools versions forward
