### 2024-04-30 e8901a3
Various documentation updates (#748)
* typo

* doc: from the vault / fork

* fix: `cargo doc` complaint

* `cargo fmt`

* test.yml: slide rtools versions forward
