+++ trait.impl/serde/ser/trait.Error.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl <a class=\"trait\" href=\"https://docs.rs/serde/1.0.198/serde/ser/trait.Error.html\" title=\"trait serde::ser::Error\">Error</a> for <a class=\"enum\" href=\"extendr_api/error/enum.Error.html\" title=\"enum extendr_api::error::Error\">Error</a>"]]
+"extendr_api":[["impl <a class=\"trait\" href=\"https://docs.rs/serde/1.0.199/serde/ser/trait.Error.html\" title=\"trait serde::ser::Error\">Error</a> for <a class=\"enum\" href=\"extendr_api/error/enum.Error.html\" title=\"enum extendr_api::error::Error\">Error</a>"]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
