+++ trait.impl/serde/de/trait.EnumAccess.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.198/serde/de/trait.EnumAccess.html\" title=\"trait serde::de::EnumAccess\">EnumAccess</a>&lt;'de&gt; for &amp;'de <a class=\"struct\" href=\"extendr_api/robj/struct.Robj.html\" title=\"struct extendr_api::robj::Robj\">Robj</a>"]]
+"extendr_api":[["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.199/serde/de/trait.EnumAccess.html\" title=\"trait serde::de::EnumAccess\">EnumAccess</a>&lt;'de&gt; for &amp;'de <a class=\"struct\" href=\"extendr_api/robj/struct.Robj.html\" title=\"struct extendr_api::robj::Robj\">Robj</a>"]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
