+++ tests/extendrtests/src/rust/src/optional_either.rs
@@ -1,7 +1,7 @@
 use extendr_api::prelude::*;
 
 #[extendr(use_try_from = true)]
-fn type_aware_sum(input : Either<Integers, Doubles>) -> Either<Rint, Rfloat> {
+fn type_aware_sum(input: Either<Integers, Doubles>) -> Either<Rint, Rfloat> {
     match input {
         Left(left) => Left(left.iter().sum()),
         Right(right) => Right(right.iter().sum()),
