+++ extendr-api/src/deserializer.rs
@@ -802,8 +802,9 @@ impl<'de> Visitor<'de> for RobjVisitor {
             keys.push(key);
             values.push(value);
         }
-
-        Ok(List::from_values(values).set_names(keys).unwrap())
+        let mut result = List::from_values(values);
+        result.set_names(keys).unwrap();
+        Ok(result.into())
     }
 
     fn visit_unit<E>(self) -> std::result::Result<Self::Value, E>
