+++ trait.impl/core/ops/deref/trait.DerefMut.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl <a class=\"trait\" href=\"extendr_api/trait.DerefMut.html\" title=\"trait extendr_api::DerefMut\">DerefMut</a> for <a class=\"struct\" href=\"extendr_api/wrapper/complexes/struct.Complexes.html\" title=\"struct extendr_api::wrapper::complexes::Complexes\">Complexes</a>"],["impl <a class=\"trait\" href=\"extendr_api/trait.DerefMut.html\" title=\"trait extendr_api::DerefMut\">DerefMut</a> for <a class=\"struct\" href=\"extendr_api/wrapper/doubles/struct.Doubles.html\" title=\"struct extendr_api::wrapper::doubles::Doubles\">Doubles</a>"],["impl <a class=\"trait\" href=\"extendr_api/trait.DerefMut.html\" title=\"trait extendr_api::DerefMut\">DerefMut</a> for <a class=\"struct\" href=\"extendr_api/wrapper/integers/struct.Integers.html\" title=\"struct extendr_api::wrapper::integers::Integers\">Integers</a>"],["impl <a class=\"trait\" href=\"extendr_api/trait.DerefMut.html\" title=\"trait extendr_api::DerefMut\">DerefMut</a> for <a class=\"struct\" href=\"extendr_api/wrapper/logicals/struct.Logicals.html\" title=\"struct extendr_api::wrapper::logicals::Logicals\">Logicals</a>"],["impl&lt;T, D&gt; <a class=\"trait\" href=\"extendr_api/trait.DerefMut.html\" title=\"trait extendr_api::DerefMut\">DerefMut</a> for <a class=\"struct\" href=\"extendr_api/wrapper/matrix/struct.RArray.html\" title=\"struct extendr_api::wrapper::matrix::RArray\">RArray</a>&lt;T, D&gt;"],["impl&lt;T: <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html\" title=\"trait core::fmt::Debug\">Debug</a> + 'static&gt; <a class=\"trait\" href=\"extendr_api/trait.DerefMut.html\" title=\"trait extendr_api::DerefMut\">DerefMut</a> for <a class=\"struct\" href=\"extendr_api/wrapper/externalptr/struct.ExternalPtr.html\" title=\"struct extendr_api::wrapper::externalptr::ExternalPtr\">ExternalPtr</a>&lt;T&gt;"]]
[…]

[…]
