### 2024-04-30 5fcb2f9
doc tests: fixed `cargo test --doc`
