+++ extendr-api/src/metadata.rs
@@ -2,7 +2,6 @@
 //!
 //! This data is returned by get_module_metadata()
 //! which is generated by [extendr_module!].
-use crate::robj::IntoRobj;
 use crate::*;
 use std::io::Write;
 
