### 2024-05-01 0d0f899
fixed the `result_list` and `result_condition`
