+++ tests/extendrtests/man/dbls_named.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R
+\name{dbls_named}
+\alias{dbls_named}
+\title{Adds an attribute to a vector of doubles}
+\usage{
+dbls_named(x)
+}
+\description{
+Adds an attribute to a vector of doubles
+}
