+++ tests/extendrtests/R/extendr-wrappers.R
@@ -83,6 +83,9 @@ tst_altstring <- function() .Call(wrap__tst_altstring)
 tst_altinteger <- function() .Call(wrap__tst_altinteger)
 
 #' Adds an attribute to a vector of doubles
+#' 
+#' @param x Vector of doubles 
+#' 
 #' @export
 dbls_named <- function(x) .Call(wrap__dbls_named, x)
 
