+++ tests/extendrtests/src/rust/src/attributes.rs
@@ -1,6 +1,9 @@
 use extendr_api::prelude::*;
 
 /// Adds an attribute to a vector of doubles
+/// 
+/// @param x Vector of doubles 
+/// 
 /// @export
 #[extendr]
 fn dbls_named(x: Doubles) -> Doubles {
