+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -529,6 +529,9 @@
       mod attributes {
           use extendr_api::prelude::*;
           /// Adds an attribute to a vector of doubles
+          ///
+          /// @param x Vector of doubles
+          ///
           /// @export
           fn dbls_named(x: Doubles) -> Doubles {
               let mut x = x;
@@ -615,7 +618,7 @@
               );
               metadata
                   .push(extendr_api::metadata::Func {
-                      doc: " Adds an attribute to a vector of doubles\n @export",
+                      doc: " Adds an attribute to a vector of doubles\n \n @param x Vector of doubles \n \n @export",
                       rust_name: "dbls_named",
                       r_name: "dbls_named",
                       mod_name: "dbls_named",
