### 2024-05-01 46693c4
Add R tests
