+++ tests/extendrtests/R/extendr-wrappers.R
@@ -110,6 +110,8 @@ type_aware_sum <- function(input) .Call(wrap__type_aware_sum, input)
 #' @export
 euclidean_dist <- function(a) .Call(wrap__euclidean_dist, a)
 
+faer_mat <- function(x) .Call(wrap__faer_mat, x)
+
 #' Test raw identifiers (`r#`) in function arguments are parsed correctly.
 #' See [Issue #582](https://github.com/extendr/extendr/issues/528) for details.
 #' @param type : i32 or `NULL`
