### 2024-05-01 720de2d
Add faer to full-functionality and test lists
