### 2024-05-01 fd60bd5
faer: added the FromRobj conversions
