### 2024-05-01 b1e474d
Improve checks and add from RMatrix<i32>
