+++ tests/extendrtests/src/rust/src/attributes.rs
@@ -1,10 +1,5 @@
 use extendr_api::prelude::*;
 
-/// Adds an attribute to a vector of doubles
-/// 
-/// @param x Vector of doubles 
-/// 
-/// @export
 #[extendr]
 fn dbls_named(x: Doubles) -> Doubles {
     let mut x = x;
