### 2024-05-03 71be443
Introduce optional faer support  (#706)
* Introduce optional faer support

* Add tests

* Prepare change to mono crate faer

* Move faer to optional

* Implement traits for MatRef<f64>

* Add TryFrom

* Add better impl traits

* Bump faer to v0.18.0

* Fix lint

* Add R tests

* Add faer to full-functionality and test lists

* Remove unnecessary matrix check

* Refactor with is_matrix()

* Fix CI

* safer conversions

* WIP: Apply suggestions [skip ci]

* faer: added the FromRobj conversions

* fix rebase issue

* missing

* progress I guess?

* bug fix: optional_either was given twice

* `car
[…]
