+++ extendr-api/src/optional/faer.rs
@@ -1,0 +1,280 @@
+use faer::{mat, Mat, MatRef};
+
+use crate::scalar::Rfloat;
+use crate::scalar::Scalar;
+use crate::*;
+
+/// Convert a `faer::Mat<f64>` into an `RMatrix<f64>` which is not NA aware.
+impl From<Mat<f64>> for RMatrix<f64> {
+    fn from(value: Mat<f64>) -> Self {
+        RMatrix::new_matrix(value.nrows(), value.ncols(), |i, j| value.read(i, j))
+    }
+}
+
+impl From<Mat<f64>> for Robj {
+    fn from(value: Mat<f64>) -> Self {
+        RMatrix::<f64>::from(value).into()
+    }
+}
+
+/// Convert a `faer::Mat<f64>` into an `RMatrix<f64>` which is not NA aware.
+impl From<MatRef<'_, f64>> for RMatrix<f64> {
+    /// Convert a faer MatRef<f64> into Robj.
+    fn from(value: MatRef<'_, f64>) -> Self {
+        RMatrix::new_matrix(value.nrows(), value.ncols(), |i, j| value.read(i, j))
+    }
+}
+
+impl From<MatRef<'_, f64>> for Robj {
+    fn from(value: MatRef<'_, f64>) -> Self {
+        RMatrix::<f64>::from(value).into_robj()
+    }
+}
+
+impl From<Mat<f64>> for RMatrix<Rfloat> {
+    fn from(value: Mat<f64>) -> Self {
+        RMatrix::new_matrix(value.nrows(), value.ncols(), |i, j| value.read(i, j).into())
+    }
+}
+
+impl From<MatRef<'_, f64>> for RMatrix<Rfloat> {
+    fn from(value: MatRef<f64>) -> Self {
+        RMatrix::new_matrix(value.nrows(), value.ncols(), |i, j| {
+            Rfloat::from(value.read(i, j))
+        })
+    }
+}
+
+impl From<RMatrix<f64>> for Mat<f64> {
+    fn from(value: RMatrix<f64>) -> Self {
+        let nrow = value.nrows();
+        let ncol = value.ncols();
+        let slice = value.as_real_slice().expect("RMatrix should be doubles");
+        Mat::from_fn(nrow, ncol, |i, j| slice[i + j * nrow])
+    }
+}
+
+impl<'a> From<&'_ RMatrix<f64>> for MatRef<'a, f64> {
+    fn from(value: &RMatrix<f64>) -> Self {
+        let nrow = value.nrows();
+        let ncol = value.ncols();
+        let slice = value.as_typed_slice().expect("RMatrix should be doubles");
+        let mat_ref = faer::mat::from_column_major_slice::<f64>(slice, nrow, ncol);
+        mat_ref
+    }
+}
+
+impl TryFrom<&Robj> for Mat<f64> {
+    type Error = Error;
+
+    fn try_from(robj: &Robj) -> Result<Self> {
+        let rmat = &RMatrix::<f64>::from_robj(robj)?;
+        let nrow = rmat.nrows();
+        let ncol = rmat.ncols();
+
+        if let Some(slice) = robj.as_real_slice() {
+            let fmat = Mat::from_fn(nrow, ncol, |i, j| slice[i + j * nrow]);
+            Ok(fmat)
+        } else {
+            Err(Error::ExpectedReal(robj.clone()))
+        }
+    }
+}
+
+impl<'a> TryFrom<&'_ Robj> for MatRef<'a, f64> {
+    type Error = Error;
+
+    fn try_from(robj: &Robj) -> Result<Self> {
+        let rmat = &RMatrix::<f64>::from_robj(robj)?;
+        let nrows = rmat.nrows();
+        let ncols = rmat.ncols();
+
+        if let Some(slice) = robj.as_typed_slice() {
[…]
