### 2024-05-03 0485497
extendr-wrapper: `cargo extendr document`
