+++ tests/extendrtests/NAMESPACE
@@ -21,7 +21,6 @@ export(MyClass)
 export(MyClassTryFrom)
 export(MySubmoduleClass)
 export(MySubmoduleClassTryFrom)
-export(dbls_named)
 export(euclidean_dist)
 export(false)
 export(hello_submodule)
