### 2024-05-03 e4c961a
alternative: Take `Self` and return  `Self` from attributes instead
