+++ extendr-api/src/robj/into_robj.rs
@@ -83,12 +83,9 @@ where
         use crate as extendr_api;
         match res {
             Ok(x) => x.into(),
-            Err(x) => {
-                let mut err = list!(message = "extendr_err", value = x.into());
-                err.set_class(["extendr_error", "error", "condition"])
-                    .expect("internal error: failed to set class");
-                err.into()
-            }
+            Err(x) => { list!(message = "extendr_err", value = x.into()) }
+                .set_class(["extendr_error", "error", "condition"])
+                .expect("internal error: failed to set class"),
         }
     }
 }
@@ -107,7 +104,7 @@ where
 {
     fn from(res: std::result::Result<T, E>) -> Self {
         use crate as extendr_api;
-        let mut result = match res {
+        match res {
             Ok(x) => list!(ok = x.into(), err = NULL),
             Err(x) => {
                 let err_robj = x.into();
@@ -116,11 +113,10 @@ where
                 }
                 list!(ok = NULL, err = err_robj)
             }
-        };
-        result
-            .set_class(&["extendr_result"])
-            .expect("Internal error: failed to set class");
-        result.into()
+        }
+        .set_class(&["extendr_result"])
+        .expect("Internal error: failed to set class")
+        .into()
     }
 }
 
@@ -586,7 +582,7 @@ pub trait RobjItertools: Iterator {
         Robj: AsTypedSlice<'a, Self::Item>,
         Self::Item: 'a,
     {
-        let mut vector = self.collect_robj();
+        let vector = self.collect_robj();
         let prod = dims.iter().product::<usize>();
         if prod != vector.len() {
             return Err(Error::Other(format!(
@@ -595,11 +591,11 @@ pub trait RobjItertools: Iterator {
                 prod
             )));
         }
-        vector.set_attrib(wrapper::symbol::dim_symbol(), dims.iter().collect_robj())?;
-        let _data = vector.as_typed_slice().ok_or(Error::Other(
+        let robj = vector.set_attrib(wrapper::symbol::dim_symbol(), dims.iter().collect_robj())?;
+        let _data = robj.as_typed_slice().ok_or(Error::Other(
             "Unknown error in converting to slice".to_string(),
         ))?;
-        Ok(RArray::from_parts(vector, dims))
+        Ok(RArray::from_parts(robj, dims))
     }
 }
 
