### 2024-05-03 0cfed6a
Add `Attributes` to all vector types (#745)
* extendr-api: enable the ability to set attributes on vector-types

* extendr-api: Set attributes for Strings / character-vectors

* suggestion

* ci: fix doc test

* doc tests: fixed `cargo test --doc`

* add integration tests

* revert extendr-api path

* snapshot tests generated
various warnings fixed

* deleted snapshot temporary file

* fixed the `result_list` and `result_condition`

* accept snapshot

* r-cmd-check: finally fixed

* update changelog and attribute doc

* remove roxygen from integration test

* fix typo

* extendr-wrapper: `carg
[…]
