+++ extendr-api/src/metadata.rs
@@ -2,7 +2,6 @@
 //!
 //! This data is returned by get_module_metadata()
 //! which is generated by [extendr_module!].
-use crate::robj::IntoRobj;
 use crate::*;
 use std::io::Write;
 
@@ -76,16 +75,17 @@ impl From<&Arg> for RArg {
 
 impl From<Arg> for Robj {
     fn from(val: Arg) -> Self {
-        List::from_values(&[r!(val.name), r!(val.arg_type)])
-            .into_robj()
+        let mut result = List::from_values(&[r!(val.name), r!(val.arg_type)]);
+        result
             .set_names(&["name", "arg_type"])
-            .expect("From<Arg> failed")
+            .expect("From<Arg> failed");
+        result.into()
     }
 }
 
 impl From<Func> for Robj {
     fn from(val: Func) -> Self {
-        List::from_values(&[
+        let mut result = List::from_values(&[
             r!(val.doc),
             r!(val.rust_name),
             r!(val.mod_name),
@@ -93,44 +93,47 @@ impl From<Func> for Robj {
             r!(List::from_values(val.args)),
             r!(val.return_type),
             r!(val.hidden),
-        ])
-        .into_robj()
-        .set_names(&[
-            "doc",
-            "rust_name",
-            "mod_name",
-            "r_name",
-            "args",
-            "return.type",
-            "hidden",
-        ])
-        .expect("From<Func> failed")
+        ]);
+        result
+            .set_names(&[
+                "doc",
+                "rust_name",
+                "mod_name",
+                "r_name",
+                "args",
+                "return.type",
+                "hidden",
+            ])
+            .expect("From<Func> failed");
+        result.into()
     }
 }
 
 impl From<Impl> for Robj {
     fn from(val: Impl) -> Self {
-        List::from_values(&[
+        let mut result = List::from_values(&[
             r!(val.doc),
             r!(val.name),
             r!(List::from_values(val.methods)),
-        ])
-        .into_robj()
-        .set_names(&["doc", "name", "methods"])
-        .expect("From<Impl> failed")
+        ]);
+        result
+            .set_names(&["doc", "name", "methods"])
+            .expect("From<Impl> failed");
+        result.into()
     }
 }
 
 impl From<Metadata> for Robj {
     fn from(val: Metadata) -> Self {
-        List::from_values(&[
+        let mut result = List::from_values(&[
             r!(val.name),
             r!(List::from_values(val.functions)),
             r!(List::from_values(val.impls)),
-        ])
-        .into_robj()
-        .set_names(&["name", "functions", "impls"])
-        .expect("From<Metadata> failed")
+        ]);
+        result
+            .set_names(&["name", "functions", "impls"])
+            .expect("From<Metadata> failed");
+        result.into()
     }
 }
 
