+++ extendr-api/src/wrapper/macros.rs
@@ -9,6 +9,9 @@ macro_rules! gen_vector_wrapper_impl {
         doc_name: $doc_name : ident,
         altrep_constructor: $altrep_constructor : ident,
     ) => {
+
+        impl Attributes for $type {}
+
         impl Default for $type {
             fn default() -> Self {
                 $type::new(0)
