### 2024-05-04 565756a
xtask: added a line about `cargo extendr devtools-test` generating snapshots, which then can be accepted be a subsequent `cargo extendr devtools-test -a`.
