### 2024-05-04 7cd8931
extendr-macros: fix the lack of `mut`-alias in `#[extendr]` macro
