+++ extendr-macros/src/wrappers.rs
@@ -354,7 +354,16 @@ pub fn translate_formal(input: &FnArg, self_ty: Option<&syn::Type>) -> syn::Resu
         // function argument.
         FnArg::Typed(ref pattype) => {
             let pat = &pattype.pat.as_ref();
-            Ok(parse_quote! { #pat : extendr_api::SEXP })
+            let pat_ident = match pat {
+                syn::Pat::Ident(ref pat_ident) => &pat_ident.ident,
+                _ => {
+                    return Err(syn::Error::new_spanned(
+                        input,
+                        "failed to translate name of argument",
+                    ))
+                }
+            };
+            Ok(parse_quote! { #pat_ident : extendr_api::SEXP })
         }
         // &self / &mut self
         FnArg::Receiver(ref receiver) => {
@@ -381,7 +390,15 @@ fn translate_meta_arg(input: &mut FnArg, self_ty: Option<&syn::Type>) -> syn::Re
         FnArg::Typed(ref mut pattype) => {
             let pat = pattype.pat.as_ref();
             let ty = pattype.ty.as_ref();
-            let name_string = quote! { #pat }.to_string();
+            let pat_ident = if let syn::Pat::Ident(ref pat_ident) = pat {
+                &pat_ident.ident
+            } else {
+                return Err(syn::Error::new_spanned(
+                    input,
+                    "failed to translate name of argument",
+                ));
+            };
+            let name_string = quote! { #pat_ident }.to_string();
             let type_string = type_name(ty);
             let default = if let Some(default) = get_named_lit(&mut pattype.attrs, "default") {
                 quote!(Some(#default))
@@ -433,7 +450,9 @@ fn translate_to_robj(input: &FnArg) -> syn::Result<syn::Stmt> {
             let pat = &pattype.pat.as_ref();
             if let syn::Pat::Ident(ref ident) = pat {
                 let varname = format_ident!("_{}_robj", ident.ident);
-                Ok(parse_quote! { let #varname = extendr_api::robj::Robj::from_sexp(#pat); })
+                let ident = &ident.ident;
+                // TODO: these do not need protection, as they come from R
+                Ok(parse_quote! { let #varname = extendr_api::robj::Robj::from_sexp(#ident); })
             } else {
                 Err(syn::Error::new_spanned(
                     input,
