+++ type.impl/std/primitive.isize.js
@@ -1,3 +1,3 @@
 (function() {var type_impls = {
-"extendr_api":[["<details class=\"toggle implementors-toggle\" open><summary><section id=\"impl-TryFrom%3C%26Robj%3E-for-isize\" class=\"impl\"><a class=\"src rightside\" href=\"src/extendr_api/robj/try_from_robj.rs.html#100\">source</a><a href=\"#impl-TryFrom%3C%26Robj%3E-for-isize\" class=\"anchor\">§</a><h3 class=\"code-header\">impl <a class=\"trait\" href=\"extendr_api/trait.TryFrom.html\" title=\"trait extendr_api::TryFrom\">TryFrom</a>&lt;&amp;<a class=\"struct\" href=\"extendr_api/robj/struct.Robj.html\" title=\"struct extendr_api::robj::Robj\">Robj</a>&gt; for <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.isize.html\">isize</a></h3></section></summary><div class=\"impl-items\"><details class=\"toggle method-toggle\" open><summary><section id=\"method.try_from\" class=\"method trait-impl\"><a class=\"src rightside\" href=\"src/extendr_api/robj/try_from_robj.rs.html#100\">source</a><a href=\"#method.try_from\" class=\"anchor\">§</a><h4 class=\"code-header\">fn <a href=\"extendr_api/trait.TryFrom.html#tymethod.try_from\" class=\"fn\">try_from</a>(robj: &amp;<a class=\"struct\" href=\"extendr_api/robj/struct.Robj.html\" title=\"struct extendr_api::robj::Robj\">Robj</a>) -&gt; <a class=\"type\" href=\"extendr_api/error/type.Result.html\" title=\"type extendr_api::error::Result\">Result</a>&lt;Self&gt;</h4></section></summary><div class=\"docblock\"><p>Convert a numeric object to an integer value.</p>\n</div></details><details class=\"toggle\" open><summary><section id=\"associatedtype.Error\" class=\"associatedtype trait-impl\"><a href=\"#associatedtype.Error\" class=\"anchor\">§</a><h4 class=\"code-header\">type <a href=\"extendr_api/trait.TryFrom.html#associatedtype.Error\" class=\"associatedtype\">Error</a> = <a class=\"enum\" href=\"extendr_api/error/enum.Error.html\" title=\"enum extendr_api::error::Error\">Error</a></h4></section></summary><div class='docblock'>The type returned in the event of a conversion error.</div></details></div></details>","TryFrom<&Robj>","extendr_api::prelude::Ixs"],["<details class=\"toggle implementors-toggle\" open><summary><section id=\"impl-TryFrom%3CRobj%3E-for-isize\" class=\"impl\"><a class=\"src rightside\" href=\"src/extendr_api/robj/try_from_robj.rs.html#423-432\">source</a><a href=\"#impl-TryFrom%3CRobj%3E-for-isize\" class=\"anchor\">§</a><h3 class=\"code-header\">impl <a class=\"trait\" href=\"extendr_api/trait.TryFrom.html\" title=\"trait extendr_api::TryFrom\">TryFrom</a>&lt;<a class=\"struct\" href=\"extendr_api/robj/struct.Robj.html\" title=\"struct extendr_api::robj::Robj\">Robj</a>&gt; for <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.isize.html\">isize</a></h3></section></summary><div class=\"impl-items\"><details class=\"toggle\" open><summary><section id=\"associatedtype.Error\" class=\"associatedtype trait-impl\"><a href=\"#associatedtype.Error\" class=\"anchor\">§</a><h4 class=\"code-header\">type <a href=\"extendr_api/trait.TryFr
[…]
