+++ src-files.js
@@ -1,5 +1,5 @@
 var srcIndex = new Map(JSON.parse('[\
-["extendr_api",["",[["graphics",[],["color.rs","device_descriptor.rs","device_driver.rs","mod.rs"]],["io",[],["load.rs","mod.rs","save.rs"]],["optional",[],["either.rs","mod.rs","ndarray.rs"]],["robj",[["into_robj",[],["repeat_into_robj.rs"]]],["debug.rs","from_robj.rs","into_robj.rs","mod.rs","operators.rs","rinternals.rs","try_from_robj.rs"]],["scalar",[],["macros.rs","mod.rs","rbool.rs","rcplx_full.rs","rfloat.rs","rint.rs"]],["wrapper",[],["altrep.rs","complexes.rs","dataframe.rs","doubles.rs","environment.rs","expr.rs","externalptr.rs","function.rs","integers.rs","lang.rs","list.rs","logicals.rs","macros.rs","matrix.rs","mod.rs","nullable.rs","pairlist.rs","primitive.rs","promise.rs","raw.rs","rstr.rs","s4.rs","strings.rs","symbol.rs"]]],["deserializer.rs","error.rs","functions.rs","iter.rs","lang_macros.rs","lib.rs","metadata.rs","na.rs","ownership.rs","prelude.rs","rmacros.rs","serializer.rs","thread_safety.rs"]]],\
+["extendr_api",["",[["graphics",[],["color.rs","device_descriptor.rs","device_driver.rs","mod.rs"]],["io",[],["load.rs","mod.rs","save.rs"]],["optional",[],["either.rs","faer.rs","mod.rs","ndarray.rs"]],["robj",[["into_robj",[],["repeat_into_robj.rs"]]],["debug.rs","from_robj.rs","into_robj.rs","mod.rs","operators.rs","rinternals.rs","try_from_robj.rs"]],["scalar",[],["macros.rs","mod.rs","rbool.rs","rcplx_full.rs","rfloat.rs","rint.rs"]],["wrapper",[],["altrep.rs","complexes.rs","dataframe.rs","doubles.rs","environment.rs","expr.rs","externalptr.rs","function.rs","integers.rs","lang.rs","list.rs","logicals.rs","macros.rs","matrix.rs","mod.rs","nullable.rs","pairlist.rs","primitive.rs","promise.rs","raw.rs","rstr.rs","s4.rs","strings.rs","symbol.rs"]]],["deserializer.rs","error.rs","functions.rs","iter.rs","lang_macros.rs","lib.rs","metadata.rs","na.rs","ownership.rs","prelude.rs","rmacros.rs","serializer.rs","thread_safety.rs"]]],\
 ["extendr_engine",["",[],["lib.rs"]]],\
 ["extendr_macros",["",[],["R.rs","call.rs","dataframe.rs","extendr_function.rs","extendr_impl.rs","extendr_module.rs","lib.rs","list.rs","list_struct.rs","pairlist.rs","pairs.rs","wrappers.rs"]]],\
 ["xtask",["",[["cli",[],["devtools_test.rs","mod.rs","r_cmd_check.rs"]],["commands",[],["cargo_fmt.rs","cargo_fmt_check.rs","cargo_msrv.rs","devtools_test.rs","generate_docs.rs","mod.rs","r_cmd_check.rs","rextendr_document.rs"]],["extendrtests",[],["mod.rs","path_helper.rs","with_absolute_path.rs"]]],["main.rs"]]]\
