### 2024-05-07 9f783cd
use `if let Some` instead of `match
