+++ extendr-macros/src/wrappers.rs
@@ -354,14 +354,13 @@ pub fn translate_formal(input: &FnArg, self_ty: Option<&syn::Type>) -> syn::Resu
         // function argument.
         FnArg::Typed(ref pattype) => {
             let pat = &pattype.pat.as_ref();
-            let pat_ident = match pat {
-                syn::Pat::Ident(ref pat_ident) => &pat_ident.ident,
-                _ => {
-                    return Err(syn::Error::new_spanned(
-                        input,
-                        "failed to translate name of argument",
-                    ))
-                }
+            let pat_ident = if let syn::Pat::Ident(ref pat_ident) = pat {
+                &pat_ident.ident
+            } else {
+                return Err(syn::Error::new_spanned(
+                    input,
+                    "failed to translate name of argument",
+                ));
             };
             Ok(parse_quote! { #pat_ident : extendr_api::SEXP })
         }
