### 2024-05-09 431b8f0
refactor: moved clean-up to function
