+++ extendr-api/Cargo.toml
@@ -8,7 +8,7 @@ license.workspace = true
 repository.workspace = true
 
 # Note: it seems cargo-msrv doesn't support rust-version.workspace = true.
-rust-version = "1.64"
+rust-version = "1.70"
 
 [dependencies]
 libR-sys = { workspace = true }
