### 2024-05-09 6923b8e
`extendr-macros`: Accept argument alias with `mut` in front  (#752)
* extendr-macros: Use `try_from` instead of `FromRobj`.

* extendr-macros: fix the lack of `mut`-alias in `#[extendr]` macro

* use `if let Some` instead of `match

* refactor: moved the code to clear `mut` from `Ident` to its own function
