+++ extendr-macros/src/extendr_module.rs
@@ -11,7 +11,7 @@ pub fn extendr_module(item: TokenStream) -> TokenStream {
         implnames,
         usenames,
     } = module;
-    let modname = modname.unwrap();
+    let modname = modname.expect("cannot include unnamed modules");
     let modname_string = modname.to_string();
     let module_init_name = format_ident!("R_init_{}_extendr", modname);
 
@@ -102,10 +102,10 @@ pub fn extendr_module(item: TokenStream) -> TokenStream {
                 use extendr_api::robj::*;
                 use extendr_api::GetSexp;
                 let robj = Robj::from_sexp(use_symbols_sexp);
-                let use_symbols: bool = <bool>::from_robj(&robj).unwrap();
+                let use_symbols: bool = <bool>::try_from(&robj).unwrap();
 
                 let robj = Robj::from_sexp(package_name_sexp);
-                let package_name: &str = <&str>::from_robj(&robj).unwrap();
+                let package_name: &str = <&str>::try_from(&robj).unwrap();
 
                 extendr_api::Robj::from(
                     #module_metadata_name()
