+++ trait.impl/serde/de/trait.SeqAccess.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl&lt;'de&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.200/serde/de/trait.SeqAccess.html\" title=\"trait serde::de::SeqAccess\">SeqAccess</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.ListGetter.html\" title=\"struct extendr_api::deserializer::ListGetter\">ListGetter</a>&lt;'de&gt;"],["impl&lt;'de, E&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.200/serde/de/trait.SeqAccess.html\" title=\"trait serde::de::SeqAccess\">SeqAccess</a>&lt;'de&gt; for <a class=\"struct\" href=\"extendr_api/deserializer/struct.SliceGetter.html\" title=\"struct extendr_api::deserializer::SliceGetter\">SliceGetter</a>&lt;'de, E&gt;<div class=\"where\">where\n    <a class=\"enum\" href=\"extendr_api/error/enum.Error.html\" title=\"enum extendr_api::error::Error\">Error</a>: <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&lt;E as <a class=\"trait\" href=\"https://docs.rs/serde/1.0.200/serde/de/trait.Deserializer.html\" title=\"trait serde::de::Deserializer\">Deserializer</a>&lt;'de&gt;&gt;::<a class=\"associatedtype\" href=\"https://docs.rs/serde/1.0.200/serde/de/trait.Deserializer.html#associatedtype.Error\" title=\"type serde::de::Deserializer::Error\">Error</a>&gt;,\n    E: <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/marker/trait.Copy.html\" title=\"trait core::marker::Copy\">Copy</a> + <a class=\"trait\" href=\"https://docs.rs/serde/1.0.200/serde/de/trait.Deserializer.html\" title=\"trait serde::de::Deserializer\">Deserializer</a>&lt;'de&gt;,</div>"]]
[…]
