+++ trait.impl/serde/ser/trait.Serializer.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"extendr_api":[["impl&lt;'a&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.200/serde/ser/trait.Serializer.html\" title=\"trait serde::ser::Serializer\">Serializer</a> for &amp;'a mut <a class=\"struct\" href=\"extendr_api/serializer/struct.RobjSerializer.html\" title=\"struct extendr_api::serializer::RobjSerializer\">RobjSerializer</a>"]]
+"extendr_api":[["impl&lt;'a&gt; <a class=\"trait\" href=\"https://docs.rs/serde/1.0.201/serde/ser/trait.Serializer.html\" title=\"trait serde::ser::Serializer\">Serializer</a> for &amp;'a mut <a class=\"struct\" href=\"extendr_api/serializer/struct.RobjSerializer.html\" title=\"struct extendr_api::serializer::RobjSerializer\">RobjSerializer</a>"]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()

[…]
