+++ extendr-api/src/wrapper/function.rs
@@ -42,7 +42,7 @@ impl Function {
     /// ```
     pub fn from_parts(formals: Pairlist, body: Language, env: Environment) -> Result<Self> {
         single_threaded(|| unsafe {
-            let sexp = Rf_allocSExp(CLOSXP);
+            let sexp = Rf_allocSExp(SEXPTYPE::CLOSXP);
             let robj = Robj::from_sexp(sexp);
             SET_FORMALS(sexp, formals.get());
             SET_BODY(sexp, body.get());
