+++ extendr-api/src/wrapper/doubles.rs
@@ -20,6 +20,7 @@ pub struct Doubles {
     pub(crate) robj: Robj,
 }
 
+use libR_sys::SEXPTYPE::REALSXP;
 crate::wrapper::macros::gen_vector_wrapper_impl!(
     vector_type: Doubles,
     scalar_type: Rfloat,

[…]
