### 2024-05-09 2a8fb6c
Replaces `SEXPTYPE` and `Rboolean` to the new `enum`-versions (#742)
* replaces `SEXPTYPE` and `Rboolean` to the
new enum-versions from latest `libR-sys`.

* Cargo.toml: Update version of `libR-sys`, otherwise it doesn't work

* build.rs: Added a feature for using `OBJSXP` vs `S4SXP`
The cfg-feature might be excessive, but handling S4 object will
have to depend on R versions eventually, so we are starting
this branching now...

* MSRV: Updated to 1.70 due to a change
in bindgen, which uses `which`, and which
uses `home` which silently updated
their msrv to 1.70.
