+++ extendr-api/src/robj/into_robj.rs
@@ -13,7 +13,7 @@ pub(crate) fn str_to_character(s: &str) -> SEXP {
                 Rf_mkCharLenCE(
                     s.as_ptr() as *const raw::c_char,
                     s.len() as i32,
-                    cetype_t_CE_UTF8,
+                    cetype_t::CE_UTF8,
                 )
             })
         }
@@ -167,7 +167,7 @@ where
 /// to `collect_robj()`.
 pub trait ToVectorValue {
     fn sexptype() -> SEXPTYPE {
-        0
+        SEXPTYPE::NILSXP
     }
 
     fn to_real(&self) -> f64
@@ -217,7 +217,7 @@ macro_rules! impl_real_tvv {
     ($t: ty) => {
         impl ToVectorValue for $t {
             fn sexptype() -> SEXPTYPE {
-                REALSXP
+                SEXPTYPE::REALSXP
             }
 
             fn to_real(&self) -> f64 {
@@ -227,7 +227,7 @@ macro_rules! impl_real_tvv {
 
         impl ToVectorValue for &$t {
             fn sexptype() -> SEXPTYPE {
-                REALSXP
+                SEXPTYPE::REALSXP
             }
 
             fn to_real(&self) -> f64 {
@@ -237,7 +237,7 @@ macro_rules! impl_real_tvv {
 
         impl ToVectorValue for Option<$t> {
             fn sexptype() -> SEXPTYPE {
-                REALSXP
+                SEXPTYPE::REALSXP
             }
 
             fn to_real(&self) -> f64 {
@@ -265,7 +265,7 @@ macro_rules! impl_complex_tvv {
     ($t: ty) => {
         impl ToVectorValue for $t {
             fn sexptype() -> SEXPTYPE {
-                CPLXSXP
+                SEXPTYPE::CPLXSXP
             }
 
             fn to_complex(&self) -> Rcomplex {
@@ -275,7 +275,7 @@ macro_rules! impl_complex_tvv {
 
         impl ToVectorValue for &$t {
             fn sexptype() -> SEXPTYPE {
-                CPLXSXP
+                SEXPTYPE::CPLXSXP
             }
 
             fn to_complex(&self) -> Rcomplex {
@@ -293,7 +293,7 @@ macro_rules! impl_integer_tvv {
     ($t: ty) => {
         impl ToVectorValue for $t {
             fn sexptype() -> SEXPTYPE {
-                INTSXP
+                SEXPTYPE::INTSXP
             }
 
             fn to_integer(&self) -> i32 {
@@ -303,7 +303,7 @@ macro_rules! impl_integer_tvv {
 
         impl ToVectorValue for &$t {
             fn sexptype() -> SEXPTYPE {
-                INTSXP
+                SEXPTYPE::INTSXP
             }
 
             fn to_integer(&self) -> i32 {
@@ -313,7 +313,7 @@ macro_rules! impl_integer_tvv {
 
         impl ToVectorValue for Option<$t> {
             fn sexptype() -> SEXPTYPE {
-                INTSXP
+                SEXPTYPE::INTSXP
             }
 
             fn to_integer(&self) -> i32 {
@@ -334,7 +334,7 @@ impl_integer_tvv!(u16);
 
 impl ToVectorValue for u8 {
     fn sexptype() -> SEXPTYPE {
-        RAWSXP
+        SEXPTYPE::RAWSXP
     }
 
     fn to_raw(&self) -> u8 {
@@ -344,7 +344,7 @@ impl ToVectorValue for u8 {
 
 impl ToVectorValue for &u8 {
     fn sexptype() -> SEXPTYPE {
-        RAWSXP
[…]
