+++ extendr-api/build.rs
@@ -30,4 +30,8 @@ fn main() {
     if &*major >= "4" && &*minor >= "3" {
         println!("cargo:rustc-cfg=use_r_altlist");
     }
+
+    if &*major >= "4" && &*minor >= "4" {
+        println!("cargo:rustc-cfg=use_objsxp");
+    }
 }
