### 2024-05-11 17b33a3
xtask: add helpful message when missing dependencies for `r-cmd-check`
