### 2024-05-11 5dc6c11
README: add instruction to use developer extendr in rextendr.
