### 2024-05-11 8479e40
added the `StrIter` representation
