### 2024-05-11 99462ff
extendr-api: added `non-api` flag
