+++ extendr-api/src/wrapper/environment.rs
@@ -80,6 +80,7 @@ impl Environment {
         }
     }
 
+    #[cfg(feature = "non-api")]
     /// Set the enclosing (parent) environment.
     pub fn set_parent(&mut self, parent: Environment) -> &mut Self {
         single_threaded(|| unsafe {
@@ -97,6 +98,7 @@ impl Environment {
         }
     }
 
+    #[cfg(feature = "non-api")]
     /// Set the environment flags.
     pub fn set_envflags(&mut self, flags: i32) -> &mut Self {
         single_threaded(|| unsafe {
