+++ Cargo.toml
@@ -23,7 +23,7 @@ rust-version = "1.70"
 # When updating extendr's version, this version also needs to be updated
 extendr-macros = { path = "./extendr-macros", version = "0.6.0" }
 
-libR-sys = { version = "0.7.0" }
+libR-sys = "0.7.0"
 
 [patch.crates-io]
 # When uncommenting this, do not forget to uncomment the same line in
