+++ .github/workflows/test.yml
@@ -238,7 +238,7 @@ jobs:
                 "\" }"),
 
               # uncomment this line when we need to depend on the dev version of libR-sys
-              'libR-sys = { git = "https://github.com/extendr/libR-sys", features = list("non-api") }',
+              'libR-sys = { git = "https://github.com/extendr/libR-sys" }',
 
               sep = ";")
           
