### 2024-05-11 363597e
ci: temporary hack to get past this circular testing BS
