### 2024-05-11 1511ed9
ci: instruction by @Ilia-Kosenkov
