+++ .github/workflows/test.yml
@@ -246,6 +246,7 @@ jobs:
 
           github_env_file <- Sys.getenv("GITHUB_ENV")
           cat(patch.crates_io, file = github_env_file, append = TRUE)
+          cat("REXTENDR_SKIP_DEV_TESTS=TRUE", file = github_env_file, append = TRUE)
 
           cat("::endgroup::\n")
         shell: Rscript {0}
