### 2024-05-11 1eb7011
Skip rextendr dev tests
