### 2024-05-11 f2934cf
Update .github/workflows/test.yml
Co-authored-by: CGMossa <cgmossa@gmail.com>
