### 2024-05-11 32ca239
ci: added ` REXTENDR_SKIP_DEV_TESTS: TRUE`
