+++ .github/workflows/test.yml
@@ -239,7 +239,7 @@ jobs:
                 "\" }"),
 
               # uncomment this line when we need to depend on the dev version of libR-sys
-              'libR-sys = { git = "https://github.com/extendr/libR-sys", features = array("non-api", 1) }',
+              # 'libR-sys = { git = "https://github.com/extendr/libR-sys", features = array("non-api", 1) }',
 
               sep = ";")
           
