### 2024-05-11 4798a02
Various maintenance (#754)
* xtask: add helpful message when missing dependencies for `r-cmd-check`

* Makevars: color is auto, not always

* README: add instruction to use developer extendr in rextendr.

* extendr-api: added `non-api` flag

* ci: temporary hack to get past this circular testing BS

* Cargo.toml: we will no longer rely on `patch.crates-io`.

* ci: added ` REXTENDR_SKIP_DEV_TESTS: TRUE`
