+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 
 - Breaking: the `Attributes` trait now returns a mutable reference to `Self`. [[#745]](https://github.com/extendr/extendr/pull/745). Previously `.set_attrib()` would modify an object in place, and then return an untyped owned pointer (Robj). Instead, now we return `&mut Self`.
 - Potentially breaking: `RArray::from_parts` no longer requires a pointer to the underlying data vector [[#657]](https://github.com/extendr/extendr/pull/657)
-- Breaking change: In `AltRep` the `unserialize_ex` is now hidden behind the feature flag `non-api`
+- Breaking change: In `AltRep` the `unserialize_ex`, `set_parent`, `set_envflags` are
+now hidden behind the feature flag `non-api`. Also, `Promise::from_parts` is marked as non-API.
 
 ### Fixed
 
