### 2024-05-12 8776fcd
Update CHANGELOG (#756)
* CHANGELOG: added entries

* freeze dev `libR-sys` to avoid CI chaos.

* suggestions

* It is necessary to freeze the `extendrtests`'s libR-sys aswell
