+++ CHANGELOG.md
@@ -10,12 +10,15 @@
 - The new `ExpectedExternalNonNullPtr` error variant provides a more informative error when a null pointer is accessed
 - `RArray::data_mut` provides a mutable slice to the underlying array data [[#657]](https://github.com/extendr/extendr/pull/657)
 - Implements the `Attributes` trait for all R vector wrapper structs (e.g. `Integers` , `Doubles`, `Strings`, etc.), allowing for easy access and setting of the attributes of an R object [[#745]](https://github.com/extendr/extendr/pull/745). This comes with breaking changes. See below.
+- feature `non-api` that gives access to non-API items; Requires compile-time generation of bindings
+[[#754]](https://github.com/extendr/extendr/pull/754)
 
 ### Changed
 
 - Breaking: the `Attributes` trait now returns a mutable reference to `Self`. [[#745]](https://github.com/extendr/extendr/pull/745). Previously `.set_attrib()` would modify an object in place, and then return an untyped owned pointer (Robj). Instead, now we return `&mut Self`.
 - Potentially breaking: `RArray::from_parts` no longer requires a pointer to the underlying data vector [[#657]](https://github.com/extendr/extendr/pull/657)
-
+- Breaking change: In `AltRep` the `unserialize_ex`, `set_parent`, `set_envflags` are
+now hidden behind the feature flag `non-api`. Also, `Promise::from_parts` is marked as non-API.
 
 ### Fixed
 
