+++ tests/extendrtests/src/rust/Cargo.toml
@@ -37,4 +37,4 @@ extendr-api = { path = "../../../../../../../extendr/extendr-api" }
 #extendr-api = { git = "https://github.com/extendr/extendr"}
 
 # Build against current libR-sys version on github
-libR-sys = { git = "https://github.com/extendr/libR-sys" }
+libR-sys = { git = "https://github.com/extendr/libR-sys", rev = "09d76ada0cd54aa4481d9f06bbdfa50bcca2814a" }
