+++ search.desc/extendr_api/extendr_api-desc-0-.js
@@ -1,1 +1,1 @@
-searchState.loadedDescShard("extendr_api", 0, "A safe and user friendly R extension interface.\nUsed for immutable dereferencing operations, like <code>*v</code>.\nUsed for mutable dereferencing operations, like in <code>*v = 1;</code>.\nThe type returned in the event of a conversion error.\nThe type returned in the event of a conversion error.\nFALSE value eg. <code>r!(FALSE)</code>\nEnable the construction of dataframes from arrays of …\nDerives an implementation of <code>From&lt;Struct&gt; for Robj</code> and …\nNA value for integers eg. <code>r!(NA_INTEGER)</code>\nNA value for logical. <code>r!(NA_LOGICAL)</code>\nNA value for real values eg. <code>r!(NA_REAL)</code>\nNA value for strings. <code>r!(NA_STRING)</code>\nNULL value eg. <code>r!(NULL)</code>\nExecute R code by parsing and evaluating tokens.\nEnum use to unpack R objects into their specialist …\nExecute R code by parsing and evaluating tokens but …\nType of R objects used by Robj::rtype.\nTRUE value eg. <code>r!(TRUE)</code>\nThe resulting type after dereferencing.\nSimple and safe type conversions that may fail in a …\nDerives an implementation of <code>TryFrom&lt;Robj&gt; for Struct</code> and …\nAn attempted conversion that consumes <code>self</code>, which may or …\nCall a function or primitive defined by a text expression …\nCreate a dataframe.\nDereferences the value.\nMutably dereferences the value.\nConvert R objects to a wide variety of types.\nError handling in Rust called from R.\nThe <code>#[extendr]</code>-macro may be placed on three items\nDefine a module and export symbols to R Example:\nCreate a factor.\nReturns the argument unchanged.\nReturns the argument unchanged.\nGet a global variable.\nGraphic Device Operations\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nA macro for constructing R langage objects.\nArgument parsing and checking.\nCreate a List R object from a list of name-value pairs.\nModule metadata\nA set of optional features and third-party crate …\nMaintain ownership of R objects.\nCreate a Pairlist R object from a list of name-value pairs.\nCommon exports for extendr-api.\nConvert a rust expression to an R object.\nPrint via the R error stream.\nPrint with a newline via the R output stream.\nrmacros - a set of macros to call actual R functions in a …\nR object handling.\nPrint via the R output stream.\nPrint with a newline via the R output stream.\nConvert extendr’s Rtype to R’s SEXPTYPE. Panics if the …\nSee https://serde.rs/impl-serializer.html\nConvert R’s SEXPTYPE to extendr’s Rtype.\nThe sym! macro install symbols. You should cache your …\nMacro for running tests.\nProvide limited protection for multithreaded access to the …\nPerforms the conversion.\nPerforms the conversion.\nGet a local variable from the calling function or a global …\nWrappers are lightweight proxies for references to R …\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nReturns the argument unchanged.\nConvert any R object to a Deserialize object.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)
[…]
