+++ xtask/src/cli/devtools_test.rs
@@ -10,10 +10,6 @@ pub(crate) struct DevtoolsTestArg {
     )]
     pub(crate) accept_snapshot: bool,
 
-    #[arg(
-        long,
-        short,
-        help = "Filter tests by name",
-    )]
+    #[arg(long, short, help = "Filter tests by name")]
     pub(crate) filter: Option<String>,
 }
