+++ xtask/src/commands/devtools_test.rs
@@ -4,7 +4,7 @@ use xshell::{cmd, Shell};
 
 use crate::{
     cli::devtools_test::DevtoolsTestArg,
-    extendrtests::with_absolute_path::{R_FOLDER_PATH, swap_extendr_api_path},
+    extendrtests::with_absolute_path::{swap_extendr_api_path, R_FOLDER_PATH},
 };
 
 pub(crate) fn run(shell: &Shell, args: DevtoolsTestArg) -> Result<(), Box<dyn Error>> {
@@ -25,7 +25,8 @@ fn run_tests(shell: &Shell, args: DevtoolsTestArg) -> Result<(), Box<dyn Error>>
         .run()?;
     }
     if let Some(filter) = args.filter {
-        shell.cmd("Rscript")
+        shell
+            .cmd("Rscript")
             .arg("-e")
             .arg(format!("devtools::test(filter = \"{filter}\")"))
             .run()?;
