+++ extendr-api/src/lib.rs
@@ -388,7 +388,6 @@ pub use error::*;
 pub use functions::*;
 pub use lang_macros::*;
 pub use na::*;
-pub use rmacros::*;
 pub use robj::*;
 pub use thread_safety::{catch_r_error, handle_panic, single_threaded, throw_r_error};
 pub use wrapper::*;
