+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - Potentially breaking: `RArray::from_parts` no longer requires a pointer to the underlying data vector [[#657]](https://github.com/extendr/extendr/pull/657)
 - Breaking change: In `AltRep` the `unserialize_ex`, `set_parent`, `set_envflags` are
 now hidden behind the feature flag `non-api`. Also, `Promise::from_parts` is marked as non-API.
+- Floating point numbers with decimal part can no longer be converted to integer types via
+  rounding [[#757]](https://github.com/extendr/extendr/pull/757)
 
 ### Fixed
 
