### 2024-05-12 24c25ad
Fix integerish conversions (#757)
* Prototype IntoIntegerish

* More tests

* Refactor

* Refactor from_robj conversions

* Fix logic

* Update bool converter as well

* Update TryFrom

* Cleanup

* Delegate implementation

* Update error

* cargo fmt

* Fix error display

* Filter extendr tests

* Fix 0.0 parsing error

* Cleanup

* Rename function

* Update changelog

* Fix integration tests

* Update CHANGELOG

* Annotate breaking changes
