+++ extendr-api/src/scalar/rint.rs
@@ -156,45 +156,13 @@ impl TryFrom<&Robj> for Rint {
     type Error = Error;
 
     fn try_from(robj: &Robj) -> Result<Self> {
-        // Check if the value is a scalar
-        match robj.len() {
-            0 => return Err(Error::ExpectedNonZeroLength(robj.clone())),
-            1 => {}
-            _ => return Err(Error::ExpectedScalar(robj.clone())),
-        };
-
-        // Check if the value is not a missing value
-        if robj.is_na() {
-            return Ok(Rint::na());
+        let i32_val: Result<i32> = robj.try_into();
+        match i32_val {
+            Ok(v) => Ok(Rint::from(v)),
+            // TODO: Currently this results in an extra protection of robj
+            Err(Error::MustNotBeNA(_)) => Ok(Rint::na()),
+            Err(e) => Err(e),
         }
-
-        // If the conversion is int-to-int, check the limits. This
-        // needs to be done by `TryFrom` because the conversion by `as`
-        // is problematic when converting a negative value to unsigned
-        // integer types (e.g. `-1i32 as u8` becomes 255).
-        if let Some(v) = robj.as_integer() {
-            if let Ok(v) = Self::try_from(v) {
-                return Ok(v);
-            } else {
-                return Err(Error::OutOfLimits(robj.clone()));
-            }
-        }
-
-        // If the conversion is float-to-int, check if the value is
-        // integer-like (i.e., an integer, or a float representing a
-        // whole number). This needs to be down with `as`, as no
-        // `TryFrom` is implemented for float types. `FloatToInt` trait
-        // might eventually become available in future, though.
-        if let Some(v) = robj.as_real() {
-            let result = v as i32;
-            if (result as f64 - v).abs() < f64::EPSILON {
-                return Ok(Rint::from(result));
-            } else {
-                return Err(Error::ExpectedWholeNumber(robj.clone()));
-            }
-        }
-
-        Err(Error::ExpectedNumeric(robj.clone()))
     }
 }
 
