+++ extendr_api/robj/from_robj/sidebar-items.js
@@ -1,1 +1,1 @@
-window.SIDEBAR_ITEMS = {"macro":["impl_iter_from_robj","impl_prim_from_robj"],"trait":["FromRobj"]};
+window.SIDEBAR_ITEMS = {"fn":["convert_scalar"],"macro":["impl_float_prim_from_robj","impl_int_prim_from_robj","impl_iter_from_robj"],"trait":["FromRobj"]};
