+++ extendr_api/conversions/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"mod":["try_into_int"]};
