+++ extendr_api/sidebar-items.js
@@ -1,1 +1,1 @@
-window.SIDEBAR_ITEMS = {"attr":["extendr"],"constant":["FALSE","NA_INTEGER","NA_LOGICAL","NA_REAL","NA_STRING","NULL","PRINTF_NO_FMT_CSTRING","TRUE"],"derive":["IntoDataFrameRow","IntoRobj","TryFromRobj"],"enum":["Rany","Rtype"],"fn":["make_method_def","rtype_to_sxp","sxp_to_rtype"],"macro":["R","Rraw","call","data_frame","extendr_module","factor","global","lang","list","pairlist","r","reprint","reprintln","rprint","rprintln","sym","test","var"],"mod":["deserializer","error","functions","graphics","io","iter","lang_macros","metadata","na","optional","ownership","prelude","rmacros","robj","scalar","serializer","thread_safety","wrapper"],"trait":["Deref","DerefMut","TryFrom","TryInto"]};
+window.SIDEBAR_ITEMS = {"attr":["extendr"],"constant":["FALSE","NA_INTEGER","NA_LOGICAL","NA_REAL","NA_STRING","NULL","PRINTF_NO_FMT_CSTRING","TRUE"],"derive":["IntoDataFrameRow","IntoRobj","TryFromRobj"],"enum":["Rany","Rtype"],"fn":["make_method_def","rtype_to_sxp","sxp_to_rtype"],"macro":["R","Rraw","call","data_frame","extendr_module","factor","global","lang","list","pairlist","r","reprint","reprintln","rprint","rprintln","sym","test","var"],"mod":["conversions","deserializer","error","functions","graphics","io","iter","lang_macros","metadata","na","optional","ownership","prelude","rmacros","robj","scalar","serializer","thread_safety","wrapper"],"trait":["Deref","DerefMut","TryFrom","TryInto"]};

[…]
