### 2024-05-13 4df996a
Rewrite try_into_int in terms of .classify()
