### 2024-05-13 d7185da
refactor: extendr-options is used by many scripts already
