+++ extendr-api/src/wrapper/matrix.rs
@@ -1,7 +1,6 @@
 //! Wrappers for matrices with deferred arithmetic.
-
+use self::robj::{AsTypedSlice, Robj};
 use super::*;
-use crate::robj::GetSexp;
 use crate::scalar::Scalar;
 use std::ops::{Index, IndexMut};
 
@@ -31,6 +30,7 @@ pub struct RArray<T, D> {
 
     /// Dimensions of the array.
     dim: D,
+
     _data: std::marker::PhantomData<T>,
 }
 
@@ -66,7 +66,6 @@ where
     pub fn new_with_na(nrow: usize, ncol: usize) -> Self {
         let mut matrix = Self::new(nrow, ncol);
         if nrow != 0 || ncol != 0 {
-            // matrix.data_mut().iter_mut().for_each(|x| {
             matrix
                 .as_typed_slice_mut()
                 .unwrap()
@@ -125,10 +124,9 @@ impl<T> Offset<[usize; 3]> for RArray<T, [usize; 3]> {
     }
 }
 
-impl<'a, T, D> RArray<T, D>
+impl<T, D> RArray<T, D>
 where
-    T: 'a,
-    Robj: AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     pub fn from_parts(robj: Robj, dim: D) -> Self {
         Self {
@@ -139,12 +137,12 @@ where
     }
 
     /// Returns a flat representation of the array in col-major.
-    pub fn data(&self) -> &'a [T] {
+    pub fn data(&self) -> &[T] {
         self.as_typed_slice().unwrap()
     }
 
-    /// Returns a flat mutable representation of the array in col-major.
-    pub fn data_mut(&mut self) -> &'a mut [T] {
+    /// Returns a flat, mutable representation of the array in col-major.
+    pub fn data_mut(&mut self) -> &mut [T] {
         self.as_typed_slice_mut().unwrap()
     }
 
@@ -154,9 +152,10 @@ where
     }
 }
 
-impl<'a, T: ToVectorValue + 'a> RColumn<T>
+impl<T> RColumn<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    T: ToVectorValue,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     /// Make a new column type.
     pub fn new_column<F: FnMut(usize) -> T>(nrows: usize, f: F) -> Self {
@@ -172,9 +171,10 @@ where
     }
 }
 
-impl<'a, T: ToVectorValue + 'a> RMatrix<T>
+impl<T> RMatrix<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    T: ToVectorValue,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     /// Create a new matrix wrapper.
     ///
@@ -214,9 +214,10 @@ where
     }
 }
 
-impl<'a, T: ToVectorValue + 'a> RMatrix3D<T>
+impl<T> RMatrix3D<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    T: ToVectorValue,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     pub fn new_matrix3d<F: Clone + FnMut(usize, usize, usize) -> T>(
         nrows: usize,
@@ -254,71 +255,120 @@ where
     }
 }
 
-impl<'a, T: 'a> TryFrom<Robj> for RColumn<T>
+impl<T> TryFrom<&Robj> for RColumn<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     type Error = Error;
 
[…]
