+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -33,34 +33,37 @@
           }
           #[cfg(use_r_altlist)]
           impl VecUsize {}
-          impl<'a> extendr_api::FromRobj<'a> for &VecUsize {
-              fn from_robj(robj: &'a Robj) -> std::result::Result<Self, &'static str> {
-                  if robj.check_external_ptr_type::<VecUsize>() {
-                      #[allow(clippy::transmute_ptr_to_ref)]
-                      Ok(unsafe { std::mem::transmute(robj.external_ptr_addr::<VecUsize>()) })
-                  } else {
-                      Err("expected VecUsize")
-                  }
+          impl TryFrom<Robj> for &VecUsize {
+              type Error = Error;
+              fn try_from(robj: Robj) -> Result<Self> {
+                  Self::try_from(&robj)
               }
           }
-          impl<'a> extendr_api::FromRobj<'a> for &mut VecUsize {
-              fn from_robj(robj: &'a Robj) -> std::result::Result<Self, &'static str> {
-                  if robj.check_external_ptr_type::<VecUsize>() {
-                      #[allow(clippy::transmute_ptr_to_ref)]
-                      Ok(unsafe { std::mem::transmute(robj.external_ptr_addr::<VecUsize>()) })
-                  } else {
-                      Err("expected VecUsize")
-                  }
+          impl TryFrom<Robj> for &mut VecUsize {
+              type Error = Error;
+              fn try_from(mut robj: Robj) -> Result<Self> {
+                  Self::try_from(&mut robj)
               }
           }
-          impl<'a> From<&'a VecUsize> for Robj {
-              fn from(value: &'a VecUsize) -> Self {
+          impl TryFrom<&Robj> for &VecUsize {
+              type Error = Error;
+              fn try_from(robj: &Robj) -> Result<Self> {
+                  use libR_sys::R_ExternalPtrAddr;
                   unsafe {
-                      let ptr = Box::into_raw(Box::new(value));
-                      let mut res = Robj::make_external_ptr(ptr, Robj::from(()));
-                      res.set_attrib(class_symbol(), "VecUsize").unwrap();
-                      res.register_c_finalizer(Some(__finalize__VecUsize));
-                      res
+                      let ptr = R_ExternalPtrAddr(robj.get()).cast::<VecUsize>();
+                      ptr.as_ref()
+                          .ok_or_else(|| Error::ExpectedExternalNonNullPtr(robj.clone()))
+                  }
+              }
+          }
+          impl TryFrom<&mut Robj> for &mut VecUsize {
+              type Error = Error;
+              fn try_from(robj: &mut Robj) -> Result<Self> {
+                  use libR_sys::R_ExternalPtrAddr;
+                  unsafe {
+                      let ptr = R_ExternalPtrAddr(robj.get_mut()).cast::<VecUsize>();
+                      ptr.as_mut()
+                          .ok_or_else(|| Error::ExpectedExternalNonNullPtr(robj.clone()))
                   }
               }
           }
@@ -97,7 +100,8 @@
           #[cfg(use_r_altlist)]
           impl AltListImpl for VecUsize {
               fn elt(&self, index: usize) -> Robj {
-                  self.into_robj()
+                  Self(<[_]>::into_vec(#[rustc_box] ::alloc::boxed::Box::new([self.0[index]])))
+                      .into()
               }
           }
           #[cfg(use_r_altlist)]
@@ -127,13 +131,7 @@
                       std::panic::AssertUnwindSafe(|| -> std::result::Result<
                           Robj,
                           extendr_api::Error,
-                      > {
-                          Ok(
-                              extendr_api::Robj::from(
-                                  new_usize(<Integers>::from_robj(&_robj_robj)?),
-                              ),
-                          )
-                      }),
[…]
