### 2024-05-13 ce2bc17
remove specific error message name from testing
