+++ tests/extendrtests/tests/testthat/test-type-conversion.R
@@ -1,49 +1,49 @@
 test_that("From conversion of R types to Rust types and vice versa works", {
   expect_equal(double_scalar(.45), .45)
   expect_equal(double_scalar(15L), 15)
-  expect_error(double_scalar(TRUE), "Unable to convert R object to primitive")
-  expect_error(double_scalar("abcxyz"), "Unable to convert R object to primitive")
-  expect_error(double_scalar(NA_real_), "Input must not be NA")
-  expect_error(double_scalar(c(.45, .46)), "Input must be of length 1")
+  expect_error(double_scalar(TRUE), "Expected Numeric, got Logicals")
+  expect_error(double_scalar("abcxyz"), "Expected Numeric, got Strings")
+  expect_error(double_scalar(NA_real_), "Must not be NA")
+  expect_error(double_scalar(c(.45, .46)), "Expected Scalar, got Doubles")
   
   expect_equal(int_scalar(15L), 15L)
-  expect_error(int_scalar(4.4), "Input must be a whole integer") # is this deliberate? seems dangerous
-  expect_error(int_scalar(TRUE), "Unable to convert R object to primitive")
-  expect_error(int_scalar("abcxyz"), "Unable to convert R object to primitive")
-  expect_error(int_scalar(NA_integer_), "Input must not be NA")
-  expect_error(int_scalar(1L:5L), "Input must be of length 1")
+  expect_error(int_scalar(4.4))
+  expect_error(int_scalar(TRUE))
+  expect_error(int_scalar("abcxyz"))
+  expect_error(int_scalar(NA_integer_))
+  expect_error(int_scalar(1L:5L))
   
   expect_equal(bool_scalar(TRUE), TRUE)
   expect_equal(bool_scalar(FALSE), FALSE)
-  expect_error(bool_scalar(.45), "Not a logical object")
-  expect_error(bool_scalar(15L), "Not a logical object")
-  expect_error(bool_scalar("abcxyz"), "Not a logical object")
-  expect_error(bool_scalar(NA), "Input must not be NA")
-  expect_error(bool_scalar(c(TRUE, FALSE, TRUE)), "Input must be of length 1")
+  expect_error(bool_scalar(.45))
+  expect_error(bool_scalar(15L))
+  expect_error(bool_scalar("abcxyz"))
+  expect_error(bool_scalar(NA))
+  expect_error(bool_scalar(c(TRUE, FALSE, TRUE)))
   
   expect_equal(char_scalar("abcxyz"), "abcxyz")
-  expect_error(char_scalar(.45), "not a string object")
-  expect_error(char_scalar(15L), "not a string object")
-  expect_error(char_scalar(TRUE), "not a string object")
-  expect_error(char_scalar(NA_character_), "Input must not be NA")
-  expect_error(char_scalar(c("hello", "world")), "not a string object") # why this error message and not "Input must be of length 1"?
+  expect_error(char_scalar(.45))
+  expect_error(char_scalar(15L))
+  expect_error(char_scalar(TRUE))
+  expect_error(char_scalar(NA_character_))
+  expect_error(char_scalar(c("hello", "world")))
 
   expect_equal(char_vec(c("hello", "world")), c("hello", "world"))
-  expect_error(char_vec(.45), "Input must be a character vector")
-  expect_error(char_vec(15L), "Input must be a character vector")
-  expect_error(char_vec(TRUE), "Input must be a character vector")
-  expect_error(char_vec(NA_character_), "Input must be a character vector. Got 'NA'.")
-  expect_error(char_vec(c("hello", NA)), "Input vector cannot contain NA's")
+  expect_error(char_vec(.45))
+  expect_error(char_vec(15L))
[…]
