### 2024-05-13 c2ba204
MSRV fix `cargo clippy` lints
