+++ extendr-api/src/wrapper/doubles.rs
@@ -96,9 +96,7 @@ impl TryFrom<Vec<f64>> for Doubles {
     type Error = Error;
 
     fn try_from(value: Vec<f64>) -> std::result::Result<Self, Self::Error> {
-        Ok(Self {
-            robj: <Robj>::try_from(value)?,
-        })
+        Ok(Self { robj: value.into() })
     }
 }
 
