+++ extendr-api/tests/dataframe_tests.rs
@@ -62,7 +62,7 @@ fn dataframe_conversion(_data_frame: Dataframe<Row>) -> Robj {
     vec![Row { name: 42 }].into_dataframe().unwrap().into_robj()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn dataframe_conversion_try_from(_data_frame: Dataframe<Row>) -> Robj {
     vec![Row { name: 42 }].into_dataframe().unwrap().into_robj()
 }
@@ -72,7 +72,7 @@ fn return_dataframe(_data_frame: Dataframe<Row>) -> Dataframe<Row> {
     vec![Row { name: 42 }].into_dataframe().unwrap()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn return_dataframe_try_from(_data_frame: Dataframe<Row>) -> Dataframe<Row> {
     vec![Row { name: 42 }].into_dataframe().unwrap()
 }
@@ -107,7 +107,7 @@ fn test_dataframe_row_and_externalptr() {
         data_frame
     }
 
-    #[extendr(use_try_from = true)]
+    #[extendr]
     fn use_dataframe_exptr3(data_frame: Dataframe<RowAndExptr>) -> Dataframe<RowAndExptr> {
         data_frame
     }
