### 2024-05-13 #6394 @jbytecode
@CGMossa - Please consider reviewing/changing/adding the suggestions of our reviewer. You can of course give answers and make comments on the suggestions. Those are the nautral steps of the review process of JOSS.

While you prepare a proposition and make changes, you can suggest me a new list of potential authors if you have. If you don't have one, I can search and assign a new reviewer.

Another option is to finish the first step review with @dccsillag and then wait for the response of @luciorq again. What do you thing?

@luciorq - We really need your thoughts and review, otherwise, we will need to find and assign another reviewer. Could you please start your review and help us?
