### 2024-05-13 25c5c3d
Convert integer-ish values based on `f64::classify()` (#758)
* Rewrite try_into_int in terms of .classify()

* Format

* Use consts
