### 2024-05-13 14f2205
Update extendr-api/src/iter.rs
Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
