### 2024-05-13 e4d0431
Update CHANGELOG.md
Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
