### 2024-05-13 ae0d9e6
`use-try-from = true` as default (#759)
* refactor: extendr-options is used by many scripts already

* remove `FromRobj`
passing `cargo check --tests`

* `tests/extendrtests` compiles!

* updated snapshots

* remove specific error message name
from testing

* rebase was funky plus lints

* MSRV fix
`cargo clippy` lints

* removed all mention of
`use_try_from` from the codebase.

Removed duplicated integration tests
for `TryFrom`-standards.

* added the previous tests back in

* removed `use_try_from` completely

* fixed compile tests

* ndarray: it still had some `FromRobj`. It is fixed now.


[…]
