+++ tests/extendrtests/src/rust/src/lib.rs
@@ -21,7 +21,7 @@ fn hello_world() -> &'static str {
 #[extendr]
 fn do_nothing() {}
 
-// From: input/output conversion
+// TryFrom: conversions
 
 // Atomic types
 
@@ -68,60 +68,50 @@ fn double_vec(x: Vec<f64>) -> Vec<f64> {
 }
 
 // NA-related tests
-#[extendr(use_try_from = true)]
+#[extendr]
 fn try_rfloat_na() -> Rfloat {
     Rfloat::na()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn try_rint_na() -> Rint {
     Rint::na()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn check_rfloat_na(x: Rfloat) -> bool {
     x.is_na()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn check_rint_na(x: Rint) -> bool {
     x.is_na()
 }
 
 // Non-atomic types
-// TODO
-
-// TryFrom: conversions
-
+// TOD
 // Atomic types
-
-// Convert a vector of doubles to itself
-// x a vector of doubles
-#[extendr(use_try_from = true)]
-fn try_double_vec(x: Vec<f64>) -> Vec<f64> {
-    x
-}
 
 // Non-atomic types
 // TODO
 
 // Vector wrappers
-#[extendr(use_try_from = true)]
+#[extendr]
 fn get_doubles_element(x: Doubles, i: i32) -> Rfloat {
     x.elt(i as usize)
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn get_integers_element(x: Integers, i: i32) -> Rint {
     x.elt(i as usize)
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn get_logicals_element(x: Logicals, i: i32) -> Rbool {
     x.elt(i as usize)
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn doubles_square(input: Doubles) -> Doubles {
     let mut result = Doubles::new(input.len());
 
@@ -132,7 +122,7 @@ fn doubles_square(input: Doubles) -> Doubles {
     result
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn complexes_square(input: Complexes) -> Complexes {
     let mut result = Complexes::new(input.len());
 
@@ -143,7 +133,7 @@ fn complexes_square(input: Complexes) -> Complexes {
     result
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn integers_square(input: Integers) -> Integers {
     let mut result = Integers::new(input.len());
 
@@ -154,7 +144,7 @@ fn integers_square(input: Integers) -> Integers {
     result
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn logicals_not(input: Logicals) -> Logicals {
     let mut result = Logicals::new(input.len());
 
@@ -167,7 +157,7 @@ fn logicals_not(input: Logicals) -> Logicals {
 
 // Parsing
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn check_default(#[default = "NULL"] x: Robj) -> bool {
     x.is_null()
 }
@@ -191,11 +181,7 @@ fn special_param_names(_x: i32, _y: i32) -> i32 {
 #[allow(non_snake_case)]
 fn __00__special_function_name() {}
 
-#[extendr(
-    use_try_from = true,
-    r_name = "test.rename.rlike",
-    mod_name = "test_rename_mymod"
-)]
[…]
