### 2024-05-14 db408f2
producing a string should use `length` of `CHARSXP`
