+++ extendr-api/src/wrapper/pairlist.rs
@@ -138,8 +138,9 @@ impl Iterator for PairlistIter {
                 let value = Robj::from_sexp(CAR(sexp));
                 self.list_elem = CDR(sexp);
                 if TYPEOF(tag) == SEXPTYPE::SYMSXP {
+                    // printname is always a CHARSXP
                     let printname = PRINTNAME(tag);
-                    assert_eq!(TYPEOF(printname),  SEXPTYPE::CHARSXP); // printname is always a CHARSXP
+                    assert_eq!(TYPEOF(printname), SEXPTYPE::CHARSXP);
                     //TODO: use RawRobj here once it is a thing
                     let length = Rf_xlength(printname);
                     let all_bytes = std::slice::from_raw_parts(R_CHAR(printname) as _, length as _);
