### 2024-05-14 2324cc1
either the iterator is finished, or the string is nothing, or the i'th charsxp has to be extracted, or it is a factor
