### 2024-05-14 0d43cd0
Rstr is a synonym for charsxp. And the provided `From` basically converts CHARSXP properly
