### 2024-05-14 96a907f
extendr-api: Add a `Factors` type
