### 2024-05-14 8609f41
Moved duplicated code into its own function
