### 2024-05-14 0ad24e8
trimming / refactoring
