### 2024-05-15 a3f1f2f
Removed explicit reference to `libR-sys` in derive-macro. Complete reliance on `ExternalPtr<T>` for implementation of `#[extendr]`-impl => Therefore removed finalizer.
