+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -48,42 +48,27 @@
           impl TryFrom<&Robj> for &VecUsize {
               type Error = Error;
               fn try_from(robj: &Robj) -> Result<Self> {
-                  use libR_sys::R_ExternalPtrAddr;
                   unsafe {
-                      let ptr = R_ExternalPtrAddr(robj.get()).cast::<VecUsize>();
-                      ptr.as_ref()
-                          .ok_or_else(|| Error::ExpectedExternalNonNullPtr(robj.clone()))
+                      let external_ptr: &ExternalPtr<VecUsize> = robj.try_into()?;
+                      external_ptr.try_addr()
                   }
               }
           }
           impl TryFrom<&mut Robj> for &mut VecUsize {
               type Error = Error;
               fn try_from(robj: &mut Robj) -> Result<Self> {
-                  use libR_sys::R_ExternalPtrAddr;
                   unsafe {
-                      let ptr = R_ExternalPtrAddr(robj.get_mut()).cast::<VecUsize>();
-                      ptr.as_mut()
-                          .ok_or_else(|| Error::ExpectedExternalNonNullPtr(robj.clone()))
+                      let external_ptr: &mut ExternalPtr<VecUsize> = robj.try_into()?;
+                      external_ptr.try_addr_mut()
                   }
               }
           }
           impl From<VecUsize> for Robj {
               fn from(value: VecUsize) -> Self {
                   unsafe {
-                      let ptr = Box::into_raw(Box::new(value));
-                      let mut res = Robj::make_external_ptr(ptr, Robj::from(()));
+                      let mut res: ExternalPtr<VecUsize> = ExternalPtr::new(value);
                       res.set_attrib(class_symbol(), "VecUsize").unwrap();
-                      res.register_c_finalizer(Some(__finalize__VecUsize));
-                      res
-                  }
-              }
-          }
-          extern "C" fn __finalize__VecUsize(sexp: extendr_api::SEXP) {
-              unsafe {
-                  let robj = extendr_api::robj::Robj::from_sexp(sexp);
-                  if robj.check_external_ptr_type::<VecUsize>() {
-                      let ptr = robj.external_ptr_addr::<VecUsize>();
-                      drop(Box::from_raw(ptr));
+                      res.into()
                   }
               }
           }
@@ -4244,42 +4229,27 @@
           impl TryFrom<&Robj> for &MySubmoduleClass {
               type Error = Error;
               fn try_from(robj: &Robj) -> Result<Self> {
-                  use libR_sys::R_ExternalPtrAddr;
                   unsafe {
-                      let ptr = R_ExternalPtrAddr(robj.get()).cast::<MySubmoduleClass>();
-                      ptr.as_ref()
-                          .ok_or_else(|| Error::ExpectedExternalNonNullPtr(robj.clone()))
+                      let external_ptr: &ExternalPtr<MySubmoduleClass> = robj.try_into()?;
+                      external_ptr.try_addr()
                   }
               }
           }
           impl TryFrom<&mut Robj> for &mut MySubmoduleClass {
               type Error = Error;
               fn try_from(robj: &mut Robj) -> Result<Self> {
-                  use libR_sys::R_ExternalPtrAddr;
                   unsafe {
-                      let ptr = R_ExternalPtrAddr(robj.get_mut()).cast::<MySubmoduleClass>();
-                      ptr.as_mut()
-                          .ok_or_else(|| Error::ExpectedExternalNonNullPtr(robj.clone()))
+                      let external_ptr: &mut ExternalPtr<MySubmoduleClass> = robj.try_into()?;
[…]
