### 2024-05-15 5b0fbea
[urgent!] `#[extendr]`-impl requires `libR-sys` (#762)
* Removed explicit reference to `libR-sys` in derive-macro.
Complete reliance on `ExternalPtr<T>` for
implementation of `#[extendr]`-impl
=> Therefore removed finalizer.
