+++ extendr_macros/wrappers/sidebar-items.js
@@ -1,1 +1,1 @@
-window.SIDEBAR_ITEMS = {"constant":["META_PREFIX","WRAP_PREFIX"],"fn":["get_doc_string","get_named_lit","get_return_type","make_function_wrappers","mangled_type_name","sanitize_identifier","translate_actual","translate_formal","translate_meta_arg","translate_only_alias","translate_to_robj","type_name"],"struct":["ExtendrOptions"]};
+window.SIDEBAR_ITEMS = {"constant":["META_PREFIX","WRAP_PREFIX"],"fn":["get_doc_string","get_named_lit","get_return_type","make_function_wrappers","mangled_type_name","sanitize_identifier","translate_actual","translate_formal","translate_meta_arg","translate_only_alias","translate_to_robj","type_name"]};
