### 2024-05-15 e03c9b7
Clean-up after #759 (#760)
* merged tests

* remove explicit dependence on `rextendr` branch
