+++ extendr-api/src/robj/tests.rs
@@ -3,7 +3,7 @@ use crate::scalar::*;
 use crate::*;
 
 #[test]
-fn test_try_from() {
+fn test_try_from_robj() {
     test! {
         assert_eq!(<bool>::try_from(&Robj::from(true)), Ok(true));
         assert_eq!(<u8>::try_from(&Robj::from(1)), Ok(1));
@@ -19,11 +19,33 @@ fn test_try_from() {
 
         assert_eq!(<Vec::<i32>>::try_from(&Robj::from(1)), Ok(vec![1]));
         assert_eq!(<Vec::<f64>>::try_from(&Robj::from(1.)), Ok(vec![1.]));
+        assert_eq!(<Vec::<Rint>>::try_from(Robj::from(1)), Ok(vec![Rint::from(1)]));
+        assert_eq!(<Vec::<Rfloat>>::try_from(Robj::from(1.)), Ok(vec![Rfloat::from(1.0)]));
+        assert_eq!(<Vec::<Rbool>>::try_from(Robj::from(TRUE)), Ok(vec![TRUE]));
+        assert_eq!(<Vec::<u8>>::try_from(Robj::from(0_u8)), Ok(vec![0_u8]));
+
+        // conversion from non-integer-ish value to integer should fail
+        let robj = Robj::from(1.5);
+        assert_eq!(<i32>::try_from(robj.clone()), Err(Error::ExpectedWholeNumber(robj, ConversionError::NotIntegerish)));
+
+        // conversion from out-of-limit value should fail
+        let robj = Robj::from(32768);
+        assert_eq!(<i16>::try_from(robj.clone()), Err(Error::OutOfLimits(robj)));
+        let robj = Robj::from(-1);
+        assert_eq!(<u32>::try_from(robj.clone()), Err(Error::OutOfLimits(robj)));
 
         let hello = Robj::from("hello");
         assert_eq!(<&str>::try_from(&hello), Ok("hello"));
+        let hello = Robj::from("hello");
+        assert_eq!(<String>::try_from(hello), Ok("hello".into()));
 
         // conversion from a vector to a scalar value
+
+        let robj = Robj::from(vec![].as_slice() as &[i32]);
+        assert_eq!(
+            <i32>::try_from(robj.clone()),
+            Err(Error::ExpectedNonZeroLength(robj))
+        );
         assert!(
             <i32>::try_from(&Robj::from(vec![].as_slice() as &[i32])).is_err()
         );
@@ -33,6 +55,11 @@ fn test_try_from() {
         );
         assert!(
             <i32>::try_from(&Robj::from(vec![1, 2].as_slice() as &[i32])).is_err()
+        );
+        let robj = Robj::from(vec![1, 2].as_slice() as &[i32]);
+        assert_eq!(
+            <i32>::try_from(robj.clone()),
+            Err(Error::ExpectedScalar(robj))
         );
 
         use std::collections::HashMap;
@@ -55,6 +82,8 @@ fn test_try_from() {
 
         assert_eq!(hmap_borrowed["a"], Robj::from(1));
         assert_eq!(hmap_borrowed["b"], Robj::from(2));
+        let hmap_borrowed = list.as_list().unwrap().into_hashmap();
+        assert_eq!(hmap_borrowed, hmap2);
 
         let na_integer = eval_string("NA_integer_").unwrap();
         assert!(<i32>::try_from(&na_integer).is_err());
@@ -88,37 +117,7 @@ fn test_try_from() {
             Ok(Some("1".to_string()))
         );
         assert!(<Option<String>>::try_from(&Robj::from(["1", "2"])).is_err());
-    }
-}
 
-#[test]
-fn test_try_from_robj() {
[…]
