### 2024-05-15 4c6d3c7
extendrtests: added an integration test
