+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -3505,6 +3505,16 @@
               fn me_explicit_mut(&mut self) -> &mut MySubmoduleClass {
                   self
               }
+              fn max_ref(&self, other: &'static MySubmoduleClass) -> &Self {
+                  if self.a > other.a { self } else { other }
+              }
+              fn max_ref_offset(
+                  &self,
+                  other: &'static MySubmoduleClass,
+                  _offset: i32,
+              ) -> &Self {
+                  if self.a > other.a { self } else { other }
+              }
           }
           #[no_mangle]
           #[allow(non_snake_case, clippy::not_unsafe_ptr_arg_deref)]
@@ -3874,11 +3884,17 @@
                           Robj,
                           extendr_api::Error,
                       > {
-                          let _return_ref_to_self = extendr_api::unwrap_or_throw_error(
+                          let return_ref_to_self = extendr_api::unwrap_or_throw_error(
                                   <&MySubmoduleClass>::try_from(&_self_robj),
                               )
                               .me_ref();
-                          Ok(_self_robj)
+                          if std::ptr::addr_eq(
+                              libR_sys::R_ExternalPtrAddr(_self),
+                              std::ptr::from_ref(return_ref_to_self),
+                          ) {
+                              return Ok(extendr_api::Robj::from_sexp(_self));
+                          }
+                          Err(Error::ExpectedExternalPtrReference)
                       }),
                   )
               };
@@ -3962,11 +3978,17 @@
                           Robj,
                           extendr_api::Error,
                       > {
-                          let _return_ref_to_self = extendr_api::unwrap_or_throw_error(
+                          let return_ref_to_self = extendr_api::unwrap_or_throw_error(
                                   <&mut MySubmoduleClass>::try_from(&mut _self_robj),
                               )
                               .me_mut();
-                          Ok(_self_robj)
+                          if std::ptr::addr_eq(
+                              libR_sys::R_ExternalPtrAddr(_self),
+                              std::ptr::from_ref(return_ref_to_self),
+                          ) {
+                              return Ok(extendr_api::Robj::from_sexp(_self));
+                          }
+                          Err(Error::ExpectedExternalPtrReference)
                       }),
                   )
               };
@@ -4050,11 +4072,17 @@
                           Robj,
                           extendr_api::Error,
                       > {
-                          let _return_ref_to_self = extendr_api::unwrap_or_throw_error(
+                          let return_ref_to_self = extendr_api::unwrap_or_throw_error(
                                   <&MySubmoduleClass>::try_from(&_self_robj),
                               )
                               .me_explicit_ref();
-                          Ok(_self_robj)
+                          if std::ptr::addr_eq(
+                              libR_sys::R_ExternalPtrAddr(_self),
+                              std::ptr::from_ref(return_ref_to_self),
+                          ) {
+                              return Ok(extendr_api::Robj::from_sexp(_self));
+                          }
+                          Err(Error::ExpectedExternalPtrReference)
                       }),
                   )
               };
@@ -4140,11 +4168,17 @@
                           Robj,
                           extendr_api::Error,
                       > {
-                          let _return_ref_to_self = extendr_api::unwrap_or_throw_error(
+                          let return_ref_to_self = extendr_api::unwrap_or_throw_error(
                                   <&mut MySubmoduleClass>::try_from(&mut _self_robj),
                               )
                               .me_explicit_mut();
-                          Ok(_self_robj)
+                          if std::ptr::addr_eq(
+                              libR_sys::R_ExternalPtrAddr(_self),
+                              std::ptr::from_ref(return_ref_to_self),
+                          ) {
+                              return Ok(extendr_api::Robj::from_sexp(_self));
+                          }
[…]
