### 2024-05-15 #6394 @alpaylan
> @kdpsingh, @alpaylan - One of our reviewers hasn't responded for a while. I can't reach them even in Twitter. A need a second reviewer for this submission. Would you willing to assist in reviewing this submission for JOSS?

Hi, I can be the second reviewer for this submission.
