### 2024-05-15 d4797e6
add test where other is of type `&Self`. why not?
