+++ tests/extendrtests/src/rust/src/submodule.rs
@@ -70,7 +70,6 @@ impl MySubmoduleClass {
         self
     }
 
-    // fn max_ref(&self, other: &Self) -> &Self {
     fn max_ref(&self, other: &'static MySubmoduleClass) -> &Self {
         if self.a > other.a {
             self
@@ -86,6 +85,15 @@ impl MySubmoduleClass {
             other
         }
     }
+
+    fn max_ref2(&self, other: &'static Self) -> &Self {
+        if self.a > other.a {
+            self
+        } else {
+            other
+        }
+    }
+
 }
 
 // Macro to generate exports
