### 2024-05-16 7800395
use `cast` instead of `as`
