### 2024-05-16 c3fa6c7
STANDALONE: ExternalPtr is treated like a pointer, and comparisons don't work directly on it. Therefore, this behavior is now documented, and a way to circumvent this limitation is explained.
Make into a separate PR probably...
