### 2024-05-16 e990a60
STANDALONE: Clean-up.. We actually don't use this bit anywhere, and frankly the stringly type checking needs to go away..
