+++ tests/extendrtests/src/rust/src/submodule.rs
@@ -93,7 +93,6 @@ impl MySubmoduleClass {
             other
         }
     }
-
 }
 
 // Macro to generate exports
