### 2024-05-16 0267fad
out of bound error case handled
