### 2024-05-16 7c680e4
segfault if the index is _negative_ so we ought to be sure about that
