+++ extendr-api/src/iter.rs
@@ -214,22 +214,3 @@ pub trait AsStrIter: GetSexp + Types + Length + Attributes + Rinternals {
 }
 
 impl AsStrIter for Robj {}
-
-#[cfg(test)]
-mod tests {
-    use extendr_engine::with_r;
-
-    use super::*;
-
-    #[test]
-    fn string_elt_beyond_limit() {
-        with_r(|| unsafe {
-            let a = Rf_ScalarString(R_NaString);
-            Rf_PrintValue(STRING_ELT(a, 0));
-            Rf_PrintValue(STRING_ELT(a, 2));
-            // Rf_PrintValue(STRING_ELT(a, -1)); // segfault
-            // let bad = STRING_ELT(a, 2);
-            // dbg!(str_from_strsxp(a, 2));
-        });
-    }
-}
