### 2024-05-16 b6be3fc
added a test for `CHARSXP` to StrIter conversion
