### 2024-05-16 fd7da4a
refactored things.. safer
