+++ extendr-api/src/iter.rs
@@ -233,57 +233,4 @@ mod tests {
             assert_eq!(s2.len(), 1);
         });
     }
-
-    #[test]
-    fn feature() {
-        with_r(|| {
-            let obj = Robj::from(vec!["a", "b", "c"]);
-            assert_eq!(
-                obj.as_str_iter().unwrap().collect::<Vec<_>>(),
-                vec!["a", "b", "c"]
-            );
-
-            let factor = factor!(vec!["abcd", "def", "fg", "fg"]);
-            assert_eq!(
-                factor.levels().unwrap().collect::<Vec<_>>(),
-                vec!["abcd", "def", "fg"]
-            );
-            assert_eq!(factor.as_integer_vector().unwrap(), vec![1, 2, 3, 3]);
-            assert_eq!(
-                factor.as_str_iter().unwrap().collect::<Vec<_>>(),
-                vec!["abcd", "def", "fg", "fg"]
-            );
-            assert_eq!(
-                factor.as_str_iter().unwrap().collect::<Vec<_>>(),
-                vec!["abcd", "def", "fg", "fg"]
-            );
-
-            let obj = Robj::from(vec![Some("a"), Some("b"), None]);
-            assert_eq!(
-                obj.as_str_iter()
-                    .unwrap()
-                    .map(|s| s.is_na())
-                    .collect::<Vec<_>>(),
-                vec![false, false, true]
-            );
-
-            let obj = Robj::from(vec!["a", "b", <&str>::na()]);
-            assert_eq!(
-                obj.as_str_iter()
-                    .unwrap()
-                    .map(|s| s.is_na())
-                    .collect::<Vec<_>>(),
-                vec![false, false, true]
-            );
-
-            let obj = Robj::from(vec!["a", "b", "NA"]);
-            assert_eq!(
-                obj.as_str_iter()
-                    .unwrap()
-                    .map(|s| s.is_na())
-                    .collect::<Vec<_>>(),
-                vec![false, false, false]
-            );
-        })
-    }
 }
