+++ extendr-api/src/robj/tests.rs
@@ -1,6 +1,7 @@
 use crate::conversions::try_into_int::ConversionError;
 use crate::scalar::*;
 use crate::*;
+use extendr_engine::with_r;
 
 #[test]
 fn test_try_from_robj() {
@@ -288,4 +289,17 @@ fn input_iterator_test() {
         assert_eq!(iter.collect::<Vec<_>>(), src);
         */
     }
+}
+
+
+#[test]
+fn as_typed_slice_u32() {
+    with_r(|| unsafe{
+        let r = [-32, 3, i32::MAX].into_robj();
+        Rf_PrintValue(r.get());
+        dbg!(r.sexptype());
+        let r_slice: Option<&[u32]> = r.as_typed_slice();
+        dbg!(r_slice);
+        assert!(r_slice.is_none());
+    });
 }
