+++ extendr-api/src/prelude.rs
@@ -77,4 +77,3 @@ pub use ::either::*;
 
 #[cfg(feature = "faer")]
 pub use ::faer::*;
-
