### 2024-05-20 7a11963
added a (private) symmbol from libR-sys necessary in `#[extendr]`-impl.
