### 2024-05-20 6e66bcf
fix: remove wrong `u32` casting
