### 2024-05-20 6e36c2d
fix: removed invalid `u32` conversion
