### 2024-05-20 d9c1d58
Added the range checking for the unsigned integer slices
