### 2024-05-20 87b82d8
extendrtests: added null externalptr test this happens when an externalptr is deserialized from `RDS` file or similar
