### 2024-05-22 7e212bd
extendr-api: added `ToVectorValue` for `Rstr`
