+++ extendr-api/src/wrapper/mod.rs
@@ -175,7 +175,7 @@ make_conversions!(
 
 make_conversions!(Raw, ExpectedRaw, is_raw, "Not a raw object");
 
-make_conversions!(Rstr, ExpectedRstr, is_char, "Not a character object");
+make_getsexp!(Rstr, impl);
 
 make_conversions!(
     Environment,
