### 2024-05-22 2a221a8
tests: added RMatrix test with `Rstr`
