### 2024-05-22 7dc9fd0
standalone: no need to clone these, as the try-from definition is already defined for references
