### 2024-05-23 #6394 @CGMossa
Dear Editor,

extendr contributors and I have started working on a general, User Guide, for using R, Rust and extendr. It is a bit of an on-going work, see here https://extendr.github.io/user-guide/, but it already addresses many of the concerns presented in the first review.

Since the submission of the paper, we've merged 39 PRs, with 6,381 additions and 3,358 deletions. (see this https://github.com/extendr/extendr/compare/paper...master)
This is only mentioning extendr crate, as the accompanying `libR-sys` crate has also seen major changes.

Most of the changes has come about based on the feedback we've received here. We have not be too direct in our design choices, and mostly left it up to the users, to define certain behaviours, like the error handling. But since then, I've changed my mind, and now, I'm streamlining the design language of extendr. We've removed for instance the confusing `#[extendr(use_try_from = true)]` option.

I still need a bit of time to implement and to figure out what the default behavior for error handling should be. Even if it will be simply adopting one of the existing strategies. And the content of the user guide ought to have a few more additions.

While I'm here, let me address a few (not all!) the points brought up in the review:

- `r!` is indeed unnecessary, as `.into()` call is sufficient. However, the thing is with `.into()` is that it lacks the
target type. Usually, type inference can do the job, but it isn't explicit.
Thus, we have extension trait that gives you `.into_robj()`, which is explicit. And finally, `r!` is then a convenience macro.
- `test!` is a macro that embeds R into the test-runner. From R, you don't need to embed R, as it is present. But when
testing things in rust alone, for testing convenience, one may use `test!` macro that ensures R is running. This macro
also provides an ability to use `?` within test functions. We used to have a lot of inquiries about why calls to R would crash in Rust, when people would call R functions outside R. Thus, we have the `test!` macro everywhere to avoid this.
One more here, is that CRAN dislike code that embeds R in its packages. So, we encourage people to only use this
capability within test-code, that won't be part of the library, that's in R-packages.

Finally, I'd like to say, that the paper contained a "Getting Started" segment that contains the minimally necessary functions from `rextendr` to get started writing rust-powered R packages. I am convinced by the reviewer's opinion, that we should overhaul the first information presented to the user when entering https://extendr.github.io/. However, the current style of presenting API capabilities isn't different to what other packages in the same space do. Other contributors have also promised to take on this task, as they agree with the reviewer's take.
