### 2024-05-23 b4bb724
added the necessary elt impls on `RArray` that's good!
