+++ extendr-api/src/wrapper/strings.rs
@@ -105,6 +105,9 @@ impl Elt<Rstr> for Strings {
     }
 }
 
+//TODO: implement Elt<&str> maybe?
+
+
 impl Attributes for Strings {}
 
 impl<T: AsRef<str>> FromIterator<T> for Strings {
