### 2024-05-23 253f042
removed `u64`, `u128` from slice-conversion, becase there simply are not enough bytes for them in a [i32] slice.
