### 2024-05-25 aef5664
Update extendr-api/src/robj/mod.rs
Co-authored-by: Ilia Kosenkov <ilia.kosenkov@outlook.com>
