### 2024-05-25 0e29d32
Update extendr-api/src/wrapper/rstr.rs
Co-authored-by: Ilia Kosenkov <ilia.kosenkov@outlook.com>
