+++ extendr-api/src/wrapper/rstr.rs
@@ -17,7 +17,7 @@ pub struct Rstr {
     pub(crate) robj: Robj,
 }
 
-/// Returns a rust string-slice based on the provided `SEXP`, who is of type
+/// Returns a rust string-slice based on the provided `SEXP`, which is of type
 /// [`SEXPTYPE::CHARSXP`]. Note that the length of a `CHARSXP` is exactly
 /// the number of non-null bytes in said R character.
 pub(crate) unsafe fn charsxp_to_str(charsxp: SEXP) -> Option<&'static str> {
