### 2024-05-25 1d33e97
Faster string conversion (#761)
* either the iterator is finished,
or the string is nothing,
or the i'th charsxp has to be extracted, or it is a factor

* Rstr is a synonym for charsxp.
And the provided `From` basically converts CHARSXP properly

* segfault if the index is _negative_ so we ought to be sure about that

* added a test for `CHARSXP` to StrIter conversion


* feedback from @Ilia-Kosenkov

* Update extendr-api/src/robj/mod.rs

Co-authored-by: Ilia Kosenkov <ilia.kosenkov@outlook.com>

* Update extendr-api/src/robj/mod.rs

Co-authored-by: Ilia Kosenkov <ilia.kosenkov@outlook.com>

*
[…]
