+++ extendr_api/robj/sidebar-items.js
@@ -1,1 +1,1 @@
-window.SIDEBAR_ITEMS = {"fn":["to_str"],"macro":["make_typed_slice"],"mod":["debug","into_robj","operators","rinternals","try_from_robj"],"struct":["Robj"],"trait":["AsTypedSlice","Attributes","Eval","GetSexp","IntoRobj","Length","Operators","Rinternals","RobjItertools","Slices","ToVectorValue","Types"]};
+window.SIDEBAR_ITEMS = {"macro":["make_typed_slice"],"mod":["debug","into_robj","operators","rinternals","try_from_robj"],"struct":["Robj"],"trait":["AsTypedSlice","Attributes","Eval","GetSexp","IntoRobj","Length","Operators","Rinternals","RobjItertools","Slices","ToVectorValue","Types"]};
