### 2024-05-25 feef4cb
STANDALONE: `ExternalPtr` is a pointer (#765)
* STANDALONE: ExternalPtr is treated like a pointer,
and comparisons don't work directly on it. Therefore,
this behavior is now documented, and a way to circumvent
this limitation is explained.
