+++ trait.impl/core/convert/trait.AsMut.js
@@ -1,0 +1,3 @@
+(function() {var implementors = {
+"extendr_api":[["impl&lt;T&gt; <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.AsMut.html\" title=\"trait core::convert::AsMut\">AsMut</a>&lt;T&gt; for <a class=\"struct\" href=\"extendr_api/wrapper/externalptr/struct.ExternalPtr.html\" title=\"struct extendr_api::wrapper::externalptr::ExternalPtr\">ExternalPtr</a>&lt;T&gt;"]]
+};if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
