### 2024-05-26 13e4e4b
STANDALONE: faer isn't in the prelude (#764)
* STANDALONE: faer isn't in the prelude
