+++ extendr-api/src/prelude.rs
@@ -74,3 +74,6 @@ pub use ::ndarray::*;
 
 #[cfg(feature = "either")]
 pub use ::either::*;
+
+#[cfg(feature = "faer")]
+pub use ::faer::*;
