+++ extendr_api/prelude/modules/qr/col_pivoting/solve/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"fn":["solve","solve_in_place","solve_in_place_req","solve_req","solve_transpose","solve_transpose_in_place","solve_transpose_in_place_req","solve_transpose_req"]};
