+++ extendr_api/prelude/utils/simd/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"struct":["NoConj","Prefix","PrefixMut","SimdFor","Suffix","SuffixMut","YesConj"],"trait":["ConjTy","Read","Write"]};
