+++ extendr_api/prelude/modules/cholesky/bunch_kaufman/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"mod":["compute","solve"]};
