+++ extendr_api/prelude/utils/constrained/perm/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"struct":["PermRef"]};
