+++ extendr_api/prelude/dyn_stack/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"mod":["mem"],"struct":["DynArray","DynStack","GlobalMemBuffer","GlobalPodBuffer","PodStack","SizeOverflow","StackReq"],"trait":["ReborrowMut"]};
