+++ trait.impl/faer/linalg/zip/trait.MaybeContiguous.js
@@ -1,0 +1,3 @@
+(function() {var implementors = {
+"extendr_api":[]
+};if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
