+++ extendr_api/prelude/modules/qr/no_pivoting/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"mod":["compute","inverse","reconstruct","solve"]};
