+++ extendr_api/prelude/modules/cholesky/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"fn":["compute_cholesky_permutation"],"mod":["bunch_kaufman","ldlt_diagonal","llt"]};
