### 2024-05-26 ea3438c
Apply suggestions from code review
Co-authored-by: Ilia Kosenkov <ilia.kosenkov.at.gm@gmail.com>
