### 2024-05-26 b2ee9c9
split off externalptr tests instead of submodule
