+++ tests/extendrtests/src/rust/src/submodule.rs
@@ -37,62 +37,6 @@ impl MySubmoduleClass {
     fn a(&self) -> i32 {
         self.a
     }
-
-    // NOTE: Cannot move ownership, as that concept is incompatible with bridging
-    // data from R to Rust
-    // fn myself(self) -> Self {
-    //     self
-    // }
-
-    /// Method for getting one's (by way of a copy) self.
-    fn me_owned(&self) -> Self {
-        // only possible due to `derive(Clone, Copy)`
-        *self
-    }
-
-    /// Method for getting one's (ref) self.
-    fn me_ref(&self) -> &Self {
-        self
-    }
-
-    /// Method for getting one's (ref mut) self.
-    fn me_mut(&mut self) -> &mut Self {
-        self
-    }
-
-    /// Method for getting one's ref (explicit) self.
-    fn me_explicit_ref(&self) -> &MySubmoduleClass {
-        self
-    }
-
-    /// Method for getting one's ref mut (explicit) self.
-    fn me_explicit_mut(&mut self) -> &mut MySubmoduleClass {
-        self
-    }
-
-    fn max_ref(&self, other: &'static MySubmoduleClass) -> &Self {
-        if self.a > other.a {
-            self
-        } else {
-            other
-        }
-    }
-
-    fn max_ref_offset(&self, other: &'static MySubmoduleClass, _offset: i32) -> &Self {
-        if self.a > other.a {
-            self
-        } else {
-            other
-        }
-    }
-
-    fn max_ref2(&self, other: &'static Self) -> &Self {
-        if self.a > other.a {
-            self
-        } else {
-            other
-        }
-    }
 }
 
 // Macro to generate exports
