### 2024-05-27 9d548d3
`#[extendr]`-impl: Return the right reference (#726)
* extendrtests: added an integration test

* add test where other is of type `&Self`. why not?

* use `cast` instead of `as`

* added a (private) symbol from libR-sys necessary in
`#[extendr]`-impl.

* Apply suggestions from code review

Co-authored-by: Ilia Kosenkov <ilia.kosenkov.at.gm@gmail.com>

* split off externalptr tests instead of submodule

---------

Co-authored-by: Ilia Kosenkov <ilia.kosenkov.at.gm@gmail.com>
