+++ tests/extendrtests/NAMESPACE
@@ -4,13 +4,16 @@ S3method("$","__MyClass")
 S3method("$",MyClass)
 S3method("$",MyClassUnexported)
 S3method("$",MySubmoduleClass)
+S3method("$",Wrapper)
 S3method("[[","__MyClass")
 S3method("[[",MyClass)
 S3method("[[",MyClassUnexported)
 S3method("[[",MySubmoduleClass)
+S3method("[[",Wrapper)
 export("__00__special_function_name")
 export(MyClass)
 export(MySubmoduleClass)
+export(Wrapper)
 export(euclidean_dist)
 export(false)
 export(hello_submodule)

[…]
