### 2024-05-27 d8f8d37
wp: example of dropping resources due to `catch_unwind`.
