### 2024-05-27 e1cfb8e
wp: added split closure to c callback. Very nice.
