### 2024-05-27 defc4a3
wp: added diagnostic utility to `Ownership`.. Maybe we ought to make it public, but the internals are still private.. which is fine?
