### 2024-05-27 904d945
Update faer requirement from 0.18.0 to 0.19.0
---
updated-dependencies:
- dependency-name: faer
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
