### 2024-05-29 d0ba478
CHANGELOG: made a breaking change section, added description of PR
