+++ CHANGELOG.md
@@ -21,13 +21,17 @@
   place, and then return an untyped owned pointer (Robj). Instead, now we return `&mut Self`.
 - [_Potentially breaking_]: `RArray::from_parts` no longer requires a pointer to the underlying data
   vector [[#657]](https://github.com/extendr/extendr/pull/657)
-- [**Breaking change**]: In `AltRep` the `unserialize_ex`, `set_parent`, `set_envflags` are
-now hidden behind the feature flag `non-api`. Also, `Promise::from_parts` is marked as non-API.
-- [**Breaking change**]: Floating point numbers with decimal part can no longer be converted to integer types via
-  rounding [[#757]](https://github.com/extendr/extendr/pull/757)
-- [**Breaking change**]: You can no longer create an `Robj` from a reference `&T`, where `T` is an `extendr`-impl. [[#759]](https://github.com/extendr/extendr/pull/759)
-- [**Breaking change**]: You can no longer use `from_robj`, as the trait `FromRobj` as been removed. Instead, use `try_from`.
 - `#[extendr(use_try_from = true)` is now the default setting, therefore the option has `use_try_from` has been removed [[#759]](https://github.com/extendr/extendr/pull/759)
+
+#### Breaking changes
+
+- In `AltRep` the `unserialize_ex`, `set_parent`, `set_envflags` are
+now hidden behind the feature flag `non-api`. Also, `Promise::from_parts` is marked as non-API.
+- Floating point numbers with decimal part can no longer be converted to integer types via
+  rounding [[#757]](https://github.com/extendr/extendr/pull/757)
+- You can no longer create an `Robj` from a reference `&T`, where `T` is an `extendr`-impl. [[#759]](https://github.com/extendr/extendr/pull/759)
+- You can no longer use `from_robj`, as the trait `FromRobj` as been removed. Instead, use `try_from`.
+- It is no longer possible to access R integer vector as a `&[u32]
 
 ### Fixed
 
