+++ CHANGELOG.md
@@ -36,7 +36,7 @@ now hidden behind the feature flag `non-api`. Also, `Promise::from_parts` is mar
 ### Fixed
 
 - returning `&Self` or `&mut Self` from a method in an `#[extendr]`-impl would
-result in unintended cloning  [[[#614](https://github.com/extendr/extendr/issues/614)]]
+result in unintended cloning  [[#614]](https://github.com/extendr/extendr/issues/614)
 - `TryFrom<&Robj>` and `FromRobj` for integer scalars now correctly handles conversions
   from `f64` [[#757]](https://github.com/extendr/extendr/pull/757)
 
