### 2024-05-30 6ea456a
bug: removed `&[u32]` interpretation of `Robj`  (#767)
* fix: remove wrong `u32` casting

* fix: removed invalid `u32` conversion

* CHANGELOG: made a breaking change section,
added description of PR

* fix changelog

* CHANGELOG: quick fix
