+++ extendr-api/Cargo.toml
@@ -24,7 +24,7 @@ faer = { version = "0.18.0", optional = true }
 
 [dev-dependencies]
 extendr-engine = { path = "../extendr-engine" }
-rstest = "0.19.0"
+rstest = "0.20.0"
 
 [features]
 
