### 2024-05-31 f2bfa49
added two new ideas
