### 2024-05-31 40e699e
extendr-api: we should not have panics. conversions, especially those we provide, must not panic.
