### 2024-06-01 e7e6787
wp: slowly working on the fix.. look up `easy_ffi` crate
