### 2024-06-06 d95497a
Cargo.toml: Upgraded faer
