+++ tests/extendrtests/src/rust/Cargo.toml
@@ -17,7 +17,7 @@ crate-type = ["staticlib"]
 
 [dependencies]
 extendr-api = { version = "*", features = ["graphics", "ndarray", "faer", "either"] }
-faer = "0.18.0"
+faer = "0.19"
 
 # TODO: I couldn't find any nice way to add the condition based on the R version
 # except for using libR-sys just for "DEP_R_*" envvars.
