### 2024-06-07 9daf4b2
It is not allowed to panic in the FFI boundary. We can instead panic, and catch the panic. This will drop everything, and proper clean-up is ensured.
