### 2024-06-07 5c85811
all the imported arguments are moved to the closure, ensuring that when the closure terminates, they  are all _dropped_.
