### 2024-06-07 20a2cc4
update macro-snapshot
