+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -111,12 +111,14 @@
                   std::result::Result<Robj, extendr_api::Error>,
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
-                  let _robj_robj = extendr_api::robj::Robj::from_sexp(robj);
                   std::panic::catch_unwind(
-                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                      std::panic::AssertUnwindSafe(move || -> std::result::Result<
                           Robj,
                           extendr_api::Error,
-                      > { Ok(extendr_api::Robj::from(new_usize(_robj_robj.try_into()?))) }),
+                      > {
+                          let _robj_robj = extendr_api::robj::Robj::from_sexp(robj);
+                          Ok(extendr_api::Robj::from(new_usize(_robj_robj.try_into()?)))
+                      }),
                   )
               };
               match wrap_result_state {
@@ -235,7 +237,7 @@
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
                   std::panic::catch_unwind(
-                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                      std::panic::AssertUnwindSafe(move || -> std::result::Result<
                           Robj,
                           extendr_api::Error,
                       > { Ok(extendr_api::Robj::from(tst_altstring())) }),
@@ -365,7 +367,7 @@
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
                   std::panic::catch_unwind(
-                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                      std::panic::AssertUnwindSafe(move || -> std::result::Result<
                           Robj,
                           extendr_api::Error,
                       > { Ok(extendr_api::Robj::from(tst_altinteger())) }),
@@ -527,12 +529,14 @@
                   std::result::Result<Robj, extendr_api::Error>,
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
-                  let _x_robj = extendr_api::robj::Robj::from_sexp(x);
                   std::panic::catch_unwind(
-                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                      std::panic::AssertUnwindSafe(move || -> std::result::Result<
                           Robj,
                           extendr_api::Error,
-                      > { Ok(extendr_api::Robj::from(dbls_named(_x_robj.try_into()?))) }),
+                      > {
+                          let _x_robj = extendr_api::robj::Robj::from_sexp(x);
+                          Ok(extendr_api::Robj::from(dbls_named(_x_robj.try_into()?)))
+                      }),
                   )
               };
               match wrap_result_state {
@@ -615,12 +619,14 @@
                   std::result::Result<Robj, extendr_api::Error>,
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
-                  let _x_robj = extendr_api::robj::Robj::from_sexp(x);
                   std::panic::catch_unwind(
-                      std::panic::AssertUnwindSafe(|| -> std::result::Result<
+                      std::panic::AssertUnwindSafe(move || -> std::result::Result<
                           Robj,
                           extendr_api::Error,
-                      > { Ok(extendr_api::Robj::from(strings_named(_x_robj.try_into()?))) }),
+                      > {
+                          let _x_robj = extendr_api::robj::Robj::from_sexp(x);
+                          Ok(extendr_api::Robj::from(strings_named(_x_robj.try_into()?)))
+                      }),
                   )
               };
               match wrap_result_state {
@@ -702,13 +708,13 @@
                   std::result::Result<Robj, extendr_api::Error>,
                   Box<dyn std::any::Any + Send>,
               > = unsafe {
-                  let _x_robj = extendr_api::robj::Robj::from_sexp(x);
[…]
