### 2024-06-09 9e2d14c
fix: relying on specific behavior for `List::iter`
