### 2024-06-10 9d6a90b
restructured `StrIter` completely
