+++ extendr-api/src/iter.rs
@@ -37,9 +37,10 @@ impl Default for StrIter {
 impl StrIter {
     /// Make an empty str iterator.
     pub fn new(len: usize) -> Self {
+        let vector = if len == 0 { nil_value() } else { na_string() };
         unsafe {
             Self {
-                vector: ().into(),
+                vector,
                 i: 0,
                 len,
                 levels: R_NilValue,
@@ -78,9 +79,11 @@ impl Iterator for StrIter {
             if i >= self.len {
                 None
             } else if TYPEOF(vector) == SEXPTYPE::NILSXP {
-                Some(<&str>::na())
+                None
             } else if TYPEOF(vector) == SEXPTYPE::STRSXP {
                 str_from_strsxp(vector, i)
+            } else if vector == R_NaString {
+                Some(<&str>::na())
             } else if TYPEOF(vector) == SEXPTYPE::CHARSXP {
                 rstr::charsxp_to_str(vector)
             } else if Rf_isFactor(vector).into() {
@@ -234,8 +237,13 @@ mod tests {
     fn test_new_constructor() {
         with_r(|| {
             let str_iter = StrIter::new(10);
-
             assert_eq!(str_iter.collect::<Vec<_>>().len(), 10);
+            let str_iter = StrIter::new(0);
+            let str_iter_collect = str_iter.collect::<Vec<_>>();
+            assert_eq!(str_iter_collect.len(), 0);
+            assert!(str_iter_collect.is_empty());
+            let mut str_iter = StrIter::new(0);
+            assert!(str_iter.next().is_none());
         });
     }
 }
