+++ CHANGELOG.md
@@ -16,15 +16,15 @@
 
 ### Changed
 
-- [**Breaking change**]: the `Attributes` trait now returns a mutable reference
-  to `Self`. [[#745]](https://github.com/extendr/extendr/pull/745). Previously `.set_attrib()` would modify an object in
-  place, and then return an untyped owned pointer (Robj). Instead, now we return `&mut Self`.
 - [_Potentially breaking_]: `RArray::from_parts` no longer requires a pointer to the underlying data
   vector [[#657]](https://github.com/extendr/extendr/pull/657)
-- `#[extendr(use_try_from = true)` is now the default setting, therefore the option has `use_try_from` has been removed [[#759]](https://github.com/extendr/extendr/pull/759)
+- `#[extendr(use_try_from = true)` is now the default setting, therefore the option `use_try_from` has been removed [[#759]](https://github.com/extendr/extendr/pull/759)
 
 #### Breaking changes
 
+- `Attributes` trait now returns a mutable reference
+  to `Self`. [[#745]](https://github.com/extendr/extendr/pull/745). Previously `.set_attrib()` would modify an object in
+  place, and then return an untyped owned pointer (Robj). Instead, now we return `&mut Self`.
 - In `AltRep` the `unserialize_ex`, `set_parent`, `set_envflags` are
 now hidden behind the feature flag `non-api`. Also, `Promise::from_parts` is marked as non-API.
 - Floating point numbers with decimal part can no longer be converted to integer types via
