### 2024-06-10 967b1a2
doc fix: comments from unimplemented features snuck in
