+++ extendr-macros/src/lib.rs
@@ -76,11 +76,9 @@ use syn::{parse_macro_input, Item};
 ///
 /// - `fn` for wrapped rust-functions, see [`extendr-fn`]
 /// - `impl`-blocks, see [`extendr-impl`]
-/// - Rust `enum`s as R factors, see [`extendr-enum`]
 ///
 /// [`extendr-fn`]: extendr-macros/extendr_function/fn.extendr_function.html
 /// [`extendr-impl`]: extendr_impl/fn.extendr_impl.html
-/// [`extendr-enum`]: extendr_enum/fn.extendr_enum.html
 ///
 /// There is also [extendr_module!], which is used for defining what rust
 /// wrapped items should be visible to the surrounding R-package.
