### 2024-06-10 e7ab15d
doc: editted the hardware-less performance numbers
