### 2024-06-10 80c5671
doc: updated features list
