### 2024-06-10 5690f79
doc fix: we are not discouraging the default state of extendr
