+++ extendr-api/README.md
@@ -1,6 +1,5 @@
 # extendr-api
 
-
 A safe and user friendly R extension interface.
 
 * Build rust extensions to R.
@@ -13,6 +12,7 @@ See [Robj] for much of the content of this crate.
 [Robj] provides a safe wrapper for the R object type.
 
 Use attributes and macros to export to R.
+
 ```rust
 use extendr_api::prelude::*;
 // Export a function or impl to R.
@@ -38,6 +38,7 @@ result <- fred(1)
 [Robj] is a wrapper for R objects.
 The r!() and R!() macros let you build R objects
 using Rust and R syntax respectively.
+
 ```rust
 use extendr_api::prelude::*;
 test! {
@@ -100,6 +101,7 @@ test! {
 To index a vector, first convert it to a slice and then
 remember to use 0-based indexing. In Rust, going out of bounds
 will cause and error (a panic) unlike C++ which may crash.
+
 ```rust
 use extendr_api::prelude::*;
 test! {
@@ -114,6 +116,7 @@ test! {
 ```
 
 Much slower, but more general are these methods:
+
 ```rust
 use extendr_api::prelude::*;
 test! {
@@ -135,6 +138,7 @@ The [R!] macro lets you embed R code in Rust
 and takes Rust expressions in {{ }} pairs.
 
 The [Rraw!] macro will not expand the {{ }} pairs.
+
 ```rust
 use extendr_api::prelude::*;
 test! {
@@ -165,6 +169,7 @@ test! {
 
 The [r!] macro converts a rust object to an R object
 and takes parameters.
+
 ```rust
 use extendr_api::prelude::*;
 test! {
@@ -175,6 +180,7 @@ test! {
 ```
 
 You can call R functions and primitives using the [call!] macro.
+
 ```rust
 use extendr_api::prelude::*;
 test! {
