### 2024-06-10 5ce7a47
added: checked conversion to `[&u32]` and `Vec<u32>` from integer vector
