### 2024-06-10 a39ca4f
added mutable slices to `TryFrom`
