+++ extendr-api/src/robj/try_from_robj.rs
@@ -538,50 +538,19 @@ macro_rules! impl_try_from_robj {
     };
 }
 
+#[rustfmt::skip]
 impl_try_from_robj!(
-    u8,
-    u16,
-    u32,
-    u64,
-    usize,
-    i8,
-    i16,
-    i32,
-    i64,
-    isize,
+    u8, u16, u32, u64, usize,
+    i8, i16, i32, i64, isize,
     bool,
-    Rint,
-    Rfloat,
-    Rbool,
-    Rcplx,
-    f32,
-    f64,
+    Rint, Rfloat, Rbool, Rcplx,
+    f32, f64,
     Vec::<String>,
-    HashMap::<String, Robj>,
-    HashMap::<&str, Robj>,
-    Vec::<Rint>,
-    Vec::<Rfloat>,
-    Vec::<Rbool>,
-    Vec::<Rcplx>,
-    Vec::<u8>,
-    Vec::<i32>,
-    Vec::<f64>,
-    &[Rint],
-    &[Rfloat],
-    &[Rbool],
-    &[Rcplx],
-    &[u8],
-    &[i32],
-    &[f64],
-    &mut [Rint],
-    &mut [Rfloat],
-    &mut [Rbool],
-    &mut [Rcplx],
-    &mut [u8],
-    &mut [i32],
-    &mut [f64],
-    &str,
-    String,
+    HashMap::<String, Robj>, HashMap::<&str, Robj>,
+    Vec::<Rint>, Vec::<Rfloat>, Vec::<Rbool>, Vec::<Rcplx>, Vec::<u8>, Vec::<i32>, Vec::<f64>,
+    &[Rint], &[Rfloat], &[Rbool], &[Rcplx], &[u8], &[i32], &[f64],
+    &mut [Rint], &mut [Rfloat], &mut [Rbool], &mut [Rcplx], &mut [u8], &mut [i32], &mut [f64],
+    &str, String,
 );
 
 // NOTE: this is included for compatibility with previously defined `FromRobj`
