### 2024-06-10 9b2c5b0
test: added test for `List` and `StrIter` use of it
