### 2024-06-10 d989191
Update Cargo.toml
Preparing for the new version
