+++ Cargo.toml
@@ -4,7 +4,7 @@ resolver = "2"
 members = ["extendr-api", "extendr-engine", "extendr-macros", "xtask"]
 
 [workspace.package]
-version = "0.6.0"
+version = "0.7.0"
 authors = [
     "andy-thomason <andy@andythomason.com>",
     "Thomas Down",
