### 2024-06-10 #6394 @CGMossa
Dear editor @jbytecode,

I've scheduled tomorrow to write an update to all the points raised by @dccsillag. A quick update from me, is that we've merged all the _code_ changes that are necessary to address the issues mentioned. Today, I've called for testing on our discord server to all the users, because I want to release a new version of extendr, now that these extensive changes have landed.

So everything is progressing nicely.
