+++ Cargo.toml
@@ -21,7 +21,7 @@ rust-version = "1.70"
 
 [workspace.dependencies]
 # When updating extendr's version, this version also needs to be updated
-extendr-macros = { path = "./extendr-macros", version = "0.6.0" }
+extendr-macros = { path = "./extendr-macros", version = "0.7.0" }
 
 # When uncommenting this, do not forget to uncomment the same line in
 # ./tests/extendrtests/src/rust/Cargo.toml, and "Run R integration tests using
