### 2024-06-10 864cd87
wp: remove `libR-sys` from `tests/extendrtests`
