### 2024-06-11 #6394 @alpaylan
Dear @jbytecode, @CGMossa,

Thank you for the patience. I checked all the boxes except the `State of the field` in the submission checkbox, I think the package is well documented, as well as easy to use, solves an important problem that I see very important. The accompanying paper is also clear, and addresses the problems `extendr` aims to solve. Additionally, I believe the comparisons of the paper with the rest of the literature needs further details, hence I did not check `State of the field` box yet.

**Expectations regarding comparison with the rest of the literature:**

> The savvy interface represents a distilled byproduct of ‘extendr’. However, these implementations
differ from ‘extendr’ in that ‘extendr’ aims at providing an opinionated API, with a focus on an ergonomic API design inspired by features from Rcpp and cpp11.

I believe it is important to provide relevant context to demonstrate such comparative claims if possible. Such context can be given in the form of side by side examples of using both libraries and demonstrating what makes extendr more ergonomic compared to an alternative library, or detailing how the choices made by `extendr` provides ergonomy and in what ways.

Additionally, if possible, having some quantitative measurements for comparing performances may also strengthen the paper, as was done in Rcpp paper that inspired this work.
