### 2024-06-11 544b38b
bug: `libR_sys` must not be used in macros! (#792)
