+++ search.desc/extendr_api/extendr_api-desc-1-.js
@@ -1,1 +1,1 @@
-searchState.loadedDescShard("extendr_api", 1, "Calls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nCalls <code>U::from(self)</code>.\nReturns a view over the conjugate transpose of <code>self</code>.\nReturns a view over the conjugate transpose of <code>self</code>.\nReturns a view over the conjugate of <code>self</code>.\nReturns a view over the conjugate of <code>self</code>.\nDecomposes the matrix into the symbolic part and the …\nDecomposes the matrix into the symbolic part and the …\nDecomposes the matrix into the symbolic part and the …\nDecomposes the matrix into the symbolic part and the …\nDecomposes the matrix into the symbolic part and the …\nReturns the components of the matrix in the order:\nReturns the components of the matrix in the order:\nDecomposes the matrix into the symbolic part and the …\nReturns a view over the transpose of <code>self</code> in row-major …\nReturns a view over the transpose of <code>self</code> in column-major …\nConsumes the matrix, and returns its transpose in …\nConsumes the matrix, and returns its transpose in …\nSparse linear algebra module. Contains low level routines …\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nReturns the number of columns of the matrix.\nCreates a new sparse matrix view.\nCreates a new sparse matrix view.\nCreates a new sparse matrix view.\nCreates a new sparse matrix view.\nCreates a new sparse matrix view.\nCreates a new sparse matrix view.\nCreates a new symbolic matrix view after asserting its …\nCreates a new symbolic matrix view after asserting its …\nCreates a new symbolic matrix view after asserting its …\nCreates a new symbolic matrix view after asserting its …\nCreate a new matrix from a previously created symbolic …\nCreate a new matrix from a previously created symbolic …\nCreates a new symbolic matrix view without asserting its …\nCreates a new symbolic matrix view without asserting its …\nCreates a new symbolic matrix view without asserting its …\nCreates a new symbolic matrix view without asserting its …\nCreates a new symbolic matrix view from data containing …\nCreates a new symbolic matrix view from data containing …\nCreates a new symbolic matrix view from data containing …\nCreates a new symbolic matrix view from data containing …\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per column of the matrix.\nReturns the count of non-zeros per row of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nReturns the number of rows of the matrix.\nSparse ma
[…]
