### 2024-06-11 2fb1038
Update tests/extendrtests/src/rust/build.rs
Co-authored-by: Ilia Kosenkov <ilia.kosenkov@outlook.com>
