### 2024-06-11 #6394 @CGMossa
Hello everyone!

(CC editor @jbytecode)

First, I will address the remaining points from @dccsillag, and then I'd like to address the comments from @alpaylan.

### Documentation

Yes! We need to have proper links and direction in the main `README.md` at the extendr repository. I've filed that here
https://github.com/extendr/extendr/pull/795.

The overall point about having consolidated documentation on how to use extendr, that covers the R part, the Rust part, the extendR in Rust part is a very valid point. And we have taken first steps towards this by creating the _User Guide_ over at https://extendr.github.io/user-guide/. The user-guide effort is on-going, but it already fulfils the purpose of unifying documentation. Most notably there is the step-by-step tutorial to take a rust crate, and make an R-package for it, see [tutorial on porting a rust crate](https://extendr.github.io/user-guide/heckin-case-converter.html).

The reason for the User Guide, is because extendR has multiple projects, with many maintainers. `libR-sys` is maintained by me, and people outside the extendR project, because it is useful to others. Then there is `rextendr` which is on-going, and its development process is a bit separate from extendr-crates.

Therefore, we have established a new repository, with no pre-established owners, that then contribute to one unified documentation, instead of having to ping pong with three separate entities all the time.

### Error management

We've been dealing with these concerns for years. There are certain limitations imposed by _Rust_ and others by R (and CRAN compliance). For instance, we are tied to Rust 1.67, due to CRAN machines using out of support distributions (Fedora 39), etc. However, over the years, we have merged orthogonal ideas to error management in extendR to experiment and figure out what works, and what doesn't.

Thus, the documentation that @dccsillag pointed to was correct, up until now, where some of these issues have been rectified. We figured that this critical change merged in https://github.com/extendr/extendr/pull/781/files would resolve the leaking issue, even for the default behaviour. The claims will be removed in https://github.com/extendr/extendr/pull/784.

To me, this is as far as the error management effort can go for now. New versions of Rust must get installed on CRAN servers, before I can do more ergonomic changes to error handling. For now, we no longer _leak by default_, so that's good.

### Performance

I've proposed a PR https://github.com/extendr/extendr/pull/784 to remove the performance claims present in the documentation. This also removes the snippet mentioned by @dccsillag

> From the docs: "The minimal overhead of calling an extendr function is in the ballpark of 2-4us. Returning a condition or list increases the overhead to 4-8us. Checking & handling the result on R side will likely increase overall overhead to 8-16us, depending on how efficiently the result is handled." But the numbers here certainly depends on the specific hardware, right? Maybe it would be better to clarify the kind of hardware being considered here, or to speak in terms of relations (e.g., 40% slower) and not absolute numbers. (Yes, in the case of relations it's still a bit flawed, but much better than stating raw numbers.)

Overall, I don't think we make performance claims for the justification of extendR. And comparing the performance of extendR to other frameworks that support languages would pose a question as to what we are comparing to. Also, extendR attempts to provide the same functionality as `Rcpp`, but the internal architecture of extendR is more similar to `cpp11`, then `Rcpp`. For instance, `Robj` in Rust implement `Clone`, which is not the case in either `Rcpp` nor `cpp11`.

This is not to say, that we aren't conscientious about performance. We recently merged https://github.com/extendr/extendr/pull/761, where the idea was, that C-strings have `NUL` in them to mark the end of the string, and before we were iterating every string to find its length by detecting `NUL`, and now we retrieve this information from the R runtime instead.

But I believe including performance benchmarks in the paper is out of scope of the paper, and mostly also outside the goals of extendR.

### Comparison to `savvy`

Thank you for your comments @alpaylan. We were hesitant as how to directly mention the difference between `savvy` and `extendr`.

I would refer to their own page for comparing extendR [here](https://yutannihilation.github.io/savvy/guide/extendr.html) and to quote:

> [extendr](https://extendr.github.io/) is great and ready to use, but it's not perfect in some points (e.g., [error handling](https://github.com/extendr/extendr/issues/278)) and it's kind of stuck; extendr is too feature-rich and complex that no one can introduce a big breaking change easily. So, I needed to create a new, simple framework to experiment with. The main goal of savvy is to provide a simpler option other than extendr, not to be a complete alternative to extendr.

A direct statement about differences could be
> `savvy` does not have facilities to use Rust from an R session. Nor does it provide a knitr-engine for Rust code. It also does not provide automatically generate R wrappers to Rust functions, see <https://yutannihilation.github.io/savvy/guide/key_ideas.html#no-implicit-conversions>. Also, all savvy-functions must return a `Result`. Futhermore, `savvy` is not trying to be compliant with CRAN requirements.

I personally think that this is _too direct_ of a comparison to fit in a paper, and it pivots the text a bit. `savvy` is not opinionated, but `extendr` is, seemed like a reasonable short sentence to state the difference.

Let me know what you'd advise on this matter.
