### 2024-06-12 999f90b
wp: added other methods TODO: these need to be available for Array as well
