### 2024-06-12 5ee641b
wp: added type annotations
