### 2024-06-12 abd6a29
Remove `libR-sys` from extendrtests permanently (#793)
* Removes `libR-sys` from `tests/extendrtests`

Co-authored-by: Ilia Kosenkov <ilia.kosenkov@outlook.com>

---------

Co-authored-by: Ilia Kosenkov <ilia.kosenkov@outlook.com>
