+++ tests/extendrtests/src/rust/src/lib.rs
@@ -5,6 +5,7 @@ mod attributes;
 mod dataframe;
 mod externalptr;
 mod graphic_device;
+mod matrix;
 mod memory_leaks;
 mod optional_either;
 mod optional_faer;
@@ -352,4 +353,5 @@ extendr_module! {
     use raw_identifiers;
     use submodule;
     use externalptr;
+    use matrix;
 }
