+++ tests/extendrtests/DESCRIPTION
@@ -14,7 +14,7 @@ License: MIT + file LICENSE
 SystemRequirements: Rust
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 Imports:
     rextendr,
 Suggests:
