### 2024-06-16 1350b4c
Update main `README.md` (#795)
* wp: added comments from reviewer..

* wp: markdownlint

* Updated links on top
