### 2024-06-16 cfe5469
Updating documentation everywhere (#784)
* CHANGELOG: minor edits

* doc fix: comments from unimplemented features snuck in

* doc fix: this is no longer true per #722

* doc fix: fixed in #781

* doc: editted the hardware-less performance numbers

* doc fix: outdated roadmap, we should propose a new one, see #783

* doc: updated features list

* doc fix: we are not discouraging the default state of extendr

* doc: markdownlint

* doc: tiny

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
