+++ tests/extendrtests/src/rust/src/lib.rs
@@ -11,6 +11,7 @@ mod optional_faer;
 mod optional_ndarray;
 mod raw_identifiers;
 mod submodule;
+mod typedsliceargs;
 
 // Return string `"Hello world!"` to R.
 #[extendr]
@@ -352,4 +353,5 @@ extendr_module! {
     use raw_identifiers;
     use submodule;
     use externalptr;
+    use typedsliceargs;
 }
