+++ tests/extendrtests/R/extendr-wrappers.R
@@ -145,6 +145,12 @@ false <- function(type) .Call(wrap__false, type)
 #' @export
 hello_submodule <- function() .Call(wrap__hello_submodule)
 
+fetch_colnames <- function(x) .Call(wrap__fetch_colnames, x)
+
+fetch_rownames <- function(x) .Call(wrap__fetch_rownames, x)
+
+change_dimnames <- function(x) .Call(wrap__change_dimnames, x)
+
 #' Class for testing (exported)
 #' @examples
 #' x <- MyClass$new()
