### 2024-06-16 29f02e1
Added mutable slices from `AsTypedSlice` to `TryFrom` (#790)
* added mutable slices to `TryFrom`

* REALLY? `cargo fmt`

* don't mangle this part..

* add integration tests

* updated snapshots

* [skip_ci] updte changelog

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
