+++ extendr-api/src/robj/try_from_robj.rs
@@ -324,7 +324,7 @@ impl TryFrom<&Robj> for &[Rbool] {
 impl TryFrom<&Robj> for &[Rcplx] {
     type Error = Error;
 
-    /// Convert a complex object into a slice of Rbool
+    /// Convert a complex object into a slice of Rcplx
     /// Use `value.is_na()` to detect NA values.
     fn try_from(robj: &Robj) -> Result<Self> {
         robj.as_typed_slice()
@@ -349,6 +349,82 @@ impl TryFrom<&Robj> for &[f64] {
     /// Use `value.is_na()` to detect NA values.
     fn try_from(robj: &Robj) -> Result<Self> {
         robj.as_typed_slice()
+            .ok_or_else(|| Error::ExpectedReal(robj.clone()))
+    }
+}
+
+impl TryFrom<&mut Robj> for &mut [i32] {
+    type Error = Error;
+
+    /// Convert an INTSXP object into a mutable slice of i32 (integer).
+    /// Use `value.is_na()` to detect NA values.
+    fn try_from(robj: &mut Robj) -> Result<Self> {
+        robj.as_typed_slice_mut()
+            .ok_or_else(|| Error::ExpectedInteger(robj.clone()))
+    }
+}
+
+impl TryFrom<&mut Robj> for &mut [Rint] {
+    type Error = Error;
+
+    /// Convert an integer object into a mutable slice of Rint (tri-state booleans).
+    /// Use `value.is_na()` to detect NA values.
+    fn try_from(robj: &mut Robj) -> Result<Self> {
+        robj.as_typed_slice_mut()
+            .ok_or_else(|| Error::ExpectedInteger(robj.clone()))
+    }
+}
+
+impl TryFrom<&mut Robj> for &mut [Rfloat] {
+    type Error = Error;
+
+    /// Convert a doubles object into a mutable slice of Rfloat (tri-state booleans).
+    /// Use `value.is_na()` to detect NA values.
+    fn try_from(robj: &mut Robj) -> Result<Self> {
+        robj.as_typed_slice_mut()
+            .ok_or_else(|| Error::ExpectedReal(robj.clone()))
+    }
+}
+
+impl TryFrom<&mut Robj> for &mut [Rbool] {
+    type Error = Error;
+
+    /// Convert a logical object into a mutable slice of Rbool (tri-state booleans).
+    /// Use `value.is_na()` to detect NA values.
+    fn try_from(robj: &mut Robj) -> Result<Self> {
+        robj.as_typed_slice_mut()
+            .ok_or_else(|| Error::ExpectedLogical(robj.clone()))
+    }
+}
+
+impl TryFrom<&mut Robj> for &mut [Rcplx] {
+    type Error = Error;
+
+    /// Convert a complex object into a mutable slice of Rcplx
+    /// Use `value.is_na()` to detect NA values.
+    fn try_from(robj: &mut Robj) -> Result<Self> {
+        robj.as_typed_slice_mut()
+            .ok_or_else(|| Error::ExpectedComplex(robj.clone()))
+    }
+}
+
+impl TryFrom<&mut Robj> for &mut [u8] {
+    type Error = Error;
+
+    /// Convert a RAWSXP object into a mutable slice of bytes.
+    fn try_from(robj: &mut Robj) -> Result<Self> {
+        robj.as_typed_slice_mut()
+            .ok_or_else(|| Error::ExpectedRaw(robj.clone()))
+    }
+}
+
+impl TryFrom<&mut Robj> for &mut [f64] {
+    type Error = Error;
+
+    /// Convert a REALSXP object into a mutable slice of f64 (double precision floating point).
+    /// Use `value.is_na()` to detect NA values.
+    fn try_from(robj: &mut Robj) -> Result<Self> {
[…]
