+++ tests/extendrtests/tests/testthat/_snaps/macro-snapshot.md
@@ -5193,6 +5193,375 @@
               unsafe { extendr_api::register_call_methods(info, get_submodule_metadata()) };
           }
       }
+      mod typedsliceargs {
+          use extendr_api::prelude::*;
+          fn middle_zero(integers: &mut [Rint]) {
+              let middle = integers.len() / 2;
+              integers[middle] = 42_i32.into();
+          }
+          #[no_mangle]
+          #[allow(non_snake_case, clippy::not_unsafe_ptr_arg_deref)]
+          pub extern "C" fn wrap__middle_zero(
+              integers: extendr_api::SEXP,
+          ) -> extendr_api::SEXP {
+              use extendr_api::robj::*;
+              let wrap_result_state: std::result::Result<
+                  std::result::Result<Robj, extendr_api::Error>,
+                  Box<dyn std::any::Any + Send>,
+              > = unsafe {
+                  std::panic::catch_unwind(
+                      std::panic::AssertUnwindSafe(move || -> std::result::Result<
+                          Robj,
+                          extendr_api::Error,
+                      > {
+                          let _integers_robj = extendr_api::robj::Robj::from_sexp(integers);
+                          Ok(extendr_api::Robj::from(middle_zero(_integers_robj.try_into()?)))
+                      }),
+                  )
+              };
+              match wrap_result_state {
+                  Ok(Ok(zz)) => {
+                      return unsafe { zz.get() };
+                  }
+                  Ok(Err(conversion_err)) => {
+                      let err_string = conversion_err.to_string();
+                      drop(conversion_err);
+                      extendr_api::throw_r_error(&err_string);
+                  }
+                  Err(unwind_err) => {
+                      drop(unwind_err);
+                      let err_string = {
+                          let res = ::alloc::fmt::format(
+                              format_args!("User function panicked: {0}", "middle_zero"),
+                          );
+                          res
+                      };
+                      extendr_api::handle_panic(
+                          err_string.as_str(),
+                          || {
+                              #[cold]
+                              #[track_caller]
+                              #[inline(never)]
+                              const fn panic_cold_explicit() -> ! {
+                                  ::core::panicking::panic_explicit()
+                              }
+                              panic_cold_explicit();
+                          },
+                      );
+                  }
+              }
+              {
+                  ::core::panicking::panic_fmt(
+                      format_args!(
+                          "internal error: entered unreachable code: {0}",
+                          format_args!("internal extendr error, this should never happen."),
+                      ),
+                  );
+              }
+          }
+          #[allow(non_snake_case)]
+          fn meta__middle_zero(metadata: &mut Vec<extendr_api::metadata::Func>) {
+              let mut args = <[_]>::into_vec(
+                  #[rustc_box]
+                  ::alloc::boxed::Box::new([
+                      extendr_api::metadata::Arg {
+                          name: "integers",
+                          arg_type: "_5bRint_5d",
+                          default: None,
+                      },
+                  ]),
+              );
+              metadata
+                  .push(extendr_api::metadata::Func {
+                      doc: "",
+                      rust_name: "middle_zero",
+                      r_name: "middle_zero",
+                      mod_name: "middle_zero",
+                      args: args,
+                      return_type: "()",
+                      func_ptr: wrap__middle_zero as *const u8,
+                      hidden: false,
+                  })
+          }
+          fn logicals_sum(logicals: &[Rbool]) -> i32 {
+              logicals
+                  .iter()
+                  .fold(
+                      0,
+                      |acc, next| {
+                          let bool = next.is_true();
+                          acc + bool as i32
+                      },
+                  )
+          }
+          #[no_mangle]
[…]
