+++ type.impl/std/primitive.isize.js
@@ -1,3 +1,3 @@
 (function() {var type_impls = {
-"extendr_api":[["<details class=\"toggle implementors-toggle\" open><summary><section id=\"impl-SignedIndex-for-isize\" class=\"impl\"><a href=\"#impl-SignedIndex-for-isize\" class=\"anchor\">§</a><h3 class=\"code-header\">impl <a class=\"trait\" href=\"extendr_api/prelude/trait.SignedIndex.html\" title=\"trait extendr_api::prelude::SignedIndex\">SignedIndex</a> for <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.isize.html\">isize</a></h3></section></summary><div class=\"impl-items\"><details class=\"toggle\" open><summary><section id=\"associatedconstant.MAX\" class=\"associatedconstant trait-impl\"><a href=\"#associatedconstant.MAX\" class=\"anchor\">§</a><h4 class=\"code-header\">const <a href=\"extendr_api/prelude/trait.SignedIndex.html#associatedconstant.MAX\" class=\"constant\">MAX</a>: <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.isize.html\">isize</a> = 9_223_372_036_854_775_807isize</h4></section></summary><div class='docblock'>Maximum representable value.</div></details><details class=\"toggle method-toggle\" open><summary><section id=\"method.truncate\" class=\"method trait-impl\"><a href=\"#method.truncate\" class=\"anchor\">§</a><h4 class=\"code-header\">fn <a href=\"extendr_api/prelude/trait.SignedIndex.html#tymethod.truncate\" class=\"fn\">truncate</a>(value: <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a>) -&gt; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.isize.html\">isize</a></h4></section></summary><div class='docblock'>Truncate <code>value</code> to type <a href=\"extendr_api/prelude/trait.SignedIndex.html\" title=\"trait extendr_api::prelude::SignedIndex\"><code>Self</code></a>.</div></details><details class=\"toggle method-toggle\" open><summary><section id=\"method.zx\" class=\"method trait-impl\"><a href=\"#method.zx\" class=\"anchor\">§</a><h4 class=\"code-header\">fn <a href=\"extendr_api/prelude/trait.SignedIndex.html#tymethod.zx\" class=\"fn\">zx</a>(self) -&gt; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a></h4></section></summary><div class='docblock'>Zero extend <code>self</code>.</div></details><details class=\"toggle method-toggle\" open><summary><section id=\"method.sx\" class=\"method trait-impl\"><a href=\"#method.sx\" class=\"anchor\">§</a><h4 class=\"code-header\">fn <a href=\"extendr_api/prelude/trait.SignedIndex.html#tymethod.sx\" class=\"fn\">sx</a>(self) -&gt; <a class=\"primitive\" href=\"https://doc.rust-lang.org/nightly/std/primitive.usize.html\">usize</a></h4></section></summary><div class='docblock'>Sign extend <code>self</code>.</div></details><details class=\"toggle method-toggle\" open><summary><section id=\"method.sum_nonnegative\" class=\"method trait-impl\"><a href=\"#method.sum_nonnegative\" class=\"anchor\">§</a><h4 class=\"code-he
[…]
