### 2024-06-16 d2c4014
Add `names<-`,  `dim<-` and `dimnames<-` to `RArray` and friends (#796)
* wp: colnames and rownames in matrix.

* wp: added other methods
TODO: these need to be available for Array as well

* wp: added type annotations

* wp: moved the set elements to the correct impl-block.

* wp: list of used dimnames api

* wp: took the tests from #800. I couldn't find a better way to port them over.

* disabled `dimnames` retrieval..

* wp: yes, this works..

* alright! added the last bit

---------

Co-authored-by: Josiah Parry <josiah.parry@gmail.com>
