### 2024-06-16 b02b4b9
Less ambigious return type for colnames/rownames
