+++ extendr-api/src/wrapper/strings.rs
@@ -148,3 +148,12 @@ impl std::fmt::Debug for Strings {
         }
     }
 }
+
+impl From<Option<Strings>> for Robj {
+    fn from(value: Option<Strings>) -> Self {
+        match value {
+            Some(value_strings) => value_strings.into(),
+            None => nil_value(),
+        }
+    }
+}
