### 2024-06-22 #6394 @CGMossa
Hello 👋 @jbytecode

All the PRs that I've mentioned in my reply above have been merged. I'm now back to regular development cycle of extendr, and waiting for a response on the paper review.

Hopefully, we can have this resolved soon, as I would like to push a new release of extendr together with the publication of this paper.
