### 2024-06-24 #6394 @alpaylan
Dear @jbytecode and @CGMossa,

As I've found the reasons to not place the comparison in the paper sufficient, I have checked the last box on the list.
