### 2024-06-28 #6394 @dccsillag
I'm happy to see the recent advances in ExtendR's development, which I believe have already been significant improvements. I have a few minor comments still, but given the timing I consider myself essentially satisfied and hence checked the remaining items in my review list.

Regarding error management: thank you for the explanation, @CGMossa. Indeed, it's pretty frustrating to have to deal with these CRAN limitations, and I now better understand the context for the current state of things. I am very glad to see that there is no longer a leak by default, and the error management situation makes quite some sense to me now.

Regarding performance: I agree with all the raised points.

Finally, regarding the docs: the user guide is a great stride forward. It seems to be growing into a great material and certainly makes things a lot more clear. I do have some more comments on it, though:
- I suggest addnig a link to it on https://extendr.github.io/. If there is one there, I could not find it. (And it's worth keeping in mind that this seems to now be the top Google result for 'extendr'.)
- The docs overall would benefit a tad from a careful grammar/language check overall. Make sure that everything is properly capitalized, etc.. It's currently reasonable, but clearly lacks a certain level 'polish', which typically makes a big difference for an incoming user.
- Finally, the docs would largely benefit quite a bit from a bit more prose. There are a number of occasions where it hops from one idea to another with little motivation or connection in between, which can be jarring for an unfamiliar reader. Similarly, in the user guide in particular, I feel like there can be a tad more text introducing the relevant jargon; for example, in the 'rheck' example there is no context given to the term 'vectorization'. While the term is common and my first guess was correct, it is a somewhat overloaded term. It does not hurt to very briefly explain what exactly you mean by it, to avoid any potential ambiguity or confusion. Any moment the user feels confused is a moment where they consider not using your solution.

That said, the current state of the docs and orientation around the ExtendR project's pages already seem quite an improvement from before. I also understand that these changes I've suggested take a while and quite a lot of effort, and given how long this review process has already dragged on along with my faith that the ExtendR team will continue to improve on these fronts, I consider myself satisfied.
