### 2024-06-28 #6394 @CGMossa
Thank you to both reviewers for your diligence and your review.

I agree, we need a reference to the user-guide on our main webpage. I didn't mention it, but we have a contributor that wants to port the website to quarto, and then merge the user-guide with the website. Their PR is here <https://github.com/extendr/extendr.github.io/pull/14>. This effort is slowly taken place.

# Missing DOIs

CRAN now provides a DOI for packages on it, so a lot of new DOIs has appeared. I'm adding them.
There are still some missing DOIs though:
- No DOI given, and none found for title: Andrie/Pagerank
This does not have a DOI, or a CITATION file in its repository.

Then these two citations also do not have DOIs.

- No DOI given, and none found for title: Writing R Extensions in Rust
- No DOI given, and none found for title: R Consortium Funded Project Extendr Provides Rust ...
