### 2024-06-28 #6394 @CGMossa
Okay. I've added a DOI for "Extending R" book.

Right now, we are working on preparing a release of extendr (which includes `libR-sys`) to accompany this publication. I'll post the depending PRs once they are all made.
