### 2024-06-28 #6394 @jbytecode
@CGMossa - We have some minor corrections to be applied in both the manuscript and the bibliography. Could you please review and apply them?

Paper

- In line 21 put a space between language and R
- In line 59, python should be Python

Bibliography

- Writing r extensions in rust: The first letter should be upper case so use {R} and {R}ust in the bibentry
- RInside: C++ classes to embed r in c++: Please capitalize the first letters of the programming language, e.g., {R} and {C}++
- V8: Embedded JavaScript and WebAssembly engine for r:  {R} (the same issue}
- Hellorust: Minimal examples of using rust code in r: {R}ust code in {R}
- rJava: Low-level r to java interface: {R} to {J}ava
- Reticulate: Interface to `python`: {P}ython
- cpp11: A c++11 interface for r's c interface: {C}++11, {R}, {C}
