### 2024-06-28 #6394 @jbytecode
@CGMossa - We need to create a tagged release in the software repository. It is about registering the reviewed code so you don't really need to push a new real release. Using this tagged release, the software will be registered in the Zenodo archive. Please prepare the archive and report me the DOI. The metadata in the archive should match those in the manuscript including the author list, the title, and ORCIDs, etc. Then I will relate the version tag and the Zenodo DOI with the paper review. Thank you in advance.
