### 2024-06-28 97d2c67
paper: applied https://github.com/openjournals/joss-reviews/issues/6394#issuecomment-2197196610 feedback
