+++ paper.bib
@@ -63,7 +63,7 @@
   abstract = {R package to compute pagerank of CRAN packages}
 }
 @manual{cpp11,
-  title  = {cpp11: A C++11 Interface for R's C Interface},
+  title  = {cpp11: A {C}++11 Interface for {R}'s {C} Interface},
   author = {Davis Vaughan and Jim Hester and Romain François},
   year   = {2023},
   note   = {R package version 0.4.7},
@@ -71,7 +71,7 @@
   doi    = {10.32614/cran.package.cpp11}
 }
 @manual{rJava,
-  title  = {rJava: Low-Level R to Java Interface},
+  title  = {r{J}ava: Low-Level {R} to {J}ava Interface},
   author = {Simon Urbanek},
   year   = {2023},
   note   = {R package version 1.0-10},
@@ -80,7 +80,7 @@
 }
 
   @manual{v8_cran,
-  title  = {V8: Embedded JavaScript and WebAssembly Engine for R},
+  title  = {V8: Embedded JavaScript and WebAssembly Engine for {R}},
   author = {Jeroen Ooms},
   year   = {2023},
   note   = {R package version 4.4.1},
@@ -88,7 +88,7 @@
   doi    = {10.32614/cran.package.v8}
 }
 @manual{reticulate_cran,
-  title  = {reticulate: Interface to 'Python'},
+  title  = {reticulate: Interface to {P}ython},
   author = {Kevin Ushey and JJ Allaire and Yuan Tang},
   year   = {2023},
   note   = {R package version 1.34.0},
@@ -115,7 +115,7 @@
   file         = {C:\Users\minin\Zotero\storage\IHYMYSSD\d41586-020-03382-2.html}
 }
 @misc{cargo_cran,
-  title         = {Writing R Extensions in Rust},
+  title         = {Writing {R} Extensions in {R}ust},
   author        = {David B. Dahl},
   year          = {2021},
   eprint        = {2108.07179},
@@ -155,7 +155,7 @@
   doi = {10.3929/ethz-b-000585774}
 }
  @Manual{hellorust_cran,
-    title = {hellorust: Minimal Examples of Using Rust Code in R},
+    title = {hellorust: Minimal Examples of Using {R}ust Code in {R}},
     author = {Jeroen Ooms and {Authors of the dependency Rust crates}},
     year = {2023},
     note = {R package version 1.2.0},
@@ -170,7 +170,7 @@
   doi = {10.32614/cran.package.usethis}
 }
 @Manual{rinside_cran,
-  title = {RInside: C++ Classes to Embed R in C++ (and C) Applications},
+  title = {RInside: {C}++ Classes to Embed {R} in {C}++ (and {C}) Applications},
   author = {Dirk Eddelbuettel and Romain Francois and Lance Bachmeier},
   year = {2023},
   note = {R package version 0.2.18},
@@ -178,7 +178,7 @@
   doi = {10.32614/cran.package.rinside}
 }
 @Manual{geos_cran,
-  title = {geos: Open Source Geometry Engine ('GEOS') R API},
+  title = {geos: Open Source Geometry Engine ('GEOS') {R} API},
   author = {Dewey Dunnington and Edzer Pebesma},
   year = {2022},
   note = {R package version 0.2.2},
@@ -234,7 +234,7 @@
 }
 
 @inproceedings{epirust_paper,
-  title = {{{EpiRust}}: {{Towards A Framework For Large-scale Agent-based Epidemiological Simulations Using Rust Language}}},
+  title = {{{EpiRust}}: {{Towards A Framework For Large-scale Agent-based Epidemiological Simulations Using {R}ust Language}}},
[…]
