### 2024-06-28 #6394 @jbytecode
@CGMossa - The paper and the bibliography seems far better now. We will have a final proof-read at the end of the stage. We can continue with the archiving stuff by now. Thank you in advance.
