### 2024-06-28 #6394 @CGMossa
Sorry, I'm a bit lost here. Do I tag the release on GitHub based on the branch in which the paper reside in? Or `master`?
