### 2024-06-28 #6394 @CGMossa
Okay. I've tried to do this, I got this link from Zenodo: https://zenodo.org/uploads/12584076?token=eyJhbGciOiJIUzUxMiJ9.eyJpZCI6ImUzOWUwM2QyLTE5MDktNDBhMC1iODdmLWY0ZDlhYWVjNTI4ZCIsImRhdGEiOnt9LCJyYW5kb20iOiIxZjQ3YjhlYTdhZDlkZjlmOGUxNTA3NzExYjU3Nzg0YiJ9.5jDzS1wHqCaj6RwK83fG13HQmwbRCqsDa24zQiTVIoscDTsVM1DFzgwiGmR3nwzPPFrAONV-x8jkgYDeRkrpJQ

And I got this 10.5281/zenodo.12584076

Does this work? I tried to find instructions as to how to do this, but there were very little in the JOSS FAQ.
