### 2024-06-28 #6394 @jbytecode
The link provided doesn't work because it seems private.
The DOI is also not resolved. https://doi.org/10.5281/zenodo.12584076
