### 2024-06-28 #6394 @jbytecode
Here is a sample Zenodo archive in JOSS, you can trace how it looks like: https://zenodo.org/records/5534663
