### 2024-06-28 #6394 @CGMossa
I've tried it now: https://zenodo.org/records/12584076
