### 2024-07-01 #6394 @jbytecode
@CGMossa - The last author in the manuscript is `Amy Thomason` whereas it is stated as `Andy Thomason` in the Zenodo archive. Which one is the correct one? Could you please fix it and ping me again?
