### 2024-07-01 #6394 @CGMossa
It's Amy Thomason. They came out as trans, and I asked yesterday which name to add to the manuscript. I'll fix zenodo.
