### 2024-07-01 #6394 @CGMossa
I've updated zenodo. I had updated earlier, but I had to press publish, as the save button isn't enough. It saves the draft or something like that.
