### 2024-07-01 #6394 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
