### 2024-07-01 407c0e5
paper: applying recommendation in https://github.com/openjournals/joss-reviews/issues/6394#issuecomment-2200656867
