### 2024-07-01 #6394 @CGMossa
> @CGMossa - I'm the track editor, proofreading this paper before further processing it. Please
>
> * merge [Patch 1 extendr/extendr#816](https://github.com/extendr/extendr/pull/816) or let me know what you disagree with
> * change affiliation 4 to "Independent researcher"
> * add countries to affiliations 4, 5, and 6

Thank you for giving it a look! I've merged the patch and applied your changes.
I just saw that there is one type with `difffers`. I've just pushed a commit to fix that. Should be alright now.
