### 2024-07-01 #6394 @danielskatz
Congratulations to @CGMossa (Mossa Reimert) and co-authors on your publication!!

JOSS depends on volunteers, and it wouldn't be successful without you
